Crate noirc_frontend

The noir compiler is separated into the following passes which are listed in order in square brackets. The inputs and outputs of each pass are also given:

Source file -[Lexing]-> Tokens -[Parsing]-> Ast -[Name Resolution]-> Hir -[Type Checking]-> Hir -[Monomorphization]-> Monomorphized Ast

After the monomorphized ast is created, it is passed to the noirc_evaluator crate to convert it to SSA form, perform optimizations, convert to ACIR and eventually prove/verify the program.

Re-exports

pub use lexer::token;

pub use parser::ParsedModule;

pub use parser::parse_program;

pub use parser::parse_program_with_dummy_file;

pub use hir_def::types::*;

Modules

ast

The submodules of this module define the various data types required to represent Noir’s Ast. Of particular importance are ExpressionKind and Statement which can be found in expression.rs and statement.rs respectively.

debug

elaborator

The Elaborator

graph

hir

hir_def

Noir’s HIR (High-Level Intermediate Representation) is produced by the elaboration pass and represents a fully resolved and type-checked program. The HIR is the input to the monomorphization pass.

lexer

The lexer is the first pass of the noir compiler. Its goal is to convert a character stream (one input file) into a stream of tokens. See lexer.rs for the lexer itself, and token.rs for more details on tokens.

locations

modules

monomorphization

Coming after type checking, monomorphization is the last pass in Noir’s frontend. It accepts the type checked HIR as input and produces a monomorphized AST as output. This file implements the pass itself, while the AST is defined in the ast module.

node_interner

ownership

This module implements the “ownership analysis” compiler pass on the monomorphized AST. It is run after monomorphization and before SSA-gen. At this point the monomorphized AST has no polymorphic types and all functions are specialized into either constrained or unconstrained versions. This pass only operates on unconstrained functions since only Noir’s unconstrained runtime Brillig has any notion of cloning a value - specifically arrays with their brillig-only reference counts.

parser

The parser is the second pass of the noir compiler. The parser’s job is to take the output of the lexer (a stream of tokens) and parse it into a valid Abstract Syntax Tree (Ast). During this, the parser validates the grammar of the program and returns parsing errors for any syntactically invalid constructs (such as fn fn fn).

resolve_locations

shared

The shared module contains simple types which are using in multiple of Noir’s IRs.

signed_field

test_utils

This crate represents utility methods which can be useful for testing in other crates which also desire to compile the frontend.

usage_tracker