noirc_frontend

Module ast

The submodules of this module define the various data types required to represent Noir’s Ast. Of particular importance are ExpressionKind and Statement which can be found in expression.rs and statement.rs respectively.

Noir’s Ast is produced by the parser and taken as input to name resolution, where it is converted into the Hir (defined in the hir_def module).

Modules

docs 🔒

enumeration 🔒

expression 🔒

function 🔒

statement 🔒

structure 🔒

traits 🔒

type_alias 🔒

visitor 🔒

Structs

AsTraitPath

A special kind of path in the form <MyType as Trait>::ident. Note that this path must consist of exactly two segments.

AssignStatement

BinaryOpKindIter

An iterator over the variants of Self

BlockExpression

CallExpression

CastExpression

ConstrainExpression

ConstructorExpression

Documented

EnumVariant

We only support variants of the form Name(A, B, ...) currently. Enum variants like Name { a: A, b: B, .. } will be implemented later

Expression

ForBounds

ForLoopStatement

FunctionDefinition

GenericTypeArgs

Ident

IfExpression

ImportStatement

IndexExpression

InfixExpression

IntegerBitSizeIter

An iterator over the variants of Self

InvalidIntegerBitSizeError

Lambda

LetStatement

LoopStatement

MatchExpression

MemberAccessExpression

MethodCallExpression

ModuleDeclaration

NoirEnumeration

Ast node for an enum

NoirFunction

NoirStruct

Ast node for a struct

NoirTrait

AST node for trait definitions: trait name<generics> { ... items ... }

NoirTraitImpl

Ast node for an implementation of a trait for a particular type impl trait_name<trait_generics> for object_type where where_clauses { ... items ... }

Param

Path

PathSegment

PrefixExpression

Statement

StructField

TraitBound

Represents a single trait bound, such as TraitX or TraitY<U, V>

TraitImplItem

TypeAlias

Ast node for type aliases Depending on ‘numeric_type’, a Type Alias can be an alias to a normal type, or to a numeric generic type

TypeImpl

Ast node for an impl of a concrete type impl object_type<generics> { ... methods ... }

TypePath

A special kind of path in the form Type::ident::<turbofish> Unlike normal paths, the type here can be a primitive type or interned type.

UnresolvedTraitConstraint

Represents a simple trait constraint such as where Foo: TraitY<U, V> Complex trait constraints such as where Foo: Display + TraitX + TraitY<U, V> are converted in the parser to a series of simple constraints: Foo: Display Foo: TraitX Foo: TraitY<U, V>

UnresolvedType

UnsafeExpression

UnsupportedNumericGenericType

UseTree

WhileStatement

Enums

ArrayLiteral

AttributeTarget

BinaryOpKind

ConstrainKind

ExpressionKind

ForRange

FunctionKind

Currently, we support four types of functions:

FunctionReturnType

GenericTypeArg

An argument to a generic type or trait.

GenericTypeArgKind

IdentOrQuotedType

IntegerBitSize

ItemVisibility

Represents whether the definition can be referenced outside its module/crate

LValue

Represents an Ast form that can be assigned to

Literal

PathKind

Pattern

StatementKind

Ast node for statements in noir. Statements are always within a block { } of some kind and are terminated via a Semicolon, except if the statement ends in a block, such as a Statement::Expression containing an if expression.

TraitImplItemKind

TraitItem

Any declaration inside the body of a trait that a user is required to specify when implementing the trait.

UnaryOp

UnresolvedGeneric

UnresolvedTypeData

The parser parses types as ’UnresolvedType’s which require name resolution to resolve any type names used for structs within, but are otherwise identical to Types.

UnresolvedTypeExpression

The precursor to TypeExpression, this is the type that the parser allows to be used in the length position of an array type. Only constant integers, variables, and numeric binary operators are allowed here.

UseTreeKind

Constants

ERROR_IDENT

This is used when an identifier fails to parse in the parser. Instead of failing the parse, we can often recover using this as the default value instead. Further passes like name resolution should also check for this ident to avoid issuing multiple errors for an identifier that already failed to parse.

WILDCARD_TYPE

This is used to represent an UnresolvedTypeData::Unspecified in a Path

Traits

Visitor

Implements the Visitor pattern for Noir’s AST.

Type Aliases

BinaryOp

DocComment

UnresolvedGenerics

A Vec of unresolved names for type variables. For fn foo<A, B>(...) this corresponds to vec![“A”, “B”].