noirc_frontend

Module debug

Structs

DebugInstrumenter

This structure is used to collect information about variables to track for debugging during the instrumentation injection phase.

SourceFieldId

SourceVarId

Constants

MAX_MEMBER_ASSIGN_DEPTH 🔒

Functions

build_assign_member_stmt 🔒

build_assign_var_stmt 🔒

build_debug_call_stmt 🔒

build_debug_crate_file

build_drop_var_stmt 🔒

id_expr 🔒

ident 🔒

pattern_to_string 🔒

pattern_vars 🔒

sint_expr 🔒

uint_expr 🔒