noirc_frontend::elaborator::item_context

Module body_context

The part of an item’s context that only a function body can have: the loop, unsafe block, lambda and call-argument scopes an expression sits in.

Structs

BodyContext 🔒

Where in a function body the elaborator currently is.

EnclosingUnsafeBlock 🔒

The unsafe block state of the code around the block just entered, to be handed back to BodyContext::exit_unsafe_block when it is left.

Enums

UnsafeBlockStatus 🔒

Determines whether we are in an unsafe block and, if so, whether any unconstrained calls were found in it (because if not we’ll warn that the unsafe block is not needed).