noirc_frontend::elaborator

Module lints

Lint checks for function attributes, visibility, and usage restrictions.

Functions

can_return_without_recursing 🔒

Check if an expression will end up calling a specific function.

can_return_without_recursing_match 🔒

check_integer_literal_fits_its_type 🔒

Checks if an ExprId, which has to be an integer literal, fits in its type.

check_varargs 🔒

databus_on_non_entry_point 🔒

call_data and return_data visibility modifiers are only allowed on entry point functions.

deprecated_function 🔒

error_if_verify_proof_with_type 🔒

Errors if func_expr_id is std::verify_proof_with_type

func_meta_name_ident 🔒

inlining_attributes 🔒

Validate inline-related attributes based on whether the function is constrained or unconstrained.

low_level_function_outside_stdlib 🔒

Attempting to define new low level (#[builtin] or #[foreign]) functions outside of the stdlib is disallowed.

missing_pub 🔒

pub is required on return types for entry point functions

no_predicates_on_entry_point 🔒

The #[no_predicates] attribute is not allowed on entry point functions since it’s meant to control inlining into the entry point.

oracle_name_clashes_with_stdlib 🔒

Attempting to define an #[oracle] functions with a name that clashes with those in the stdlib is disallowed.

oracle_not_marked_unconstrained 🔒

Oracle definitions (functions with the #[oracle] attribute) must be marked as unconstrained.

oracle_returns_multiple_vectors 🔒

Oracle functions cannot return more than 1 vector in their output.

oracle_returns_reference 🔒

Oracle functions cannot return references

oracle_returns_vector_with_nested_array 🔒

Oracles cannot return vectors containing nested arrays because deflattening is not yet implemented in the VM.

unbounded_recursion 🔒

Check that a recursive function can return without endlessly calling itself.

unconstrained_function_args 🔒

Check that we are not passing a mutable reference from a constrained runtime to an unconstrained runtime.

unconstrained_function_return 🔒

Check that that a type returned from an unconstrained to a constrained runtime is safe:

unnecessary_pub_argument 🔒

Only arguments to entrypoint functions may have a non-private visibility modifier applied to them.

unnecessary_pub_return 🔒

Only entrypoint functions require a pub visibility modifier applied to their return types.