noirc_frontend::elaborator::lints

Function unnecessary_pub_return

pub(super) fn unnecessary_pub_return(
    func: &FuncMeta,
    modifiers: &FunctionModifiers,
    is_entry_point: bool,
) -> Option<ResolverError>

Only entrypoint functions require a pub visibility modifier applied to their return types.

Application of pub to other functions is not meaningful and is a mistake.