noirc_frontend::elaborator

Module options

Compiler frontend options and unstable feature flags.

Structs

GenericOptions

Generic options struct meant to resolve to ElaboratorOptions below when we can resolve a file path to a file id later. This generic struct is used so that FrontendOptions doesn’t need to duplicate fields and methods with ElaboratorOptions.

Enums

UnstableFeature

Type Aliases

ElaboratorOptions 🔒

Options from nargo_cli that need to be passed down to the elaborator

FrontendOptions

This is the unresolved version of ElaboratorOptions CLI options that need to be passed to the compiler frontend (the elaborator).