noirc_frontend::elaborator::options
Type Alias ElaboratorOptions

pub(crate) type ElaboratorOptions<'a> = GenericOptions<'a, FileId>;

Options from nargo_cli that need to be passed down to the elaborator

Aliased Type

struct ElaboratorOptions<'a> {
    pub debug_comptime_in_file: Option<FileId>,
    pub pedantic_solving: bool,
    pub enabled_unstable_features: &'a [UnstableFeature],
    pub disable_required_unstable_features: bool,
}

Fields
debug_comptime_in_file: Option<FileId>

The scope of –debug-comptime, or None if unset

pedantic_solving: bool

Use pedantic ACVM solving

enabled_unstable_features: &'a [UnstableFeature]

Unstable compiler features that were explicitly enabled. Any unstable features that are not in this list result in an error when used.

disable_required_unstable_features: bool

Deny crates from requiring unstable features.