noirc_frontend::elaborator

Module path_resolution

Path resolution for types, values, and trait methods across modules.

Structs

PathResolution 🔒

Turbofish

TypedPath

Like a crate::ast::Path but each segment has resolved turbofish types.

TypedPathSegment

Like a crate::ast::PathSegment but with resolved turbofish types.

Enums

IntermediatePathResolutionItem 🔒

Any item that can appear before the last segment in a path.

MethodLookupResult 🔒

PathResolutionItem 🔒

All possible items that result from resolving a Path. Note that this item doesn’t include the last turbofish in a Path, only intermediate ones, if any.

PathResolutionMode 🔒

Determines whether data-types found along a path are to be marked as referenced or used (see crate::usage_tracker::UsageTracker::mark_as_referenced and crate::usage_tracker::UsageTracker::mark_as_used)

PathResolutionTarget 🔒

Depending on where a path appears in the source code it should either resolve to a type or a value. For example, in let x: Foo::Bar = Foo::Bar {} both Foo::Bar should resolve to types, never values. On the other hand, in Foo::Bar() Foo::Bar should resolve to a value, typically a function.

Functions

get_type_alias_module_def_id 🔒

merge_intermediate_path_resolution_item_with_module_def_id 🔒

Transform a ModuleDefId into a PathResolutionItem.

Type Aliases

PathResolutionResult 🔒