noirc_frontend::elaborator
Struct Elaborator

pub struct Elaborator<'context> { /* private fields */ }

Implementations

impl<'context> Elaborator<'context>

pub fn elaborate_item_from_comptime_in_function<'a, T>( &'a mut self, current_function: Option<FuncId>, reason: Option<ElaborateReason>, f: impl FnOnce(&mut Elaborator<'a>) -> T, ) -> T

Elaborate an expression from the middle of a comptime scope. When this happens we require additional information to know what variables should be in scope.

pub fn elaborate_item_from_comptime_in_module<'a, T>( &'a mut self, module: ModuleId, reason: Option<ElaborateReason>, f: impl FnOnce(&mut Elaborator<'a>) -> T, ) -> T

pub fn setup_interpreter<'local>( &'local mut self, ) -> Interpreter<'local, 'context>

impl Elaborator<'_>

pub fn intern_expr(&mut self, expr: HirExpression, location: Location) -> ExprId

pub fn inline_comptime_value( &mut self, value: Result<Value, InterpreterError>, location: Location, ) -> (ExprId, Type)

impl Elaborator<'_>

pub fn elaborate_pattern_and_store_ids( &mut self, pattern: Pattern, expected_type: Type, definition_kind: DefinitionKind, created_ids: &mut Vec<HirIdent>, warn_if_unused: bool, ) -> HirPattern

Equivalent to elaborate_pattern, this version just also adds any new DefinitionIds that were created to the given Vec.

pub fn add_existing_variable_to_scope( &mut self, name: String, ident: HirIdent, warn_if_unused: bool, )

pub fn add_global_variable_decl( &mut self, name: Ident, global_id: GlobalId, ) -> HirIdent

pub fn get_ident_from_path( &mut self, path: Path, ) -> ((HirIdent, usize), Option<PathResolutionItem>)

impl Elaborator<'_>

pub fn module_id(&self) -> ModuleId

pub fn replace_module(&mut self, new_module: ModuleId) -> ModuleId

pub fn push_scope(&mut self)

pub fn pop_scope(&mut self)

pub fn check_for_unused_variables_in_scope_tree( &mut self, scope_decls: ScopeTree<String, ResolverMeta>, )

pub fn lookup_trait_or_error(&mut self, path: Path) -> Option<&mut Trait>

Lookup a given trait by name/path.

pub fn lookup_datatype_or_error( &mut self, path: Path, ) -> Option<Shared<DataType>>

Lookup a given struct type by name.

pub fn lookup_type_alias(&mut self, path: Path) -> Option<Shared<TypeAlias>>

impl Elaborator<'_>

pub fn collect_traits( &mut self, traits: &mut BTreeMap<TraitId, UnresolvedTrait>, )

pub fn resolve_trait_function( &mut self, trait_id: TraitId, name: &Ident, is_unconstrained: bool, generics: &UnresolvedGenerics, parameters: &[(Ident, UnresolvedType)], return_type: &FunctionReturnType, where_clause: Vec<UnresolvedTraitConstraint>, body: &Option<BlockExpression>, trait_visibility: ItemVisibility, func_id: FuncId, )

impl Elaborator<'_>

pub fn resolve_type_inner(&mut self, typ: UnresolvedType, kind: &Kind) -> Type

Translates an UnresolvedType into a Type and appends any freshly created TypeVariables created to new_variables.

pub fn find_generic(&self, target_name: &str) -> Option<&ResolvedGeneric>

pub fn lookup_generic_or_global_type(&mut self, path: &Path) -> Option<Type>

pub fn unify_without_applying_bindings( &mut self, actual: &Type, expected: &Type, make_error: impl FnOnce() -> TypeCheckError, )

Do not apply type bindings even after a successful unification. This function is used by the interpreter for some comptime code which can change types e.g. on each iteration of a for loop.

pub fn type_check_function_body( &mut self, body_type: Type, meta: &FuncMeta, body_id: ExprId, )

pub fn verify_trait_constraint( &mut self, object_type: &Type, trait_id: TraitId, trait_generics: &[Type], associated_types: &[NamedType], function_ident_id: ExprId, select_impl: bool, location: Location, )

pub fn add_existing_generics( &mut self, unresolved_generics: &UnresolvedGenerics, generics: &Generics, )

pub fn add_existing_generic( &mut self, unresolved_generic: &UnresolvedGeneric, location: Location, resolved_generic: &ResolvedGeneric, )

pub fn push_trait_constraint( &mut self, constraint: TraitConstraint, expr_id: ExprId, select_impl: bool, )

Push a trait constraint into the current FunctionContext to be solved if needed at the end of the earlier of either the current function or the current comptime scope.

pub fn bind_generics_from_trait_constraint( &self, constraint: &TraitConstraint, assumed: bool, bindings: &mut TypeBindings, )

pub fn bind_generics_from_trait_bound( &self, trait_bound: &ResolvedTraitBound, bindings: &mut TypeBindings, )

pub fn instantiate_parent_trait_bound( &self, trait_bound: &ResolvedTraitBound, parent_trait_bound: &ResolvedTraitBound, ) -> ResolvedTraitBound

impl Elaborator<'_>

pub fn find_unquoted_exprs_tokens(&mut self, tokens: Tokens) -> Tokens

Go through the given tokens looking for a ‘$’ token followed by a variable to unquote. Each time these two tokens are found, they are replaced by a new UnquoteMarker token containing the ExprId of the resolved variable to unquote.

impl<'context> Elaborator<'context>

pub fn new( interner: &'context mut NodeInterner, def_maps: &'context mut DefMaps, usage_tracker: &'context mut UsageTracker, crate_graph: &'context CrateGraph, crate_id: CrateId, interpreter_call_stack: Vector<Location>, options: GenericOptions<'context, FileId>, elaborate_reasons: Vector<ElaborateReason>, ) -> Self

pub fn from_context( context: &'context mut Context<'_, '_>, crate_id: CrateId, options: GenericOptions<'context, FileId>, ) -> Self

pub fn elaborate( context: &'context mut Context<'_, '_>, crate_id: CrateId, items: CollectedItems, options: GenericOptions<'context, FileId>, ) -> Vec<CompilationError>

pub fn elaborate_and_return_self( context: &'context mut Context<'_, '_>, crate_id: CrateId, items: CollectedItems, options: GenericOptions<'context, FileId>, ) -> Self

pub fn pedantic_solving(&self) -> bool

True if we should use pedantic ACVM solving

pub fn add_generics(&mut self, generics: &UnresolvedGenerics) -> Generics

Add the given generics to scope. Each generic will have a fresh Shared associated with it.

pub fn resolve_module_by_path(&mut self, path: Path) -> Option<ModuleId>

pub fn resolve_trait_bound( &mut self, bound: &TraitBound, ) -> Option<ResolvedTraitBound>

pub fn get_module(&self, module: ModuleId) -> &ModuleData

pub fn resolve_struct_fields( &mut self, unresolved: &NoirStruct, struct_id: TypeId, ) -> Vec<StructField>

pub fn use_unstable_feature( &mut self, feature: UnstableFeature, location: Location, )

Register a use of the given unstable feature. Errors if the feature has not been explicitly enabled in this package.

pub fn errors_occurred_in<T>( &mut self, f: impl FnOnce(&mut Self) -> T, ) -> (bool, T)

Run the given function using the resolver and return true if any errors (not warnings) occurred while running it.