noirc_frontend::elaborator::traits
Function check_function_type_matches_expected_type

fn check_function_type_matches_expected_type(
    expected: &Type,
    actual: &Type,
    method_name: &str,
    actual_parameters: &Parameters,
    actual_return_type: &FunctionReturnType,
    noir_function: &NoirFunction,
    location: Location,
    trait_name: &str,
    errors: &mut Vec<TypeCheckError>,
)

Check the given function type matches the expected one.

This is used to check if a trait impl’s function type matches the declared function in the original trait declaration - while handling the appropriate generic substitutions.