noirc_frontend::elaborator::traits
Function check_trait_impl_method_matches_declaration

pub(crate) fn check_trait_impl_method_matches_declaration(
    interner: &mut NodeInterner,
    function: FuncId,
    noir_function: &NoirFunction,
) -> Vec<TypeCheckError>

Checks that the type of a function in a trait impl matches the type of the corresponding function declaration in the trait itself.

To do this, given a trait such as: trait Foo<A> { fn foo<B>(...); }

And an impl such as: impl<C> Foo<D> for Bar<E> { fn foo<F>(...); }

We have to substitute:

Self for Bar<E>

A for D

B for F

Before we can type check. Finally, we must also check that the unification result does not introduce any new bindings. This can happen if the impl function’s type is more general than that of the trait function. E.g. fn baz<A, B>(a: A, b: B) when the impl required fn baz<A>(a: A, b: A).

This does not type check the body of the impl function.