noirc_frontend::elaborator

Module types

Type resolution, unification, and method resolution (for both types and traits).

Structs

TraitPathResolution 🔒

Enums

TraitPathResolutionMethod 🔒

WildcardAllowed

WildcardDisallowedContext

Constants

SELF_TYPE_NAME

Functions

bind_generic 🔒

Binds the type variable in a ResolvedGeneric, e.g. a generic parameter of a trait, to a Type, which itself can be an unbound type variable.

bind_named_generics 🔒

Binds the associated ResolvedGenerics of a trait to the named generics in a ResolvedTraitBound.

bind_ordered_generics 🔒

Binds the ordered ResolvedGenerics of a trait to the ordered generics in a ResolvedTraitBound.