noirc_frontend

Module hir

Modules

comptime

The comptime interpreter is used to execute Noir code which is marked to be executed at compilation time. The main uses for this are: - Metaprogramming such as deriving trait implementations - Interpreting comptime blocks, attribute functions, and macros - Evaluation of globals

def_collector

This set of modules implements the name resolution pass which converts the AST into the HIR. In comparison to the AST, the HIR is interned in the NodeInterner and attaches DefinitionIds to each Variable AST node to link them up to their definition. In doing so, this pass validates scoping requirements and is responsible for issuing name/definition related errors including missing definitions and duplicate definitions. One aspect of handling definition links is handling and tracking imports. Another result of this pass is that all imports will be removed from the program, and the AST of each file will be combined into a larger Hir stored in the NodeInterner and linked together via DefinitionIds.

def_map

printer

resolution

This set of modules implements the second half of the name resolution pass. After all definitions are collected by def_collector, resolver::Resolvers are created to resolve all names within a definition. In this context ‘resolving’ a name means validating that it has a valid definition, and that it was not redefined multiple times in the same scope. Once this is validated, it is linked to that definition via a matching DefinitionId. All references to the same definition will have the same DefinitionId.

scope

type_check

Structs

Context

Helper object which groups together several useful context objects used during name resolution. Once name resolution is finished, only the def_interner is required for type inference and monomorphization.

Enums

FunctionNameMatch

Functions

fully_qualified_function_name 🔒

get_all_test_functions_in_crate_matching

Type Aliases

ParsedFiles