noirc_frontend::hir::comptime

Module display

Structs

TokenPrettyPrinter 🔒

Tries to print tokens in a way that it’ll be easier for the user to understand a stream of tokens without having to format it themselves.

TokenPrinter

TokensPrettyPrinter 🔒

ValuePrinter

Functions

display_quoted 🔒

display_trait_constraint 🔒

remove_interned_in_array_literal 🔒

remove_interned_in_expression 🔒

remove_interned_in_expression_kind 🔒

remove_interned_in_generic_type_args 🔒

remove_interned_in_literal 🔒

remove_interned_in_lvalue 🔒

remove_interned_in_option_unresolved_type 🔒

remove_interned_in_pattern 🔒

remove_interned_in_statement 🔒

remove_interned_in_statement_expr 🔒

remove_interned_in_statement_kind 🔒

remove_interned_in_unresolved_type 🔒

remove_interned_in_unresolved_type_data 🔒

tokens_to_string

tokens_to_string_with_indent