noirc_frontend::hir::comptime::errors
Enum InterpreterError

pub enum InterpreterError {

    ArgumentCountMismatch {
        expected: usize,
        actual: usize,
        location: Location,
    },
    TypeMismatch {
        expected: String,
        actual: Type,
        location: Location,
    },
    NonComptimeVarReferenced {
        name: String,
        location: Location,
    },
    VariableNotInScope {
        location: Location,
    },
    IntegerOutOfRangeForType {
        value: SignedField,
        typ: Type,
        location: Location,
    },
    ErrorNodeEncountered {
        location: Location,
    },
    NonFunctionCalled {
        typ: Type,
        location: Location,
    },
    NonBoolUsedInIf {
        typ: Type,
        location: Location,
    },
    NonBoolUsedInWhile {
        typ: Type,
        location: Location,
    },
    NonBoolUsedInConstrain {
        typ: Type,
        location: Location,
    },
    FailingConstraint {
        message: Option<String>,
        location: Location,
        call_stack: Vector<Location>,
    },
    NonIntegerUsedInLoop {
        typ: Type,
        location: Location,
    },
    RangeBoundsTypeMismatch {
        start_type: Type,
        end_type: Type,
        location: Location,
    },
    NonPointerDereferenced {
        typ: Type,
        location: Location,
    },
    NonTupleOrStructInMemberAccess {
        typ: Type,
        location: Location,
    },
    NonArrayIndexed {
        typ: Type,
        location: Location,
    },
    NonIntegerUsedAsIndex {
        typ: Type,
        location: Location,
    },
    NonIntegerIntegerLiteral {
        typ: Type,
        location: Location,
    },
    NonIntegerArrayLength {
        typ: Type,
        err: Option<Box<TypeCheckError>>,
        location: Location,
    },
    NonIntegerAssociatedConstant {
        typ: Type,
        err: Option<Box<TypeCheckError>>,
        location: Location,
    },
    NonNumericCasted {
        typ: Type,
        location: Location,
    },
    IndexOutOfBounds {
        index: usize,
        length: usize,
        location: Location,
    },
    ExpectedStructToHaveField {
        typ: Type,
        field_name: String,
        location: Location,
    },
    TypeUnsupported {
        typ: Type,
        location: Location,
    },
    InvalidValueForUnary {
        typ: Type,
        operator: &'static str,
        location: Location,
    },
    InvalidValuesForBinary {
        lhs: Type,
        rhs: Type,
        operator: &'static str,
        location: Location,
    },
    BinaryOperationOverflow {
        operator: &'static str,
        location: Location,
    },
    NegateWithOverflow {
        location: Location,
    },
    CannotApplyMinusToType {
        location: Location,
        typ: &'static str,
    },
    CastToNonNumericType {
        typ: Type,
        location: Location,
    },
    NonStructInConstructor {
        typ: Type,
        location: Location,
    },
    NonEnumInConstructor {
        typ: Type,
        location: Location,
    },
    CannotInlineMacro {
        value: String,
        typ: Type,
        location: Location,
    },
    UnquoteFoundDuringEvaluation {
        location: Location,
    },
    DebugEvaluateComptime {
        diagnostic: CustomDiagnostic,
        location: Location,
    },
    FailedToParseMacro {
        error: Box<ParserError>,
        tokens: String,
        rule: &'static str,
        location: Location,
    },
    UnsupportedTopLevelItemUnquote {
        item: String,
        location: Location,
    },
    ComptimeDependencyCycle {
        function: String,
        location: Location,
    },
    NoImpl {
        location: Location,
    },
    NoMatchingImplFound {
        error: NoMatchingImplFoundError,
    },
    ImplMethodTypeMismatch {
        expected: Type,
        actual: Type,
        location: Location,
    },
    BreakNotInLoop {
        location: Location,
    },
    ContinueNotInLoop {
        location: Location,
    },
    BlackBoxError(BlackBoxResolutionError, Location),
    FailedToResolveTraitBound {
        trait_bound: TraitBound,
        location: Location,
    },
    TraitDefinitionMustBeAPath {
        location: Location,
    },
    FailedToResolveTraitDefinition {
        location: Location,
    },
    FunctionAlreadyResolved {
        location: Location,
    },
    MultipleMatchingImpls {
        object_type: Type,
        candidates: Vec<String>,
        location: Location,
    },
    Unimplemented {
        item: String,
        location: Location,
    },
    InvalidInComptimeContext {
        item: String,
        location: Location,
        explanation: String,
    },
    TypeAnnotationsNeededForMethodCall {
        location: Location,
    },
    ExpectedIdentForStructField {
        value: String,
        index: usize,
        location: Location,
    },
    InvalidAttribute {
        attribute: String,
        location: Location,
    },
    GenericNameShouldBeAnIdent {
        name: Rc<String>,
        location: Location,
    },
    DuplicateGeneric {
        name: Rc<String>,
        struct_name: String,
        duplicate_location: Location,
        existing_location: Location,
    },
    CannotResolveExpression {
        location: Location,
        expression: String,
    },
    CannotSetFunctionBody {
        location: Location,
        expression: String,
    },
    UnknownArrayLength {
        length: Type,
        err: Box<TypeCheckError>,
        location: Location,
    },
    CannotInterpretFormatStringWithErrors {
        location: Location,
    },
    GlobalsDependencyCycle {
        location: Location,
    },
    GlobalCouldNotBeResolved {
        location: Location,
    },
    LoopHaltedForUiResponsiveness {
        location: Location,
    },
    DuplicateStructFieldInSetFields {
        name: Ident,
        index: usize,
        previous_index: usize,
    },
    CheckedTransmuteFailed {
        actual: Type,
        expected: Type,
        location: Location,
    },
    StackOverflow {
        location: Location,
        call_stack: Vector<Location>,
    },
    EvaluationDepthOverflow {
        location: Location,
        call_stack: Vector<Location>,
    },
    Break,
    Continue,
    SkippedDueToEarlierErrors,

}

The possible errors that can halt the interpreter.

Variants

ArgumentCountMismatch

Fields

expected: usize

actual: usize

location: Location

TypeMismatch

Fields

expected: String

actual: Type

location: Location

NonComptimeVarReferenced

Fields

name: String

location: Location

VariableNotInScope

Fields

location: Location

IntegerOutOfRangeForType

Fields

value: SignedField

typ: Type

location: Location

ErrorNodeEncountered

Fields

location: Location

NonFunctionCalled

Fields

typ: Type

location: Location

NonBoolUsedInIf

Fields

typ: Type

location: Location

NonBoolUsedInWhile

Fields

typ: Type

location: Location

NonBoolUsedInConstrain

Fields

typ: Type

location: Location

FailingConstraint

Fields

message: Option<String>

location: Location

call_stack: Vector<Location>

NonIntegerUsedInLoop

Fields

typ: Type

location: Location

RangeBoundsTypeMismatch

Fields

start_type: Type

end_type: Type

location: Location

NonPointerDereferenced

Fields

typ: Type

location: Location

NonTupleOrStructInMemberAccess

Fields

typ: Type

location: Location

NonArrayIndexed

Fields

typ: Type

location: Location

NonIntegerUsedAsIndex

Fields

typ: Type

location: Location

NonIntegerIntegerLiteral

Fields

typ: Type

location: Location

NonIntegerArrayLength

Fields

typ: Type

err: Option<Box<TypeCheckError>>

location: Location

NonIntegerAssociatedConstant

Fields

typ: Type

err: Option<Box<TypeCheckError>>

location: Location

NonNumericCasted

Fields

typ: Type

location: Location

IndexOutOfBounds

Fields

index: usize

length: usize

location: Location

ExpectedStructToHaveField

Fields

typ: Type

field_name: String

location: Location

TypeUnsupported

Fields

typ: Type

location: Location

InvalidValueForUnary

Fields

typ: Type

operator: &'static str

location: Location

InvalidValuesForBinary

Fields

lhs: Type

rhs: Type

operator: &'static str

location: Location

BinaryOperationOverflow

Fields

operator: &'static str

location: Location

NegateWithOverflow

Fields

location: Location

CannotApplyMinusToType

Fields

location: Location

typ: &'static str

CastToNonNumericType

Fields

typ: Type

location: Location

NonStructInConstructor

Fields

typ: Type

location: Location

NonEnumInConstructor

Fields

typ: Type

location: Location

CannotInlineMacro

Fields

value: String

typ: Type

location: Location

UnquoteFoundDuringEvaluation

Fields

location: Location

DebugEvaluateComptime

Fields

diagnostic: CustomDiagnostic

location: Location

FailedToParseMacro

Fields

error: Box<ParserError>

tokens: String

rule: &'static str

location: Location

UnsupportedTopLevelItemUnquote

Fields

item: String

location: Location

ComptimeDependencyCycle

Fields

function: String

location: Location

NoImpl

Fields

location: Location

NoMatchingImplFound

Fields

error: NoMatchingImplFoundError

ImplMethodTypeMismatch

Fields

expected: Type

actual: Type

location: Location

BreakNotInLoop

Fields

location: Location

ContinueNotInLoop

Fields

location: Location

BlackBoxError(BlackBoxResolutionError, Location)

FailedToResolveTraitBound

Fields

trait_bound: TraitBound

location: Location

TraitDefinitionMustBeAPath

Fields

location: Location

FailedToResolveTraitDefinition

Fields

location: Location

FunctionAlreadyResolved

Fields

location: Location

MultipleMatchingImpls

Fields

object_type: Type

candidates: Vec<String>

location: Location

Unimplemented

Fields

item: String

location: Location

InvalidInComptimeContext

Fields

item: String

location: Location

explanation: String

TypeAnnotationsNeededForMethodCall

Fields

location: Location

ExpectedIdentForStructField

Fields

value: String

index: usize

location: Location

InvalidAttribute

Fields

attribute: String

location: Location

GenericNameShouldBeAnIdent

Fields

name: Rc<String>

location: Location

DuplicateGeneric

Fields

name: Rc<String>

struct_name: String

duplicate_location: Location

existing_location: Location

CannotResolveExpression

Fields

location: Location

expression: String

CannotSetFunctionBody

Fields

location: Location

expression: String

UnknownArrayLength

Fields

length: Type

err: Box<TypeCheckError>

location: Location

CannotInterpretFormatStringWithErrors

Fields

location: Location

GlobalsDependencyCycle

Fields

location: Location

GlobalCouldNotBeResolved

Fields

location: Location

LoopHaltedForUiResponsiveness

Fields

location: Location

DuplicateStructFieldInSetFields

Fields

name: Ident

index: usize

previous_index: usize

CheckedTransmuteFailed

Fields

actual: Type

expected: Type

location: Location

StackOverflow

Fields

location: Location

call_stack: Vector<Location>

EvaluationDepthOverflow

Fields

location: Location

call_stack: Vector<Location>

Break

Continue

SkippedDueToEarlierErrors

Implementations

impl InterpreterError

pub(crate) fn should_be_filtered(&self) -> bool

Returns true if this error should be filtered out and not displayed to the user. This is used for internal control flow errors and errors that indicate the interpreter was skipped due to earlier errors that were already reported.

pub fn location(&self) -> Location

pub(crate) fn debug_evaluate_comptime( expr: impl Display, location: Location, ) -> Self