noirc_frontend::hir::comptime
Module interpreter

The comptime interpreter is a tree-walking interpreter used for evaluating Noir code at compile-time. It is typically triggered (by the elaborator) in one of four scenarios:

A comptime {} block

Everything in the block is interpreted

A macro call foo!()

The interpreter calls the function foo and inlines the resulting Quoted code at the callsite.

An attribute call #[my_attr] struct Foo {}

The interpreter calls the function my_attr and, if my_attr returns a Quoted value, inlines the resulting Quoted code.

A global global FOO = expr;

The interpreter evaluates expr to simplify the global to a constant.

This means any side-effects in expr will be performed at compile-time (!).

This may change in the future.

The interpreter operates on the HIR which only requires interpreted code to be elaborated before-hand, it does not need to be translated into another IR. Operating on high-level code like this makes the interpreter more predictable, hopefully limiting bugs, but does make it rather slow in practice.

Since unquoting macros may result in new variables in scope, the elaborator must run on that code after the interpreter is run. Yet the requirement that the interpreter runs on HIR means the interpreter must run in the middle of the elaborator. The usual flow is for the elaborator to elaborate as it goes, creating new HIR. Then when it sees a comptime {} block or other item that must be interpreted, it elaborates the entire item, creates and runs an Interpreter on it, inlines the result, and continues elaborating the rest of the code.

Also note that although it runs on code that has already been elaborated, in general it is still possible to invoke the interpreter on code which contains errors, such as type errors. For this reason, the interpreter must still perform error-checking at least in cases where we cannot continue otherwise. This can result in similar errors being issued for the same code. For example, a function’s body may fail to type check, but that same function may be called in the interpreter later on where we’d presumably halt with a similar error. InterpreterError::ArgumentCountMismatch is an example of such an error.

Modules

builtin 🔒

Large file containing implementations of all the various built-in functions which can be called in the interpreter. This notably includes the entire comptime-API defined in noir_stdlib/src/meta/*

cast 🔒

foreign 🔒

The foreign function counterpart to interpreter/builtin.rs, defines how to call all foreign functions available to the interpreter.

infix 🔒

unquote 🔒

Structs

Interpreter

Constants

MAX_EVALUATION_DEPTH 🔒

Maximum depth of evaluation, limiting recursion during comptime as well as expression depth. The goal is to be able to provide Noir stack traces if we run out, rather than a Rust backtrace.

Functions

bounds_check 🔒

Bounds check the given array and index pair. This will also ensure the given arguments are in fact an array and integer.

evaluate_integer 🔒

evaluate_prefix_with_value 🔒

force_bind_all 🔒

to_i128 🔒

to_u128 🔒

unbind_all 🔒