noirc_frontend::hir::comptime
Struct Interpreter

pub struct Interpreter<'local, 'interner> {
    pub elaborator: &'local mut Elaborator<'interner>,
    in_loop: bool,
    current_function: Option<FuncId>,
    bound_generics: Vec<FxHashMap<TypeVariable, (Type, Kind)>>,
    evaluation_depth: usize,
}

Fields
elaborator: &'local mut Elaborator<'interner>

To expand macros the Interpreter needs access to the Elaborator

in_loop: bool

True if the interpreter is currently in a loop (in the current function). Used only to error if break/continue are used outside a loop.

current_function: Option<FuncId>

The current function being interpreted. This may be None if we’re interpreting the rhs of a global.

bound_generics: Vec<FxHashMap<TypeVariable, (Type, Kind)>>

Maps each generic to the binding it has in the current callstack. Since the interpreter monomorphizes as it interprets, we can bind over the same generic multiple times. Without the outer Vec, when one of these inner functions exits we would unbind the generic completely instead of resetting it to its previous binding.

evaluation_depth: usize

Current evaluation depth.

Implementations

impl Interpreter<'_, '_>

pub(super) fn call_builtin( &mut self, name: &str, arguments: Vec<(Value, Location)>, return_type: Type, location: Location, ) -> Result<Value, InterpreterError>

impl Interpreter<'_, '_>

pub(super) fn call_foreign( &mut self, name: &str, arguments: Vec<(Value, Location)>, return_type: Type, location: Location, ) -> Result<Value, InterpreterError>

impl Interpreter<'_, '_>

pub(super) fn substitute_unquoted_values_into_tokens( &mut self, tokens: Tokens, ) -> Result<Vec<LocatedToken>, InterpreterError>

Evaluates any expressions within UnquoteMarkers in the given token list and replaces the expression held by the marker with the evaluated value in expression form.

impl<'local, 'interner> Interpreter<'local, 'interner>

pub(crate) fn new( elaborator: &'local mut Elaborator<'interner>, current_function: Option<FuncId>, ) -> Self

pub(crate) fn call_function( &mut self, function: FuncId, arguments: Vec<(Value, Location)>, instantiation_bindings: TypeBindings, location: Location, ) -> Result<Value, InterpreterError>

Call the given function with the given arguments and return the result.

This will handle internal details like binding generics and error handling. Note that running code which resulted in previous errors during elaboration may result in similar errors being issued again by the interpreter.

fn call_function_inner( &mut self, function: FuncId, arguments: Vec<(Value, Location)>, location: Location, ) -> Result<Value, InterpreterError>

Helper to check parameter count and dispatch on the function kind to run the function.

fn call_user_defined_function( &mut self, function: FuncId, arguments: Vec<(Value, Location)>, location: Location, ) -> Result<Value, InterpreterError>

Call a non-builtin function

fn get_function_body( &mut self, function: FuncId, location: Location, ) -> Result<ExprId, InterpreterError>

Try to retrieve a function’s body. If the function has not yet been resolved this will attempt to lazily resolve it. Afterwards, if the function’s body is still not known or the function is still in a Resolving state we issue an error.

fn elaborate_in_function<T>( &mut self, function: Option<FuncId>, reason: Option<ElaborateReason>, f: impl FnOnce(&mut Elaborator<'_>) -> T, ) -> T

Helper to elaborate the given item in the given function’s context. If None is passed, the global context is used. This function will temporarily unbind any generics from the previous function call if they exist.

fn elaborate_in_module<T>( &mut self, module: ModuleId, reason: Option<ElaborateReason>, f: impl FnOnce(&mut Elaborator<'_>) -> T, ) -> T

Run the given function with an elaborator in the context of the given module. Temporarily undoes any generics from the previous function.

fn call_special( &mut self, function: FuncId, arguments: Vec<(Value, Location)>, return_type: Type, location: Location, ) -> Result<Value, InterpreterError>

Calls a builtin, foreign, or oracle function (not all oracles are supported).

This will ignore any oracles starting with “__debug”

fn call_closure( &mut self, closure: Closure, arguments: Vec<(Value, Location)>, call_location: Location, ) -> Result<Value, InterpreterError>

Call a closure value with the given arguments and environment, returning the result.

fn call_closure_inner( &mut self, closure: HirLambda, environment: Vec<Value>, arguments: Vec<(Value, Location)>, call_location: Location, ) -> Result<Value, InterpreterError>

Performs the bulk of the work for calling a closure function. This function is very similar to Self::call_user_defined_function with the main difference being handling of closure.captures.

pub(super) fn enter_function( &mut self, ) -> (bool, Vec<HashMap<DefinitionId, Value>>)

Enters a function, pushing a new scope and resetting any required state. Returns the previous values of the internal state, to be reset when Self::exit_function is called.

pub(super) fn exit_function( &mut self, state: (bool, Vec<HashMap<DefinitionId, Value>>), )

Resets the per-function state to the value previously returned by Self::enter_function

pub(super) fn push_scope(&mut self)

Pushes a new scope to define any variables in.

Note that the first scope is always expected to be the global scope shared by all crates, which should never be popped.

pub(super) fn pop_scope(&mut self)

Pops the topmost scope.

Note that the first scope is expected to be the global scope of comptime values shared between all crates, which should never be popped.

fn current_scope_mut(&mut self) -> &mut HashMap<DefinitionId, Value>

Returns the current scope to define comptime variables in. The stack of scopes is always non-empty so this should never panic.

fn unbind_generics_from_previous_function(&mut self)

Unbinds all of the generics at the top of self.bound_generics, then push an empty set of bindings to become the new top of the stack.

fn rebind_generics_from_previous_function(&mut self)

Pops the top of self.bound_generics then takes the new bindings at the top of that stack and force-binds each.

fn remember_bindings( &mut self, main_bindings: &TypeBindings, impl_bindings: &TypeBindings, )

Adds all of the given main_bindings and impl_bindings to the top of self.bound_generics. Note that this will not actually perform any of the type bindings.

pub(super) fn define_pattern( &mut self, pattern: &HirPattern, typ: &Type, argument: Value, location: Location, ) -> Result<(), InterpreterError>

Defines a pattern, putting all variables contained within the pattern in the current scope.

fn define(&mut self, id: DefinitionId, argument: Value)

Define a new variable in the current scope

fn mutate( &mut self, id: DefinitionId, argument: Value, location: Location, ) -> Result<(), InterpreterError>

Mutate an existing variable, potentially from a prior scope

pub(super) fn lookup(&self, ident: &HirIdent) -> Result<Value, InterpreterError>

Lookup the comptime value of the given variable

pub fn lookup_id( &self, id: DefinitionId, location: Location, ) -> Result<Value, InterpreterError>

Lookup the comptime value of the given definition

pub fn evaluate(&mut self, id: ExprId) -> Result<Value, InterpreterError>

Evaluate an expression and return the result. This will automatically dereference a mutable variable if used.

fn evaluate_no_dereference( &mut self, id: ExprId, ) -> Result<Value, InterpreterError>

Evaluating a mutable variable will dereference it automatically. This function should be used when that is not desired - e.g. when compiling a &mut var expression to grab the original reference.

pub(super) fn evaluate_ident( &mut self, ident: HirIdent, id: ExprId, ) -> Result<Value, InterpreterError>

Evaluates a variable

fn evaluate_numeric_generic( &self, value: Type, expected: &Type, id: ExprId, ) -> Result<Value, InterpreterError>

Evaluates a numeric generic with the value value (expected to be Type::Constant) and an expected integer type expected.

fn evaluate_trait_item( &mut self, item: TraitItem, id: ExprId, ) -> Result<Value, InterpreterError>

fn evaluate_literal( &mut self, literal: HirLiteral, id: ExprId, ) -> Result<Value, InterpreterError>

fn evaluate_format_string( &mut self, fragments: Vec<FmtStrFragment>, captures: Vec<ExprId>, id: ExprId, ) -> Result<Value, InterpreterError>

Evaluates a format string. Note that in doing so, the string is formatted now, there is no delayed formatting when it is later used. Effectively the result is identical to a normal string, just with a different type. This is also why when format strings are lowered into runtime code they become regular strings - because they’re already formatted.

fn evaluate_integer( &self, value: SignedField, id: ExprId, ) -> Result<Value, InterpreterError>

Since integers are polymorphic, evaluating one requires the result type. We pass down the result type the elaborator previously inferred.

pub(crate) fn evaluate_block( &mut self, block: HirBlockExpression, ) -> Result<Value, InterpreterError>

fn evaluate_array( &mut self, array: HirArrayLiteral, id: ExprId, ) -> Result<Value, InterpreterError>

fn evaluate_vector( &mut self, array: HirArrayLiteral, id: ExprId, ) -> Result<Value, InterpreterError>

fn evaluate_prefix( &mut self, prefix: HirPrefixExpression, id: ExprId, ) -> Result<Value, InterpreterError>

fn evaluate_infix( &mut self, infix: HirInfixExpression, id: ExprId, ) -> Result<Value, InterpreterError>

fn evaluate_overloaded_infix( &mut self, infix: HirInfixExpression, lhs: Value, rhs: Value, id: ExprId, ) -> Result<Value, InterpreterError>

fn evaluate_overloaded_prefix( &mut self, prefix: HirPrefixExpression, rhs: Value, id: ExprId, ) -> Result<Value, InterpreterError>

fn evaluate_ordering( &self, ordering: Value, operator: BinaryOpKind, location: Location, ) -> Result<Value, InterpreterError>

Given the result of a cmp operation, convert it into the boolean result of the given operator.

fn evaluate_index( &mut self, index: HirIndexExpression, id: ExprId, ) -> Result<Value, InterpreterError>

fn evaluate_constructor( &mut self, constructor: HirConstructorExpression, id: ExprId, ) -> Result<Value, InterpreterError>

fn evaluate_enum_constructor( &mut self, constructor: HirEnumConstructorExpression, id: ExprId, ) -> Result<Value, InterpreterError>

Unlike a struct constructor, an enum constructor inserts a tag value along with the fields

fn evaluate_access( &mut self, access: HirMemberAccess, id: ExprId, ) -> Result<Value, InterpreterError>

fn get_field( &mut self, value: Value, id: ExprId, name: &String, ) -> Result<Shared<Value>, InterpreterError>

Given a value, return the struct/tuple field with the given name, automatically dereferencing any pointers found.

fn evaluate_call( &mut self, call: HirCallExpression, id: ExprId, ) -> Result<Value, InterpreterError>

Evaluates a call expression, deferring to Self::call_function or Self::call_closure once the function is determined.

fn unify_without_binding( &mut self, actual: &Type, expected: &Type, location: Location, )

This function is used by the interpreter for some comptime code which can change types e.g. on each iteration of a for loop.

fn evaluate_cast( &mut self, cast: &HirCastExpression, id: ExprId, ) -> Result<Value, InterpreterError>

fn evaluate_if( &mut self, if_: HirIfExpression, ) -> Result<Value, InterpreterError>

fn evaluate_tuple( &mut self, tuple: Vec<ExprId>, ) -> Result<Value, InterpreterError>

fn evaluate_lambda( &mut self, lambda: HirLambda, id: ExprId, ) -> Result<Value, InterpreterError>

fn evaluate_quote(&mut self, tokens: Tokens) -> Result<Value, InterpreterError>

pub fn evaluate_statement( &mut self, statement: StmtId, ) -> Result<Value, InterpreterError>

pub(crate) fn evaluate_let( &mut self, let_: HirLetStatement, ) -> Result<Value, InterpreterError>

fn evaluate_constrain( &mut self, constrain: HirConstrainExpression, ) -> Result<Value, InterpreterError>

fn evaluate_assign( &mut self, assign: HirAssignStatement, ) -> Result<Value, InterpreterError>

fn store_lvalue( &mut self, lvalue: HirLValue, rhs: Value, ) -> Result<(), InterpreterError>

Stores rhs at the location determined by lvalue

fn store_flattened(lvalue: Shared<Value>, rvalue: Value)

When we store to a struct such as in

let mut a = (false,);
let b = &mut a.0;
a = (true,);

we must flatten the store to store to each individual field so that any existing references, such as b above, will also reflect the mutation.

fn evaluate_lvalue( &mut self, lvalue: &HirLValue, ) -> Result<Value, InterpreterError>

Returns the current value held by lvalue

fn evaluate_for( &mut self, for_: HirForStatement, ) -> Result<Value, InterpreterError>

fn evaluate_for_loop<T>( &mut self, range_iterator: impl Iterator<Item = T>, get_index: fn(_: T) -> Value, index_id: DefinitionId, block: ExprId, ) -> Result<Value, InterpreterError>

fn evaluate_loop(&mut self, expr: ExprId) -> Result<Value, InterpreterError>

fn evaluate_while( &mut self, condition: ExprId, block: ExprId, ) -> Result<Value, InterpreterError>

fn evaluate_loop_body( &mut self, body: ExprId, result: &mut Result<Value, InterpreterError>, ) -> bool

Evaluate one iteration of a loop.

Returns a flag to indicate whether the loop should be exited.

fn evaluate_break(&mut self, id: StmtId) -> Result<Value, InterpreterError>

fn evaluate_continue(&mut self, id: StmtId) -> Result<Value, InterpreterError>

pub(super) fn evaluate_comptime( &mut self, statement: StmtId, ) -> Result<Value, InterpreterError>

fn print_oracle( &mut self, arguments: Vec<(Value, Location)>, ) -> Result<Value, InterpreterError>