noirc_frontend::hir::comptime
Module value

Defines the Value type, representing a compile-time value, used by the comptime interpreter when evaluating code.

Structs

Closure

Enums

ExprValue

TypedExpr

Value

A value representing the result of evaluating a Noir expression.

Functions

parse_tokens 🔒

Helper to parse the given tokens using the given parse function.

unwrap_rc 🔒

Unwraps an Rc value without cloning the inner value if the reference count is 1. Clones otherwise.

Type Aliases

StructFields 🔒