noirc_frontend::hir::comptime::value
Enum Value

pub enum Value {

    Unit,
    Bool(bool),
    Field(SignedField),
    I8(i8),
    I16(i16),
    I32(i32),
    I64(i64),
    U1(bool),
    U8(u8),
    U16(u16),
    U32(u32),
    U64(u64),
    U128(u128),
    String(Rc<String>),
    FormatString(Rc<String>, Type),
    CtString(Rc<String>),
    Function(FuncId, Type, Rc<TypeBindings>),
    Closure(Box<Closure>),
    Tuple(Vec<Shared<Value>>),
    Struct(FxHashMap<Rc<String>, Shared<Value>>, Type),
    Enum(usize, Vec<Value>, Type),
    Pointer(Shared<Value>, bool, bool),
    Array(Vector<Value>, Type),
    Slice(Vector<Value>, Type),
    Quoted(Rc<Vec<LocatedToken>>),
    TypeDefinition(TypeId),
    TraitConstraint(TraitId, TraitGenerics),
    TraitDefinition(TraitId),
    TraitImpl(TraitImplId),
    FunctionDefinition(FuncId),
    ModuleDefinition(ModuleId),
    Type(Type),
    Zeroed(Type),
    Expr(Box<ExprValue>),
    TypedExpr(TypedExpr),
    UnresolvedType(UnresolvedTypeData),

}

A value representing the result of evaluating a Noir expression.

Variants

Unit

Bool(bool)

Field(SignedField)

I8(i8)

I16(i16)

I32(i32)

I64(i64)

U1(bool)

U8(u8)

U16(u16)

U32(u32)

U64(u64)

U128(u128)

String(Rc<String>)

FormatString(Rc<String>, Type)

CtString(Rc<String>)

Function(FuncId, Type, Rc<TypeBindings>)

Closure(Box<Closure>)

Tuple(Vec<Shared<Value>>)

Tuple elements are automatically shared to support projection into a tuple: let elem = &mut tuple.0 should mutate the original element.

Struct(FxHashMap<Rc<String>, Shared<Value>>, Type)

Struct elements are automatically shared to support projection: let elem = &mut my_struct.field should mutate the original element.

Enum(usize, Vec<Value>, Type)

Pointer(Shared<Value>, bool, bool)

Array(Vector<Value>, Type)

Slice(Vector<Value>, Type)

Quoted(Rc<Vec<LocatedToken>>)

TypeDefinition(TypeId)

TraitConstraint(TraitId, TraitGenerics)

TraitDefinition(TraitId)

TraitImpl(TraitImplId)

FunctionDefinition(FuncId)

ModuleDefinition(ModuleId)

Type(Type)

Zeroed(Type)

Expr(Box<ExprValue>)

TypedExpr(TypedExpr)

UnresolvedType(UnresolvedTypeData)

Implementations

impl Value

pub fn display<'value, 'interner>( &'value self, interner: &'interner NodeInterner, ) -> ValuePrinter<'value, 'interner>

impl Value

pub(crate) fn expression(expr: ExpressionKind) -> Self

pub(crate) fn statement(statement: StatementKind) -> Self

pub(crate) fn lvalue(lvalue: LValue) -> Self

pub(crate) fn pattern(pattern: Pattern) -> Self

pub(crate) fn get_type(&self) -> Cow<'_, Type>

Retrieves the type of this value. Types can always be determined from the value, in cases where it would be ambiguous, Values store the type directly.

pub(crate) fn into_expression( self, elaborator: &mut Elaborator<'_>, location: Location, ) -> Result<Expression, InterpreterError>

Lowers this value into a runtime expression.

For literals this is often simple, e.g. Value::I8(3) translates to 3, but not all values are valid to lower. Certain values change form - e.g. format strings lowering into normal string literals. Lowering quoted code will simply return the quoted code (after parsing), this is how macros are implemented.

pub(crate) fn into_hir_expression( self, interner: &mut NodeInterner, location: Location, ) -> Result<ExprId, InterpreterError>

Lowers this compile-time value into a HIR expression. This is similar to Self::into_expression but is used in some cases in the monomorphizer where code must already be in HIR.

pub(crate) fn into_tokens( self, interner: &mut NodeInterner, location: Location, ) -> Result<Vec<LocatedToken>, InterpreterError>

Attempt to convert this value into a Vec of tokens representing this value if it appeared in source code. For example, Value::Unit is vec!['(', ')']. This is used for splicing values into quoted values when $ is used within a quote { } expression. Since Quoted code is represented as tokens, we need to convert the value into tokens.

pub(crate) fn is_integral(&self) -> bool

Returns false for non-integral Values.

pub(crate) fn is_zero(&self) -> bool

pub(crate) fn contains_function_or_closure(&self) -> bool

pub(crate) fn to_signed_field(&self) -> Option<SignedField>

Converts any integral Value into a SignedField. Returns None for non-integral Values.

pub(crate) fn into_top_level_items( self, location: Location, elaborator: &mut Elaborator<'_>, ) -> Result<Vec<Item>, InterpreterError>

Similar to Self::into_expression or Self::into_hir_expression but for converting into top-level item(s). Unlike those other methods, most expressions are invalid as top-level items (e.g. a lone 3 is not a valid top-level statement). As a result, this method is significantly simpler because we only have to parse Quoted values into top level items.

pub fn is_negative(&self) -> bool

pub(crate) fn move_struct(self) -> Value

Structs and tuples store references to their fields internally which need to be manually changed when moving them.

All references are shared by default but when we have let mut foo = Struct { .. } in code, we don’t want moving it: let mut bar = foo; to refer to the same references. This function will copy them so that mutating the fields of foo will not mutate bar.