noirc_frontend::hir::def_collector

Module dc_crate

Structs

CollectedItems

DefCollector

Given a Crate root, collect all definitions in that crate

ModuleAttribute

UnresolvedEnum

UnresolvedFunctions

Stores all of the unresolved functions in a particular file/mod

UnresolvedGlobal

UnresolvedStruct

UnresolvedTrait

UnresolvedTraitImpl

UnresolvedTypeAlias

Enums

CompilationError

Functions

add_import_reference 🔒

inject_prelude 🔒

Type Aliases

ImplMap 🔒

Maps the type and the module id in which the impl is defined to the functions contained in that impl along with the generics declared on the impl itself. This also contains the Span of the object_type of the impl, used to issue an error if the object type fails to resolve.