noirc_frontend::hir::def_collector::dc_crate

Type Alias ImplMap

pub(crate) type ImplMap = FxHashMap<(UnresolvedType, LocalModuleId), Vec<(UnresolvedGenerics, Location, UnresolvedFunctions)>>;

Maps the type and the module id in which the impl is defined to the functions contained in that impl along with the generics declared on the impl itself. This also contains the Span of the object_type of the impl, used to issue an error if the object type fails to resolve.

Note that because these are keyed by unresolved types, the impl map is one of the few instances of HashMap rather than BTreeMap. For this reason, we should be careful not to iterate over it since it would be non-deterministic.

Aliased Type

struct ImplMap { /* private fields */ }