noirc_frontend::hir::def_collector::dc_crate

Type Alias ImportDirectiveBatch

pub type ImportDirectiveBatch = Vec<ImportDirective>;

A batch of imports that resulted from desugaring a single import statement. This can happen as a result of expanding use foo::{bar, baz} into use foo:bar; use foo:baz;. Keeping the desugared imports as a batch is important because all imports in a batch will first be resolved, and just then those items will be brought in scope. This lets use foo::{foo, bar} work by first resolving foo::foo, then resolving foo::bar, then finally adding those items to scope. Compare this with use foo::foo; use foo::bar; where the second import will likely error as foo will mean foo::foo and not foo.

Aliased Type

pub struct ImportDirectiveBatch { /* private fields */ }