noirc_frontend::hir::def_collector

Module dc_mod

Structs

ModCollector 🔒

Given a module collect all definitions into ModuleData

Constants

CHOSEN_FIELD

Functions

check_duplicate_field_names 🔒

check_duplicate_variant_names 🔒

check_nargo_doc_primitive 🔒

If this submodule is an stdlib module that’s tagged with #['nargo_doc_primitive <primitive>], returns the primitive it refers to.

collect_defs

Walk a module and collect its definitions.

collect_enum

collect_function

collect_global 🔒

collect_impl

collect_struct

collect_trait_impl_items 🔒

Returns a tuple of (methods, associated types, associated constants)

find_module 🔒

is_native_field 🔒

push_child_module 🔒

Add a child module to the current def_map. On error this returns None and pushes to errors

should_check_siblings_for_module 🔒

Returns true if a module’s child modules are expected to be in the same directory. Returns false if they are expected to be in a subdirectory matching the name of the module.

Type Aliases

AssociatedConstants 🔒

AssociatedTypes 🔒