noirc_frontend::hir::def_collector::dc_mod
Struct ModCollector

struct ModCollector<'a> {
    pub(crate) def_collector: &'a mut DefCollector,
    pub(crate) file_id: FileId,
    pub(crate) module_id: LocalModuleId,
}

Given a module collect all definitions into ModuleData

Fields
def_collector: &'a mut DefCollectorfile_id: FileIdmodule_id: LocalModuleId
Implementations

impl ModCollector<'_>

fn collect_attributes( &mut self, attributes: Vec<SecondaryAttribute>, file_id: FileId, module_id: LocalModuleId, attribute_file_id: FileId, attribute_module_id: LocalModuleId, is_inner: bool, )

fn collect_globals( &mut self, context: &mut Context<'_, '_>, globals: Vec<(Documented<LetStatement>, ItemVisibility)>, crate_id: CrateId, ) -> Vec<CompilationError>

fn collect_impls( &mut self, context: &mut Context<'_, '_>, impls: Vec<TypeImpl>, krate: CrateId, ) -> Vec<CompilationError>

fn collect_trait_impls( &mut self, context: &mut Context<'_, '_>, impls: Vec<NoirTraitImpl>, krate: CrateId, ) -> Vec<CompilationError>

fn collect_functions( &mut self, context: &mut Context<'_, '_>, functions: Vec<Documented<NoirFunction>>, krate: CrateId, ) -> Vec<CompilationError>

fn collect_structs( &mut self, context: &mut Context<'_, '_>, types: Vec<Documented<NoirStruct>>, krate: CrateId, ) -> Vec<CompilationError>

Collect any struct definitions declared within the ast. Returns a vector of errors if any structs were already defined, or if a struct has duplicate fields in it.

fn collect_enums( &mut self, context: &mut Context<'_, '_>, types: Vec<Documented<NoirEnumeration>>, krate: CrateId, ) -> Vec<CompilationError>

Collect any enum definitions declared within the ast. Returns a vector of errors if any enums were already defined, or if an enum has duplicate variants in it.

fn collect_type_aliases( &mut self, context: &mut Context<'_, '_>, type_aliases: Vec<Documented<TypeAlias>>, krate: CrateId, ) -> Vec<CompilationError>

Collect any type aliases definitions declared within the ast. Returns a vector of errors if any type aliases were already defined.

fn collect_traits( &mut self, context: &mut Context<'_, '_>, traits: Vec<Documented<NoirTrait>>, krate: CrateId, ) -> Vec<CompilationError>

Collect any traits definitions declared within the ast. Returns a vector of errors if any traits were already defined.

fn collect_submodules( &mut self, context: &mut Context<'_, '_>, crate_id: CrateId, parent_module_id: LocalModuleId, submodules: Vec<Documented<SortedSubModule>>, file_id: FileId, ) -> Vec<CompilationError>

fn parse_module_declaration( &mut self, context: &mut Context<'_, '_>, mod_decl: Documented<ModuleDeclaration>, crate_id: CrateId, parent_file_id: FileId, parent_module_id: LocalModuleId, ) -> Vec<CompilationError>

Search for a module named mod_name Parse it, add it as a child to the parent module in which it was declared and then collect all definitions of the child module

fn push_child_module( &mut self, context: &mut Context<'_, '_>, mod_name: &Ident, visibility: ItemVisibility, mod_location: Location, outer_attributes: Vec<SecondaryAttribute>, inner_attributes: Vec<SecondaryAttribute>, add_to_parent_scope: bool, is_contract: bool, is_type: bool, ) -> Result<ModuleId, DefCollectorErrorKind>

Add a child module to the current def_map. On error this returns None and pushes to errors

fn resolve_associated_constant_type( &self, typ: Option<&UnresolvedType>, errors: &mut Vec<CompilationError>, ) -> Type