noirc_frontend::hir
Module def_map

Modules

item_scope 🔒

module_data 🔒

module_def 🔒

namespace 🔒

Structs

CrateDefMap

Map of all modules and scopes defined within a crate.

FuzzingHarness

ItemScope

All the definitions of Idents in scope, either as types or values.

LocalModuleId

Lets first check if this is offered by any external crate XXX: RA has made this a crate on crates.io

ModuleData

Contains the actual contents of a module: its parent (if one exists), children, and scope with all definitions defined within the scope.

ModuleId

PerNs

Result of looking up a name in type and value definitions in scope.

TestFunction

Enums

ModuleDefId

A generic ID that references either a module, function, type, interface or global

Constants

MAIN_FUNCTION

The name that is used for a non-contract program’s entry-point function.

Functions

fully_qualified_module_path

parse_file

Given a FileId, fetch the File, from the FileManager and parse it’s content

Type Aliases

DefMaps