noirc_frontend::hir::def_map::module_data
Struct ModuleData

pub struct ModuleData {
    pub parent: Option<LocalModuleId>,
    pub children: HashMap<Ident, LocalModuleId>,
    pub child_declaration_order: Vec<LocalModuleId>,
    scope: ItemScope,
    definitions: ItemScope,
    traits_in_scope: HashMap<TraitId, Ident>,
    pub location: Location,
    pub is_contract: bool,
    pub is_type: bool,
    pub attributes: Vec<SecondaryAttribute>,
}

Contains the actual contents of a module: its parent (if one exists), children, and scope with all definitions defined within the scope.

Fields
parent: Option<LocalModuleId>children: HashMap<Ident, LocalModuleId>child_declaration_order: Vec<LocalModuleId>

Each child in the order they were declared in the parent module. E.g. for a module containing mod foo; mod bar; mod baz this would be vec![foo, bar, baz].

scope: ItemScope

Contains all definitions visible to the current module. This includes all definitions in self.definitions as well as all imported definitions.

definitions: ItemScope

Contains only the definitions directly defined in the current module

traits_in_scope: HashMap<TraitId, Ident>

All traits in scope, either from use imports or trait declarations. The Ident value is the trait name or the use alias, if any. This is stored separately from scope to quickly check if a trait is in scope.

location: Locationis_contract: bool

True if this module is a contract Foo { ... } module containing contract functions

is_type: bool

True if this module is actually a type

attributes: Vec<SecondaryAttribute>
Implementations

impl ModuleData

pub fn new( parent: Option<LocalModuleId>, location: Location, outer_attributes: Vec<SecondaryAttribute>, inner_attributes: Vec<SecondaryAttribute>, is_contract: bool, is_type: bool, ) -> ModuleData

pub fn scope(&self) -> &ItemScope

pub fn definitions(&self) -> &ItemScope

fn declare( &mut self, name: Ident, visibility: ItemVisibility, item_id: ModuleDefId, trait_id: Option<TraitId>, ) -> Result<(), (Ident, Ident)>

pub fn declare_function( &mut self, name: Ident, visibility: ItemVisibility, id: FuncId, ) -> Result<(), (Ident, Ident)>

pub fn declare_trait_function( &mut self, name: Ident, id: FuncId, trait_id: TraitId, ) -> Result<(), (Ident, Ident)>

pub fn remove_function(&mut self, name: &Ident)

pub fn declare_global( &mut self, name: Ident, visibility: ItemVisibility, id: GlobalId, ) -> Result<(), (Ident, Ident)>

pub fn declare_type( &mut self, name: Ident, visibility: ItemVisibility, id: TypeId, ) -> Result<(), (Ident, Ident)>

pub fn declare_type_alias( &mut self, name: Ident, visibility: ItemVisibility, id: TypeAliasId, ) -> Result<(), (Ident, Ident)>

pub fn declare_trait( &mut self, name: Ident, visibility: ItemVisibility, id: TraitId, ) -> Result<(), (Ident, Ident)>

pub fn declare_trait_associated_type( &mut self, name: Ident, id: TraitAssociatedTypeId, ) -> Result<(), (Ident, Ident)>

pub fn declare_child_module( &mut self, name: Ident, visibility: ItemVisibility, child_id: ModuleId, ) -> Result<(), (Ident, Ident)>

pub fn find_func_with_name(&self, name: &Ident) -> Option<FuncId>

pub fn import( &mut self, name: Ident, visibility: ItemVisibility, id: ModuleDefId, is_prelude: bool, ) -> Result<(), (Ident, Ident)>

pub fn find_name(&self, name: &Ident) -> PerNs

Find an Ident in the types and values in scope.

Returns the preferred, unambiguous result in both.

pub fn find_trait_in_scope(&self, trait_id: TraitId) -> Option<&Ident>

Finds a trait in scope and returns its name (either the trait name, or a use alias if it was brought to scope like that)

pub fn type_definitions(&self) -> impl Iterator<Item = ModuleDefId> + '_

pub fn value_definitions(&self) -> impl Iterator<Item = ModuleDefId> + '_

Return an iterator over all definitions defined within this module, excluding any type definitions.