noirc_frontend::hir::printer
Module items

This module is responsible for building a list of items that represent the final code that is going to be monomorphized and turned into SSA. This final code has all macros expanded and is mainly gathered from data inside NodeInterner, modules in DefMaps and function bodies from HirExpressions.

Modules

hir_def 🔒

types 🔒

Structs

DataType

Impl

Import

ItemBuilder 🔒

Module

PrimitiveType

Trait

TraitImpl

Enums

Item

Functions

gather_named_type_vars 🔒

type_mentions_data_type 🔒

type_mentions_primitive_type 🔒