noirc_frontend::hir::printer
Struct ItemPrinter

struct ItemPrinter<'context, 'string> {
    crate_id: CrateId,
    interner: &'context NodeInterner,
    def_maps: &'context DefMaps,
    dependencies: &'context Vec<Dependency>,
    string: &'string mut String,
    indent: usize,
    module_id: ModuleId,
    imports: HashMap<ModuleDefId, Ident>,
    self_type: Option<Type>,
    trait_constraints: Vec<TraitConstraint>,
    trait_impls_printed: HashSet<TraitImplId>,
}

Fields
crate_id: CrateIdinterner: &'context NodeInternerdef_maps: &'context DefMapsdependencies: &'context Vec<Dependency>string: &'string mut Stringindent: usizemodule_id: ModuleIdimports: HashMap<ModuleDefId, Ident>self_type: Option<Type>trait_constraints: Vec<TraitConstraint>

Trait constraints in scope. These are set when a trait, trait impl or function is visited.

trait_impls_printed: HashSet<TraitImplId>

Keep track of trait impls that have been printed so we don’t show a same trait impl multiple times.

Implementations

impl ItemPrinter<'_, '_>

fn show_hir_expression_id(&mut self, expr_id: ExprId)

fn dereference_hir_expression_id(&self, expr_id: ExprId) -> ExprId

fn show_hir_expression_id_maybe_inside_parens(&mut self, expr_id: ExprId)

fn show_hir_expression_id_maybe_inside_curly_braces(&mut self, expr_id: ExprId)

pub(crate) fn show_hir_expression( &mut self, hir_expr: HirExpression, expr_id: ExprId, )

pub(crate) fn show_hir_lambda(&mut self, hir_lambda: HirLambda)

fn show_hir_match(&mut self, hir_match: HirMatch)

fn show_constructor(&mut self, constructor: Constructor)

fn try_show_hir_call_as_method( &mut self, hir_call_expression: &HirCallExpression, ) -> bool

fn show_hir_call_as_trait_as_path( &mut self, hir_call_expression: &HirCallExpression, arguments: &[ExprId], generics: Option<Vec<Type>>, func_id: FuncId, trait_method: &TraitItem, )

fn show_hir_block_expression(&mut self, block: HirBlockExpression)

fn show_hir_expression_ids_separated_by_comma(&mut self, expr_ids: &[ExprId])

fn show_hir_statement_id(&mut self, stmt_id: StmtId)

fn show_hir_statement(&mut self, statement: HirStatement)

fn show_hir_literal(&mut self, literal: HirLiteral, expr_id: ExprId)

fn show_hir_array_literal(&mut self, array: HirArrayLiteral)

fn show_hir_lvalue(&mut self, lvalue: HirLValue)

fn show_hir_pattern(&mut self, pattern: HirPattern)

fn show_definition_id(&mut self, definition_id: DefinitionId)

fn show_hir_ident(&mut self, ident: HirIdent, expr_id: Option<ExprId>)

fn statement_id_has_unsafe(&self, stmt_id: StmtId) -> bool

fn statement_has_unsafe(&self, statement: &HirStatement) -> bool

fn expression_id_has_unsafe(&self, expr_id: ExprId) -> bool

fn expression_has_unsafe(&self, expr: HirExpression) -> bool

fn hir_match_has_unsafe(&self, hir_match: &HirMatch) -> bool

impl ItemPrinter<'_, '_>

pub(crate) fn show_types_separated_by_comma(&mut self, types: &[Type])

pub(crate) fn show_type(&mut self, typ: &Type)

pub(super) fn show_type_as_expression(&mut self, typ: &Type)

fn show_type_impl(&mut self, typ: &Type, as_expression: bool)

fn show_type_maybe_in_parentheses(&mut self, typ: &Type)

impl<'context, 'string> ItemPrinter<'context, 'string>

pub(super) fn new( crate_id: CrateId, interner: &'context NodeInterner, def_maps: &'context DefMaps, dependencies: &'context Vec<Dependency>, string: &'string mut String, ) -> Self

pub(super) fn show_item(&mut self, item: Item)

fn show_module(&mut self, module: Module)

fn show_item_with_visibility(&mut self, item: Item, visibility: ItemVisibility)

fn show_doc_comments(&mut self, reference_id: ReferenceId)

fn show_module_def_id_attributes(&mut self, module_def_id: ModuleDefId)

fn show_secondary_attributes(&mut self, attributes: &[SecondaryAttribute])

fn show_item_visibility(&mut self, visibility: ItemVisibility)

fn show_visibility(&mut self, visibility: Visibility)

fn show_data_type(&mut self, item_data_type: DataType)

fn show_struct(&mut self, data_type: &DataType)

fn show_enum(&mut self, data_type: &DataType)

fn show_data_type_impls(&mut self, impls: Vec<Impl>)

fn show_impl(&mut self, impl_: Impl)

fn show_trait_impls(&mut self, trait_impls: &[&TraitImpl])

fn show_type_alias(&mut self, type_alias_id: TypeAliasId)

fn show_trait(&mut self, item_trait: Trait)

fn show_trait_impl(&mut self, item_trait_impl: &TraitImpl)

fn show_global(&mut self, global_id: GlobalId)

fn show_function(&mut self, func_id: FuncId)

fn show_generic_types(&mut self, types: &[Type], use_colons: bool)

fn show_generics(&mut self, generics: &Generics)

fn show_generic_kind(&mut self, name: &str, kind: &Kind)

fn show_trait_generics(&mut self, generics: &TraitGenerics)

fn show_generic_type_variables(&mut self, generics: &BTreeSet<(String, Kind)>)

fn show_where_clause(&mut self, constraints: &[TraitConstraint])

fn show_trait_bound(&mut self, bound: &ResolvedTraitBound)

fn show_trait_bounds(&mut self, bounds: &[ResolvedTraitBound])

fn show_pattern(&mut self, pattern: &HirPattern)

fn show_value(&mut self, value: &Value)

fn show_type_name_as_data_type(&mut self, typ: &Type)

fn show_imports(&mut self, imports: Vec<Import>)

fn show_reference_to_module_def_id( &mut self, module_def_id: ModuleDefId, visibility: ItemVisibility, use_import: bool, ) -> String

fn show_quoted(&mut self, tokens: &[LocatedToken])

fn pattern_is_self(&self, pattern: &HirPattern) -> bool

fn pattern_is_self_or_underscore_self(&self, pattern: &HirPattern) -> bool

fn module_def_id_name(&self, module_def_id: ModuleDefId) -> String

fn module_def_id_visibility(&self, module_def_id: ModuleDefId) -> ItemVisibility

fn show_separated_by_comma<Item, F>(&mut self, items: &[Item], f: F)
where F: Fn(&mut Self, &Item),

fn increase_indent(&mut self)

fn decrease_indent(&mut self)

fn write_indent(&mut self)

fn push_str(&mut self, str: &str)

fn push(&mut self, char: char)