noirc_frontend::hir::resolution
Module import

Structs

ImportDirective

ImportSolver 🔒

PathResolutionTargetResolver 🔒

ResolvedImport

Enums

PathResolutionError

Functions

get_module 🔒

path_segment_to_typed_path_segment 🔒

path_to_typed_path 🔒

resolve_import

Resolves a Path in a use statement, assuming it’s located in importing_module.

resolve_path_kind

Given a Path and a ModuleId it’s being used in, this function returns a plain Path and a ModuleId where that plain Path should be resolved. That is, this method will resolve the Path kind and translate it to a plain path.

Type Aliases

ImportResolutionResult 🔒