noirc_frontend::hir::resolution::import
Function resolve_import

pub fn resolve_import(
    path: Path,
    importing_module: ModuleId,
    def_maps: &DefMaps,
    usage_tracker: &mut UsageTracker,
    references_tracker: Option<ReferencesTracker<'_>>,
) -> Result<ResolvedImport, PathResolutionError>

Resolves a Path in a use statement, assuming it’s located in importing_module.

If the imported name can’t be found, Err will be returned. If it can be found, Ok will be returned with a potential list of errors if, for example, one of the segments is not accessible from the importing module (e.g. because it’s private).