noirc_frontend::hir::resolution::import
Function resolve_path_kind

pub fn resolve_path_kind<'r>(
    path: TypedPath,
    importing_module: ModuleId,
    def_maps: &DefMaps,
    references_tracker: Option<ReferencesTracker<'r>>,
) -> Result<(TypedPath, ModuleId, Option<ReferencesTracker<'r>>), PathResolutionError>

Given a TypedPath and a ModuleId it’s being used in, this function returns a TypedPath and a ModuleId where that TypedPath should be resolved.

For a PathKind::Dep with a value such as dep::foo::bar::baz, the path will be turned into a PathKind::Plain with the first segment (the crate foo) removed, leaving just bar::baz to be resolved within foo. For other cases the path kind stays the same, it’s just paired up with the module where it should be looked up. If the module cannot be found, and error is returned.

The third value in the tuple is a reference tracker that must be passed to this method, which is used in case the path kind is dep: the segment after dep will be linked to the root module of the external dependency.