noirc_frontend::hir::resolution

Module visibility

Functions

is_same_type_regardless_generics 🔒

Returns whether two types are the same disregarding their generic arguments.

item_in_module_is_visible

Returns true if an item with the given visibility in the target module is visible from the current module. For example:

method_call_is_visible

Returns whether a method call func_id on an object of type object_type is visible from current_module.

module_def_id_visibility

module_is_descendant_of_target 🔒

Returns true if current is a (potentially nested) child module of target. This is also true if current == target.

module_is_parent_of_struct_module 🔒

Returns true if target is a struct and its parent is current.

struct_member_is_visible

Returns whether a struct member with the given visibility is visible from current_module_id.

trait_member_is_visible

Returns whether a trait member with the given visibility is visible from current_module_id.

type_member_is_visible 🔒

Returns whether a struct or trait member with the given visibility is visible from current_module_id.