noirc_frontend::hir

Module scope

Structs

Scope

This implementation uses the terminology Scope and ScopeTree A Scope is map between Keys and Values, it is possible to search for a Key in Scope, returning a mutable copy A ScopeTree is a Vector of Scopes, searching for Key in ScopeTree will recursively search from the last scope until the first scope to find the Key, returning None, if the Key is in None of the Scopes. A ScopeTree is useful for implementing the scoping logic around for-loops, if statements and ClosureCalls. Implementers will usually store a Vector of ScopeTrees to implement the logic needed for FunctionCalls

ScopeForest

ScopeTree

ScopeTree allows one to specify that maps within the same vector are scope extensions This allows one to extend the scope and then remove the extension, without affecting the unextended part of the scope