noirc_frontend::hir
Struct Context

pub struct Context<'file_manager, 'parsed_files> {
    pub def_interner: NodeInterner,
    pub crate_graph: CrateGraph,
    pub def_maps: BTreeMap<CrateId, CrateDefMap>,
    pub usage_tracker: UsageTracker,
    pub file_manager: Cow<'file_manager, FileManager>,
    pub debug_instrumenter: DebugInstrumenter,
    pub visited_files: BTreeMap<FileId, Location>,
    pub parsed_files: Cow<'parsed_files, ParsedFiles>,
    pub package_build_path: PathBuf,
    pub interpreter_output: Option<Rc<RefCell<dyn Write>>>,
    pub required_unstable_features: BTreeMap<CrateId, Vec<UnstableFeature>>,
}

Helper object which groups together several useful context objects used during name resolution. Once name resolution is finished, only the def_interner is required for type inference and monomorphization.

Fields
def_interner: NodeInternercrate_graph: CrateGraphdef_maps: BTreeMap<CrateId, CrateDefMap>usage_tracker: UsageTrackerfile_manager: Cow<'file_manager, FileManager>debug_instrumenter: DebugInstrumentervisited_files: BTreeMap<FileId, Location>

A map of each file that already has been visited from a prior mod foo; declaration. This is used to issue an error if a second mod foo; is declared to the same file.

parsed_files: Cow<'parsed_files, ParsedFiles>package_build_path: PathBufinterpreter_output: Option<Rc<RefCell<dyn Write>>>

Writer for comptime prints.

required_unstable_features: BTreeMap<CrateId, Vec<UnstableFeature>>

Any unstable features required by the current package or its dependencies.

Implementations

impl Context<'_, '_>

pub fn interpret_function( &mut self, main_id: FuncId, args: Vec<(Value, Location)>, ) -> Result<Value, InterpreterError>

Interprets (as comptime code) the given function in the give crate, with the given arguments. Panics if there’s no main function.

impl Context<'_, '_>

pub fn new( file_manager: FileManager, parsed_files: ParsedFiles, ) -> Context<'static, 'static>

pub fn from_ref_file_manager<'file_manager, 'parsed_files>( file_manager: &'file_manager FileManager, parsed_files: &'parsed_files ParsedFiles, ) -> Context<'file_manager, 'parsed_files>

pub fn parsed_file_results( &self, file_id: FileId, ) -> (ParsedModule, Vec<ParserError>)

pub fn def_map(&self, crate_id: &CrateId) -> Option<&CrateDefMap>

Returns the CrateDefMap for a given CrateId. It is perfectly valid for the compiler to look up a CrateDefMap and it is not available. This is how the compiler knows to compile a Crate.

pub fn def_map_mut(&mut self, crate_id: &CrateId) -> Option<&mut CrateDefMap>

pub fn crates(&self) -> impl Iterator<Item = CrateId> + '_

Return the CrateId for each crate that has been compiled successfully

pub fn root_crate_id(&self) -> &CrateId

pub fn stdlib_crate_id(&self) -> &CrateId

pub fn function_name(&self, id: &FuncId) -> &str

pub fn fully_qualified_function_name( &self, crate_id: &CrateId, id: &FuncId, ) -> String

pub fn fully_qualified_struct_path( &self, crate_id: &CrateId, id: TypeId, ) -> String

Returns a fully-qualified path to the given TypeId from the given CrateId. This function also account for the crate names of dependencies.

For example, if you project contains a main.nr and foo.nr and you provide the main_crate_id and the bar_struct_id where the Bar struct is inside foo.nr, this function would return foo::Bar as a String.

pub fn function_meta(&self, func_id: &FuncId) -> &FuncMeta

pub fn get_main_function(&self, crate_id: &CrateId) -> Option<FuncId>

Returns the FuncId of the ‘main’ function in a crate.

Expects check_crate to be called beforehand

Panics if no main function is found

pub fn get_all_test_functions_in_crate_matching( &self, crate_id: &CrateId, pattern: &FunctionNameMatch, ) -> Vec<(String, TestFunction)>

Returns a list of all functions in the current crate marked with #test whose names contain the given pattern string. An empty pattern string will return all functions marked with #test.

pub fn get_all_fuzzing_harnesses_in_crate_matching( &self, crate_id: &CrateId, pattern: &FunctionNameMatch, ) -> Vec<(String, FuzzingHarness)>

Returns a list of all functions in the current crate marked with #[fuzz] whose names contain the given pattern string. An empty pattern string will return all functions marked with #[fuzz].

pub fn get_all_exported_functions_in_crate( &self, crate_id: &CrateId, ) -> Vec<(String, FuncId)>

pub fn module(&self, module_id: ModuleId) -> &ModuleData

pub(crate) fn resolve_generics( interner: &NodeInterner, generics: &UnresolvedGenerics, errors: &mut Vec<CompilationError>, ) -> Generics

Generics need to be resolved before elaboration to distinguish between normal and numeric generics. This method is expected to be used during definition collection. Each result is returned in a list rather than returned as a single result as to allow definition collection to provide an error for each ill-formed numeric generic.

pub fn crate_files(&self, crate_id: &CrateId) -> HashSet<FileId>

pub fn activate_lsp_mode(&mut self)

Activates LSP mode, which will track references for all definitions.

pub fn disable_comptime_printing(&mut self)

pub fn set_comptime_printing(&mut self, output: Rc<RefCell<dyn Write>>)