noirc_frontend::hir::type_check::generics
Trait Generic

pub trait Generic {
    // Required methods
    fn item_kind(&self) -> &'static str;
    fn item_name(&self, interner: &NodeInterner) -> String;
    fn generic_kinds(&self, interner: &NodeInterner) -> Vec<Kind>;
    fn accepts_named_type_args(&self) -> bool;
    fn named_generics(&self, interner: &NodeInterner) -> Vec<ResolvedGeneric>;
}

Represents something that can be generic over type variables such as a trait, struct type, or type alias.

Used primarily by Elaborator::resolve_type_args so that we can have one function to do this for struct types, type aliases, traits, etc.

Required Methods

fn item_kind(&self) -> &'static str

The name of this kind of item, for error messages. E.g. “trait”, “struct type”.

fn item_name(&self, interner: &NodeInterner) -> String

The name of this item, usually named by a user. E.g. “Foo” for “struct Foo {}”

fn generic_kinds(&self, interner: &NodeInterner) -> Vec<Kind>

Each ordered generic kind on this type, excluding any named generics.

fn accepts_named_type_args(&self) -> bool

True if this item kind can ever accept named type arguments. Currently, this is only true for traits. Structs & aliases can never have named args.

fn named_generics(&self, interner: &NodeInterner) -> Vec<ResolvedGeneric>