noirc_frontend::hir_def

Module expr

Structs

Case

HirBinaryOp

HirBlockExpression

HirCallExpression

HirCapturedVar

A variable captured inside a closure

HirCastExpression

HirConstrainExpression

Corresponds to assert and assert_eq in the source code. This node also contains the FileId of the file the constrain originates from. This is used later in the SSA pass to issue an error if a constrain is found to be always false.

HirConstructorExpression

HirEnumConstructorExpression

An enum constructor is an expression such as Option::Some(foo) to construct an enum. These are usually inserted by the compiler itself since Some is actually a function with the body implicitly being an enum constructor expression, but in the future these may be directly represented when using enums with named fields.

HirIdent

Corresponds to a variable in the source code

HirIfExpression

HirIndexExpression

Indexing, as in array[index]

HirInfixExpression

HirLambda

HirMemberAccess

This is always a struct field access my_struct.field and never a method call. The later is represented by HirMethodCallExpression.

HirMethodCallExpression

These nodes are temporary, they’re lowered into HirCallExpression nodes after type checking resolves the object type and the method it calls.

HirPrefixExpression

HirTraitMethodReference

TraitItem

A method or constant defined in a trait

Enums

Constructor

HirArrayLiteral

HirExpression

A HirExpression is the result of an Expression in the AST undergoing name resolution. It is almost identical to the Expression AST node, but references other HIR nodes indirectly via IDs rather than directly via boxing. Variables in HirExpressions are tagged with their DefinitionId from the definition that refers to them so there is no ambiguity with names.

HirLiteral

HirMatch

HirMethodReference

ImplKind