noirc_frontend::hir_def
Module function

Structs

FuncMeta

A FuncMeta contains the signature of the function and any associated meta data like the function’s Location, FunctionKind, and attributes. If the function’s body is needed, it can be retrieved separately via NodeInterner::function(&self, &FuncId).

HirFunction

A Hir function is a block expression with a list of statements. If the function has yet to be resolved, the body starts off empty (None).

Parameters

Enums

FunctionBody

Type Aliases

FunctionSignature

Param

An interned function parameter from a function definition