noirc_frontend::hir_def::function
Struct FuncMeta

pub struct FuncMeta {

    pub name: HirIdent,
    pub kind: FunctionKind,
    pub parameters: Parameters,
    pub parameter_idents: Vec<HirIdent>,
    pub return_type: FunctionReturnType,
    pub return_visibility: Visibility,
    pub typ: Type,
    pub direct_generics: Vec<ResolvedGeneric>,
    pub all_generics: Vec<ResolvedGeneric>,
    pub location: Location,
    pub has_body: bool,
    pub trait_constraints: Vec<TraitConstraint>,
    pub extra_trait_constraints: Vec<TraitConstraint>,
    pub type_id: Option<TypeId>,
    pub trait_id: Option<TraitId>,
    pub trait_impl: Option<TraitImplId>,
    pub enum_variant_index: Option<usize>,
    pub is_entry_point: bool,
    pub has_inline_attribute: bool,
    pub function_body: FunctionBody,
    pub source_crate: CrateId,
    pub source_module: LocalModuleId,
    pub source_file: FileId,
    pub self_type: Option<Type>,

}

A FuncMeta contains the signature of the function and any associated meta data like the function’s Location, FunctionKind, and attributes. If the function’s body is needed, it can be retrieved separately via NodeInterner::function(&self, &FuncId).

Fields
name: HirIdentkind: FunctionKindparameters: Parametersparameter_idents: Vec<HirIdent>

The HirIdent of each identifier within the parameter list. Note that this includes separate entries for each identifier in e.g. tuple patterns.

return_type: FunctionReturnTypereturn_visibility: Visibilitytyp: Type

The type of this function. Either a Type::Function or a Type::Forall for generic functions.

direct_generics: Vec<ResolvedGeneric>

The set of generics that are declared directly on this function in the source code. This does not include generics from an outer scope, like those introduced by an impl<T> block. This also does not include implicit generics added by the compiler such as a trait’s Self type variable.

all_generics: Vec<ResolvedGeneric>

All the generics used by this function, which includes any implicit generics or generics from outer scopes, such as those introduced by an impl. This is stored when the FuncMeta is first created to later be used to set the current generics when the function’s body is later resolved.

location: Locationhas_body: booltrait_constraints: Vec<TraitConstraint>

Trait constraints that were specified directly on this function.

extra_trait_constraints: Vec<TraitConstraint>

Trait constraints that came either from a parent item (for example a where clause on a trait or trait impl) or from constraints on implicitly added named generics.

type_id: Option<TypeId>

The type this method belongs to, if any

trait_id: Option<TraitId>trait_impl: Option<TraitImplId>

The trait impl this function belongs to, if any

enum_variant_index: Option<usize>

If this function is the one related to an enum variant, this holds its index (relative to type_id)

is_entry_point: bool

True if this function is an entry point to the program. For non-contracts, this means the function is main.

has_inline_attribute: bool

True if this function is marked with an attribute that indicates it should be inlined differently than the default (inline everything). For example, such as fold (never inlined) or no_predicates (inlined after flattening)

function_body: FunctionBodysource_crate: CrateId

The crate this function was defined in

source_module: LocalModuleId

The module this function was defined in

source_file: FileId

THe file this function was defined in

self_type: Option<Type>

If this function is from an impl (trait or regular impl), this is the object type of the impl. Otherwise this is None.

Implementations

impl FuncMeta

pub fn is_stub(&self) -> bool

A stub function does not have a body. This includes Builtin, LowLevel, and Oracle functions in addition to method declarations within a trait without a body.

We don’t check the return type of these functions since it will always have an empty body, and we don’t check for unused parameters.

pub fn function_signature(&self) -> FunctionSignature

pub fn return_type(&self) -> &Type

Gives the (uninstantiated) return type of this function.

pub fn take_body(&mut self) -> FunctionBody

Take this function body, returning an owned version while avoiding cloning any large Expressions inside by replacing a Unresolved with a Resolving variant.

pub fn all_trait_constraints(&self) -> impl Iterator<Item = &TraitConstraint>