noirc_frontend::hir_def::traits
Struct TraitImpl

pub struct TraitImpl {
    pub ident: Ident,
    pub location: Location,
    pub typ: Type,
    pub trait_id: TraitId,
    pub trait_generics: Vec<Type>,
    pub file: FileId,
    pub crate_id: CrateId,
    pub methods: Vec<FuncId>,
    pub where_clause: Vec<TraitConstraint>,
}

Fields
ident: Identlocation: Locationtyp: Typetrait_id: TraitIdtrait_generics: Vec<Type>

Any ordered type arguments on the trait this impl is for. E.g. A, B in impl Foo<A, B, C = D> for Bar

Note that named arguments (associated types) are stored separately in the NodeInterner. This is because they’re required to resolve types before the impl as a whole is finished resolving.

file: FileIdcrate_id: CrateIdmethods: Vec<FuncId>where_clause: Vec<TraitConstraint>

The where clause, if present, contains each trait requirement which must be satisfied for this impl to be selected. E.g. in impl Eq for [T] where T: Eq, where_clause would contain the one T: Eq constraint. If there is no where clause, this Vec is empty.