noirc_frontend::hir_def

Module types

Modules

arithmetic 🔒

recursion 🔒

unification 🔒

validity 🔒

Structs

DataType

Represents a struct or enum type in the type system. Each instance of this rust struct will be shared across all Type::DataType variants that represent the same struct or enum type.

EnumVariant

NamedGeneric

QuotedTypeIter

An iterator over the variants of Self

ResolvedGeneric

Shared

A shared, mutable reference to some T. Wrapper is required for Hash impl of RefCell.

StructField

TraitAssociatedType

TypeAlias

Wrap around an unsolved type

TypeVariable

A TypeVariable is a mutable reference that is either bound to some type, or unbound with a given TypeVariableId.

TypeVariableId

A unique ID used to differentiate different type variables

UnificationError

Enums

BinaryTypeOperator

A restricted subset of binary operators useable on type level integers for use in the array length positions of types.

Kind

A Kind is the type of a Type. These are used since only certain kinds of types are allowed in certain positions.

MustUse

Convenience enum to avoid using Option<Option<String>> to indicate whether #[must_use] is present (outer option) and the optional message (inner option).

QuotedType

Type

TypeBinding

TypeBindings are the mutable insides of a TypeVariable. They are either bound to some type, or are unbound.

TypeBody 🔒

Constants

TYPE_RECURSION_LIMIT

Arbitrary recursion limit when following type variables or recurring on types some other way. Types form trees but are not likely to be more deep than just a few levels in real code.

Functions

type_bindings_to_string

Pretty print type bindings for debugging

Type Aliases

GenericTypeVars

Corresponds to generic lists such as <T, U> in the source program. Used mainly for resolved types which no longer need information such as names or kinds

ResolvedGenerics

Corresponds to generic lists such as <T, U> with additional information gathered during name resolution that is necessary correctly resolving types.

TypeBindings

A list of (TypeVariableId, Kind)’s to bind to a type. Storing the TypeVariable in addition to the matching TypeVariableId allows the binding to later be undone if needed.