noirc_frontend::hir_def::types
Enum Type

pub enum Type {

    FieldElement,
    Array(Box<Type>, Box<Type>),
    Slice(Box<Type>),
    Integer(Signedness, IntegerBitSize),
    Bool,
    String(Box<Type>),
    FmtString(Box<Type>, Box<Type>),
    Unit,
    Tuple(Vec<Type>),
    DataType(Shared<DataType>, Vec<Type>),
    Alias(Shared<TypeAlias>, Vec<Type>),
    TypeVariable(TypeVariable),
    TraitAsType(TraitId, Rc<String>, TraitGenerics),
    NamedGeneric(NamedGeneric),
    CheckedCast {
        from: Box<Type>,
        to: Box<Type>,
    },
    Function(Vec<Type>, Box<Type>, Box<Type>, bool),
    Reference(Box<Type>, bool),
    Forall(GenericTypeVars, Box<Type>),
    Constant(SignedField, Kind),
    Quoted(QuotedType),
    InfixExpr(Box<Type>, BinaryTypeOperator, Box<Type>, bool),
    Error,

}

Variants

FieldElement

A primitive Field type

Array(Box<Type>, Box<Type>)

Array(N, E) is an array of N elements of type E. It is expected that N is either a type variable of some kind or a Type::Constant.

Slice(Box<Type>)

Slice(E) is a slice of elements of type E.

Integer(Signedness, IntegerBitSize)

A primitive integer type with the given sign and bit count. E.g. u32 would be Integer(Unsigned, ThirtyTwo)

Bool

The primitive bool type.

String(Box<Type>)

String(N) is an array of characters of length N. It is expected that N is either a type variable of some kind or a Type::Constant.

FmtString(Box<Type>, Box<Type>)

FmtString(N, Vec<E>) is an array of characters of length N that contains a list of fields specified inside the string by the following regular expression r“{([\S]+)}“

Unit

The unit type ().

Tuple(Vec<Type>)

A tuple type with the given list of fields in the order they appear in source code.

DataType(Shared<DataType>, Vec<Type>)

A user-defined struct type. The Shared<StructType> field here refers to the shared definition for each instance of this struct type. The Vec<Type> represents the generic arguments (if any) to this struct type.

Alias(Shared<TypeAlias>, Vec<Type>)

A user-defined alias to another type. Similar to a Struct, this carries a shared reference to the definition of the alias along with any generics that may have been applied to the alias.

TypeVariable(TypeVariable)

TypeVariables are stand-in variables for some type which is not yet known. They are not to be confused with NamedGenerics. While the later mostly works as with normal types (ie. for two NamedGenerics T and U, T != U), TypeVariables will be automatically rebound as necessary to satisfy any calls to unify.

TypeVariables are often created when a generic function is instantiated. This is a process that replaces each NamedGeneric in a generic function with a TypeVariable. Doing this at each call site of a generic function is how they can be called with different argument types each time.

TraitAsType(TraitId, Rc<String>, TraitGenerics)

impl Trait when used in a type position. These are only matched based on the TraitId. The trait name parameter is only used for displaying error messages using the name of the trait.

NamedGeneric(NamedGeneric)

NamedGenerics are the ‘T’ or ‘U’ in a user-defined generic function like fn foo<T, U>(...) {}. Unlike TypeVariables, they cannot be bound over.

CheckedCast

A cast (to, from) that’s checked at monomorphization.

Simplifications on arithmetic generics are only allowed on the LHS.

Fields

from: Box<Type>

to: Box<Type>

Function(Vec<Type>, Box<Type>, Box<Type>, bool)

A functions with arguments, a return type and environment. the environment should be Unit by default, for closures it should contain a Tuple type with the captured variable types.

Reference(Box<Type>, bool)

&T

Forall(GenericTypeVars, Box<Type>)

A type generic over the given type variables. Storing both the TypeVariableId and TypeVariable isn’t necessary but it makes handling them both easier. The TypeVariableId should never be bound over during type checking, but during monomorphization it will be and thus needs the full TypeVariable link.

Constant(SignedField, Kind)

A type-level integer. Included to let

an Array’s size type variable bind to an integer without special checks to bind it to a non-type.

values to be used at the type level

Quoted(QuotedType)

The type of quoted code in macros. This is always a comptime-only type

InfixExpr(Box<Type>, BinaryTypeOperator, Box<Type>, bool)

An infix expression in the form lhs * rhs.

The inversion bool keeps track of whether this expression came from an expression like 4 = a / b which was transformed to a = 4 / b so that if at some point a infix expression b * (4 / b) is created, it could be simplified back to 4.

Error

The result of some type error. Remembering type errors as their own type variant lets us avoid issuing repeat type errors for the same item. For example, a lambda with an invalid type would otherwise issue a new error each time it is called if not for this variant.

Implementations

impl Type

fn to_display_ast(&self) -> UnresolvedType

Convert to AST for display (some details lost)

fn to_type_expression(&self) -> UnresolvedTypeExpression

Convert to AST for display (some details lost)

impl Type

pub fn canonicalize(&self) -> Type

Try to canonicalize the representation of this type. Currently the only type with a canonical representation is Type::Infix where for each consecutive commutative operator we sort the non-constant operands by Type: Ord and place all constant operands at the end, constant folded.

For example:

canonicalize[((1 + N) + M) + 2] = (M + N) + 3

canonicalize[A + 2 * B + 3 - 2] = A + (B * 2) + 3 - 2

pub(crate) fn canonicalize_with_simplifications( &self, run_simplifications: bool, ) -> Type

pub(crate) fn canonicalize_checked(&self) -> Type

Only simplify constants and drop/skip any CheckedCast’s

fn canonicalize_checked_helper(&self) -> Type

Only simplify constants and drop/skip any CheckedCast’s

fn canonicalize_unchecked(&self) -> Type

Run all simplifications and drop/skip any CheckedCast’s

fn canonicalize_helper( &self, found_checked_cast: bool, run_simplifications: bool, ) -> Type

If found_checked_cast, then drop additional CheckedCast’s

If run_simplifications is false, then only:

Attempt to evaluate each sub-expression to a constant

Drop nested CheckedCast’s

Otherwise also attempt try_simplify_partial_constants, sort_commutative, and other simplifications

fn sort_commutative(lhs: &Type, op: BinaryTypeOperator, rhs: &Type) -> Type

fn try_simplify_non_constants_in_lhs( lhs: &Type, op: BinaryTypeOperator, rhs: &Type, ) -> Option<Type>

Try to simplify non-constant expressions in the form (N op1 M) op2 M where the two M terms are expected to cancel out. Precondition: lhs & rhs are in canonical form

Simplifies (N +/- M) -/+ M to N

Simplifies (N * M) ÷ M to N

fn try_simplify_non_constants_in_rhs( lhs: &Type, op: BinaryTypeOperator, rhs: &Type, ) -> Option<Type>

Try to simplify non-constant expressions in the form N op1 (M op1 N) where the two M terms are expected to cancel out. Precondition: lhs & rhs are in canonical form

Unlike try_simplify_non_constants_in_lhs we can’t simplify N / (M * N) Since that should simplify to 1 / M instead of M.

Simplifies N +/- (M -/+ N) to M

Simplifies N * (M ÷ N) to M

fn parse_partial_constant_expr( lhs: &Type, rhs: &Type, ) -> Option<(Box<Type>, BinaryTypeOperator, SignedField, SignedField)>

Given: lhs = N op C1 rhs = C2 Returns: (N, op, C1, C2) if C1 and C2 are constants. Note that the operator here is within the lhs term, the operator separating lhs and rhs is not needed. Precondition: lhs & rhs are in canonical form

fn try_simplify_partial_constants( lhs: &Type, op: BinaryTypeOperator, rhs: &Type, ) -> Option<Type>

Try to simplify partially constant expressions in the form (N op1 C1) op2 C2 where C1 and C2 are constants that can be combined (e.g. N + 5 - 3 = N + 2) Precondition: lhs & rhs are in canonical form

Simplifies (N +/- C1) +/- C2 to N +/- (C1 +/- C2) if C1 and C2 are constants.

Simplifies (N * C1) ÷ C2 to N * (C1 ÷ C2) if C1 and C2 are constants which divide without a remainder.

impl Type

pub fn unify(&self, expected: &Type) -> Result<(), UnificationError>

Try to unify this type with another, setting any type variables found equal to the other type in the process. When comparing types, unification (including try_unify) are almost always preferred over Type::eq as unification will correctly handle generic types.

pub fn try_unify( &self, other: &Type, bindings: &mut TypeBindings, ) -> Result<(), UnificationError>

try_unify is a bit of a misnomer since although errors are not committed, any unified bindings are on success.

fn try_unify_with_flags( &self, other: &Type, flags: UnificationFlags, bindings: &mut TypeBindings, ) -> Result<(), UnificationError>

fn try_unify_to_type_variable( &self, type_variable: &TypeVariable, flags: UnificationFlags, bindings: &mut TypeBindings, bind_variable: impl FnOnce(&mut TypeBindings) -> Result<(), UnificationError>, ) -> Result<(), UnificationError>

Try to unify a type variable to self. This is a helper function factored out from try_unify.

pub(super) fn try_unify_by_isolating_an_unbound_type_variable( &self, other: &Type, bindings: &mut TypeBindings, ) -> Result<(), UnificationError>

Try to unify the following equations:

A + rhs = other -> A = other - rhs

A - rhs = other -> A = other + rhs

lhs + B = other -> B = other - lhs

lhs - B = other -> B = lhs - other

other = A + rhs -> A = other - rhs

other = A - rhs -> A = other + rhs

other = lhs + B -> B = other - lhs

other = lhs - B -> B = lhs - other

fn try_unify_by_isolating_an_unbound_type_variable_in_self( &self, other: &Type, bindings: &mut TypeBindings, ) -> Result<(), UnificationError>

Try to unify the following equations:

A + rhs = other -> A = other - rhs

A - rhs = other -> A = other + rhs

lhs + B = other -> B = other - lhs

lhs - B = other -> B = lhs - other

fn try_unify_by_moving_single_constant_term( &self, other: &Type, flags: UnificationFlags, bindings: &mut TypeBindings, ) -> Result<(), UnificationError>

Try to unify the following equations:

(..a..) + 1 = (..b..) -> (..a..) = (..b..) - 1

(..a..) - 1 = (..b..) -> (..a..) = (..b..) + 1

(..a..) = (..b..) + 1 -> (..b..) = (..a..) - 1

(..a..) = (..b..) - 1 -> (..b..) = (..a..) + 1

fn try_unify_by_moving_single_constant_term_in_self( &self, other: &Type, bindings: &mut TypeBindings, ) -> Result<(), UnificationError>

Try to unify the following equations:

(..a..) + 1 = (..b..) -> (..a..) = (..b..) - 1

(..a..) - 1 = (..b..) -> (..a..) = (..b..) + 1

pub fn unify_with_coercions( &self, expected: &Type, expression: ExprId, location: Location, interner: &mut NodeInterner, errors: &mut Vec<CompilationError>, make_error: impl FnOnce() -> CompilationError, )

Similar to unify but if the check fails this will attempt to coerce the argument to the target type. When this happens, the given expression is wrapped in a new expression to convert its type. E.g. array -> array.as_slice()

Currently the only type coercion in Noir is [T; N] into [T] via .as_slice().

fn try_fn_to_unconstrained_fn_coercion( &self, expected: &Type, ) -> FunctionCoercionResult

fn try_array_to_slice_coercion( &self, target: &Type, expression: ExprId, interner: &mut NodeInterner, ) -> bool

Try to apply the array to slice coercion to this given type pair and expression. If self can be converted to target this way, do so and return true to indicate success.

fn try_string_to_ctstring_coercion( &self, target: &Type, expression: ExprId, interner: &mut NodeInterner, ) -> bool

pub(crate) fn try_reference_coercion(&self, target: &Type) -> bool

Attempt to coerce &mut T to &T, returning true if this is possible.

impl Type

pub(crate) fn program_input_validity(&self) -> Option<InvalidType>

Returns this type, or a nested one, that cannot be used as a parameter to main or a contract function. This is only Some for unsized types like slices or slices that do not make sense as a program input such as named generics or mutable references.

This function should match the same check done in create_value_from_type in acir_gen. If this function does not catch a case where a type should be valid, it will later lead to a panic in that function instead of a user-facing compiler error message.

Returns None if this type and its nested types are all valid program inputs.

pub(crate) fn non_inlined_function_input_validity(&self) -> Option<InvalidType>

Returns this type, or a nested one, if this type can be used as a parameter to an ACIR function that is not main or a contract function. This encapsulates functions for which we may not want to inline during compilation.

The inputs allowed for a function entry point differ from those allowed as input to a program as there are certain types which through compilation we know what their size should be. This includes types such as numeric generics.

pub(crate) fn is_valid_for_unconstrained_boundary(&self) -> bool

Returns true if a value of this type can safely pass between constrained and unconstrained functions (and vice-versa).

impl Type

pub fn default_int_or_field_type() -> Type

pub fn default_int_type() -> Type

pub fn type_variable_with_kind( interner: &NodeInterner, type_var_kind: Kind, ) -> Type

pub fn type_variable(id: TypeVariableId) -> Type

pub fn polymorphic_integer_or_field(interner: &NodeInterner) -> Type

pub fn polymorphic_integer(interner: &NodeInterner) -> Type

pub fn is_bindable(&self) -> bool

A bit of an awkward name for this function - this function returns true for type variables or polymorphic integers which are unbound. NamedGenerics will always be false as although they are bindable, they shouldn’t be bound over until monomorphization.

pub fn is_field(&self) -> bool

pub fn is_bool(&self) -> bool

pub fn is_integer(&self) -> bool

pub fn is_signed(&self) -> bool

pub fn is_unsigned(&self) -> bool

pub fn is_numeric_value(&self) -> bool

While Kind::is_numeric refers to numeric types, this method checks for numeric values

pub fn is_primitive(&self) -> bool

pub fn is_function(&self) -> bool

pub(crate) fn is_mutable_ref(&self) -> bool

pub(crate) fn is_ref(&self) -> bool

pub(crate) fn is_message_compatible(&self, is_monomorphized: bool) -> bool

Returns true if a type is allowed to appear in an assertion message.

This should try filter out types which would cause a panic in abi_gen::abi_type_from_hir_type, but it has to be more permissive, as we don’t have all information yet; some only become apparent after monomorphization.

pub fn generic_count(&self) -> usize

Returns the number of Forall-quantified type variables on this type. Returns 0 if this is not a Type::Forall

pub(crate) fn generalize_from_substitutions( self, type_bindings: TypeBindings, ) -> Type

Takes a monomorphic type and generalizes it over each of the type variables in the given type bindings, ignoring what each type variable is bound to in the TypeBindings and their Kind’s

pub fn as_monotype(&self) -> &Type

Return this type as a monomorphic type - without a Type::Forall if there is one. This is only a shallow check since Noir’s type system prohibits Type::Forall anywhere inside other types.

pub fn unwrap_forall(&self) -> (Cow<'_, GenericTypeVars>, &Type)

Return the generics and type within this Type::Forall. Panics if self is not Type::Forall

pub fn kind(&self) -> Kind

pub fn has_cyclic_alias(&self, aliases: &mut HashSet<TypeAliasId>) -> bool

Determines if a type contains a self referring alias by tracking visited TypeAliasId.

aliases is a mutable set of TypeAliasId to track visited aliases

it returns true if a cyclic alias is detected, false otherwise

fn infix_kind(&self, other: &Self) -> Kind

Unifies self and other kinds or fails with a Kind error

pub fn infix_expr( lhs: Box<Type>, op: BinaryTypeOperator, rhs: Box<Type>, ) -> Type

Creates an InfixExpr.

pub fn inverted_infix_expr( lhs: Box<Type>, op: BinaryTypeOperator, rhs: Box<Type>, ) -> Type

Creates an InfixExpr that results from the compiler trying to unify something like 4 = a * b into a = 4 / b (where 4 / b is the “inverted” expression).

pub fn new_infix_expr( lhs: Box<Type>, op: BinaryTypeOperator, rhs: Box<Type>, inversion: bool, ) -> Type

pub fn field_count(&self, location: &Location) -> u32

Returns the number of field elements required to represent the type once encoded.

pub(crate) fn is_nested_slice(&self) -> bool

Check whether this type is an array or slice, and contains a nested slice in its element type.

pub(crate) fn contains_slice(&self) -> bool

Check whether this type is itself a slice, or a struct/enum/tuple/array which contains a slice.

pub(crate) fn contains_reference(&self) -> bool

pub(crate) fn contains_function(&self) -> bool

pub(crate) fn contains_type_variable(&self) -> bool

fn try_bind_to_polymorphic_int( &self, var: &TypeVariable, bindings: &mut TypeBindings, only_integer: bool, ) -> Result<(), UnificationError>

Try to bind a PolymorphicInt variable to self, succeeding if self is an integer, field, other PolymorphicInt type, or type variable. If successful, the binding is placed in the given TypeBindings map rather than linked immediately.

fn try_bind_to( &self, var: &TypeVariable, bindings: &mut TypeBindings, kind: Kind, ) -> Result<(), UnificationError>

Try to bind the given type variable to self. Although the given type variable is expected to be of Kind::Normal, this binding can still fail if the given type variable occurs within self as that would create a recursive type.

If successful, the binding is placed in the given TypeBindings map rather than linked immediately.

fn get_inner_type_variable(&self) -> Option<(Shared<TypeBinding>, Kind)>

pub fn apply_type_bindings(bindings: TypeBindings)

Apply the given type bindings, making them permanently visible for each clone of each type variable bound.

pub fn evaluate_to_u32(&self, location: Location) -> Result<u32, TypeCheckError>

If this type is a Type::Constant (used in array lengths), or is bound to a Type::Constant, return the constant as a u32.

pub(crate) fn evaluate_to_signed_field( &self, kind: &Kind, location: Location, ) -> Result<SignedField, TypeCheckError>

pub(crate) fn evaluate_to_signed_field_helper( &self, kind: &Kind, location: Location, run_simplifications: bool, ) -> Result<SignedField, TypeCheckError>

evaluate_to_field_element with optional generic arithmetic simplifications

pub fn get_field_type_and_visibility( &self, field_name: &str, ) -> Option<(Type, ItemVisibility)>

Retrieves the Type and ItemVisibility of the given field name:

for structs, it finds a member with a matching name

for tuples, it find the item by index, treating indexes as names “0”, “1”, …

otherwise returns None

pub fn instantiate_with_bindings( &self, bindings: TypeBindings, interner: &NodeInterner, ) -> (Type, TypeBindings)

Instantiate this type with the given type bindings. If any type variables which would be instantiated are contained in the given type bindings instead, the value from the type bindings is used.

pub fn instantiate(&self, interner: &NodeInterner) -> (Type, TypeBindings)

Instantiate this type, replacing any type variables it is quantified over with fresh type variables. If this type is not a Type::Forall, it is unchanged.

pub fn instantiate_with_bindings_and_turbofish( &self, bindings: TypeBindings, turbofish_types: Vec<Type>, interner: &NodeInterner, implicit_generic_count: usize, ) -> (Type, TypeBindings)

Instantiates a type with the given types. This differs from substitute in that only the quantified type variables are matched against the type list and are eligible for substitution - similar to normal instantiation. This function is used when the turbofish operator is used and generic substitutions are provided manually by users.

Expects the given type vector to be the same length as the Forall type variables.

fn type_variable_id(&self) -> Option<TypeVariableId>

pub fn substitute(&self, type_bindings: &TypeBindings) -> Type

Substitute any type variables found within this type with the given bindings if found. If a type variable is not found within the given TypeBindings, it is unchanged.

pub fn force_substitute(&self, type_bindings: &TypeBindings) -> Type

Forcibly substitute any type variables found within this type with the given bindings if found. If a type variable is not found within the given TypeBindings, it is unchanged.

Compared to substitute, this function will also substitute any type variables from type_bindings, even if they are bound in self. Since this can undo previous bindings, this function should be avoided unless necessary. Currently, it is only needed when handling bindings between trait methods and their corresponding impl method during monomorphization.

fn substitute_helper( &self, type_bindings: &TypeBindings, substitute_bound_typevars: bool, ) -> Type

This helper function only differs in the additional parameter which, if set, allows substitutions on already-bound type variables. This should be false for most uses, but is currently needed during monomorphization when instantiating trait functions to shed any previous bindings from recursive parent calls to the same trait.

pub fn occurs(&self, target_id: TypeVariableId) -> bool

True if the given TypeVariableId is free anywhere within self

pub fn follow_bindings(&self) -> Type

Follow any TypeVariable bindings within this type. Doing so ensures that if the bindings are rebound or unbound from under the type then the returned type will not change (because it will no longer contain the links that may be unbound).

Expected to be called on an instantiated type (with no Type::Foralls)

pub fn follow_bindings_shallow(&self) -> Cow<'_, Type>

Follow bindings if this is a type variable or generic to the first non-type-variable type. Unlike follow_bindings, this won’t recursively follow any bindings on any fields or arguments of this type.

pub fn from_generics(generics: &GenericTypeVars) -> Vec<Type>

pub fn replace_named_generics_with_type_variables(&mut self)

Replace any Type::NamedGeneric in this type with a Type::TypeVariable using to the same inner TypeVariable. This is used during monomorphization to bind to named generics since they are unbindable during type checking.

pub fn slice_element_type(&self) -> Option<&Type>

pub(crate) fn integral_maximum_size(&self) -> Option<FieldElement>

pub(crate) fn integral_minimum_size(&self) -> Option<SignedField>

pub(crate) fn substitute_kind_any_with_kind(self, kind: &Kind) -> Type

Substitute any Kind::Any in this type, for types that hold kinds (like Type::Constant) with the given kind.

pub(crate) fn as_integer_type_suffix(&self) -> Option<IntegerTypeSuffix>