noirc_frontend::hir_def::types
Struct DataType

pub struct DataType {
    pub id: TypeId,
    pub name: Ident,
    pub visibility: ItemVisibility,
    body: TypeBody,
    pub generics: ResolvedGenerics,
    pub location: Location,
    pub must_use: MustUse,
}

Represents a struct or enum type in the type system. Each instance of this rust struct will be shared across all Type::DataType variants that represent the same struct or enum type.

Fields
id: TypeId

A unique id representing this type. Used to check if two types are equal.

name: Identvisibility: ItemVisibilitybody: TypeBody

A type’s body is private to force struct fields or enum variants to only be accessed through get_field(), get_fields(), instantiate(), or similar functions since these will handle applying generic arguments to fields as well.

generics: ResolvedGenericslocation: Locationmust_use: MustUse
Implementations

impl DataType

pub fn new( id: TypeId, name: Ident, location: Location, generics: ResolvedGenerics, visibility: ItemVisibility, ) -> DataType

pub fn set_fields(&mut self, fields: Vec<StructField>)

To account for cyclic references between structs, a struct’s fields are resolved strictly after the struct itself is initially created. Therefore, this method is used to set the fields once they become known.

pub(crate) fn init_variants(&mut self)

pub(crate) fn push_variant(&mut self, variant: EnumVariant)

pub fn is_struct(&self) -> bool

pub fn is_enum(&self) -> bool

pub fn fields_raw(&self) -> Option<&[StructField]>

Retrieve the fields of this type with no modifications. Returns None if this is not a struct type.

fn variants_raw(&self) -> Option<&[EnumVariant]>

Retrieve the variants of this type with no modifications. Panics if this is not an enum type.

pub fn generic_types(&self) -> Vec<Type>

Return the generics on this type as a vector of types

pub fn get_field( &self, field_name: &str, generic_args: &[Type], ) -> Option<(Type, ItemVisibility, usize)>

Returns the field matching the given field name, as well as its visibility and field index. Always returns None if this is not a struct type.

pub fn get_fields_with_visibility( &self, generic_args: &[Type], ) -> Option<Vec<(String, ItemVisibility, Type)>>

Returns all the fields of this type, after being applied to the given generic arguments. Returns None if this is not a struct type.

pub fn get_fields( &self, generic_args: &[Type], ) -> Option<Vec<(String, Type, ItemVisibility)>>

Retrieve the fields of this type. Returns None if this is not a field type

pub fn get_variants( &self, generic_args: &[Type], ) -> Option<Vec<(String, Vec<Type>)>>

Retrieve the variants of this type. Returns None if this is not an enum type

pub fn get_variant( &self, variant_index: usize, generic_args: &[Type], ) -> Option<(String, Vec<Type>)>

Retrieve the given variant at the given variant index of this type. Returns None if this is not an enum type or variant_index is out of bounds.

fn get_fields_substitutions( &self, generic_args: &[Type], ) -> HashMap<TypeVariableId, (TypeVariable, Kind, Type)>

pub fn get_fields_as_written(&self) -> Option<Vec<StructField>>

Returns the name and raw types of each field of this type. This will not substitute any generic arguments so a generic field like x in struct Foo<T> { x: T } will return a ("x", T) pair.

This method is almost never what is wanted for type checking or monomorphization, prefer to use get_fields whenever possible.

Returns None if this is not a struct type.

pub fn get_variants_as_written(&self) -> Option<Vec<EnumVariant>>

Returns the name and raw parameters of each variant of this type. This will not substitute any generic arguments so a generic variant like X in enum Foo<T> { X(T) } will return a ("X", Vec<T>) pair.

Returns None if this is not an enum type.

pub fn get_variant_as_written( &self, variant_index: usize, ) -> Option<&EnumVariant>

Returns the name and raw parameters of the variant at the given variant index. This will not substitute any generic arguments so a generic variant like X in enum Foo<T> { X(T) } will return a ("X", Vec<T>) pair.

Returns None if this is not an enum type or the given variant index is out of bounds.

pub fn field_at(&self, index: usize) -> &StructField

Returns the field at the given index. Panics if no field exists at the given index or this is not a struct type.

pub fn variant_at(&self, index: usize) -> &EnumVariant

Returns the enum variant at the given index. Panics if no field exists at the given index or this is not an enum type.

pub fn field_names(&self) -> Option<BTreeSet<Ident>>

Returns each of this type’s field names. Returns None if this is not a struct type.

pub fn instantiate(&self, interner: &mut NodeInterner) -> Vec<Type>

Instantiate this struct type, returning a Vec of the new generic args (in the same order as self.generics)

pub fn variant_function_type( &self, variant_index: usize, this: Shared<DataType>, ) -> Type

Returns the function type of the variant at the given index of this enum. Requires the Shared<DataType> handle of self to create the given function type. Panics if this is not an enum.

The function type uses the variant “as written” ie. no generic substitutions. Although the returned function is technically generic, Type::Function is returned instead of Type::Forall.

pub fn variant_function_type_with_forall( &self, variant_index: usize, this: Shared<DataType>, ) -> Type

Returns the function type of the variant at the given index of this enum. Requires the Shared<DataType> handle of self to create the given function type. Panics if this is not an enum.