noirc_frontend

Module lexer

The lexer is the first pass of the noir compiler. Its goal is to convert a character stream (one input file) into a stream of tokens. See lexer.rs for the lexer itself, and token.rs for more details on tokens.

Re-exports

pub use lexer::Lexer;

pub use lexer::SpannedTokenResult;

Modules

errors

lexer

token