noirc_frontend::lexer
Module lexer

Structs

Lexer

The job of the lexer is to transform an iterator of characters (char_iter) into an iterator of SpannedToken. Each Token corresponds roughly to 1 word or operator. Tokens are tagged with their location in the source file (a Span) for use in error reporting.

Type Aliases

LocatedTokenResult

SpannedTokenResult