noirc_frontend::lexer
Module token

Structs

Attributes

FunctionAttribute

Primary Attributes are those which a function can only have one of. They change the FunctionKind and thus have direct impact on the IR output

KeywordIter

An iterator over the variants of Self

LocatedToken

MetaAttribute

SecondaryAttribute

Secondary attributes are those which a function can have many of. They are not able to change the FunctionKind and thus do not have direct impact on the IR output They are often consumed by libraries or used as notices for the developer

SpannedToken

Tokens

Enums

Attribute

An Attribute can be either a Primary Attribute or a Secondary Attribute A Primary Attribute can alter the function type, thus there can only be one A secondary attribute has no effect and is either consumed by a library or used as a notice for the developer

BorrowedToken

Represents a token in noir’s grammar - a word, number, or symbol that can be used in noir’s syntax. This is the smallest unit of grammar. A parser may (will) decide to parse items differently depending on the Tokens present but will never parse the same ordering of identical tokens differently.

DocStyle

FmtStrFragment

FunctionAttributeKind

Primary Attributes are those which a function can only have one of. They change the FunctionKind and thus have direct impact on the IR output

FuzzingScope

FuzzingScope is used to specify additional annotations for fuzzing harnesses

IntType

IntegerTypeSuffix

Keyword

Note that self is not present - it is a contextual keyword rather than a true one as it is only special within impls. Otherwise self functions as a normal identifier.

MetaAttributeName

SecondaryAttributeKind

TestScope

TestScope is used to specify additional annotations for test functions

Token

TokenKind

The different kinds of tokens that are possible in the target language

Functions

token_to_borrowed_token