noirc_frontend::lexer::token
Enum BorrowedToken

pub enum BorrowedToken<'input> {

    Ident(&'input str),
    Int(FieldElement, Option<IntegerTypeSuffix>),
    Bool(bool),
    Str(&'input str),
    RawStr(&'input str, u8),
    FmtStr(&'input [FmtStrFragment], u32),
    Keyword(Keyword),
    AttributeStart {
        is_inner: bool,
        is_tag: bool,
    },
    LineComment(&'input str, Option<DocStyle>),
    BlockComment(&'input str, Option<DocStyle>),
    Quote(&'input Tokens),
    QuotedType(QuotedTypeId),
    InternedExpression(InternedExpressionKind),
    InternedStatement(InternedStatementKind),
    InternedLValue(InternedExpressionKind),
    InternedUnresolvedTypeData(InternedUnresolvedTypeData),
    InternedPattern(InternedPattern),
    InternedCrate(CrateId),
    Less,
    LessEqual,
    Greater,
    GreaterEqual,
    Equal,
    NotEqual,
    Plus,
    Minus,
    Star,
    Slash,
    Percent,
    Ampersand,
    DeprecatedVectorStart,
    At,
    Caret,
    ShiftLeft,
    ShiftRight,
    Dot,
    DoubleDot,
    DoubleDotEqual,
    LeftParen,
    RightParen,
    LeftBrace,
    RightBrace,
    LeftBracket,
    RightBracket,
    Arrow,
    FatArrow,
    Pipe,
    Pound,
    Comma,
    Colon,
    DoubleColon,
    Semicolon,
    Bang,
    DollarSign,
    Assign,
    LogicalAnd,
    EOF,
    Whitespace(&'input str),
    UnquoteMarker(ExprId),
    Invalid(char),

}

Represents a token in noir’s grammar - a word, number, or symbol that can be used in noir’s syntax. This is the smallest unit of grammar. A parser may (will) decide to parse items differently depending on the Tokens present but will never parse the same ordering of identical tokens differently.

Variants

Ident(&'input str)

Int(FieldElement, Option<IntegerTypeSuffix>)

Bool(bool)

Str(&'input str)

RawStr(&'input str, u8)

the u8 is the number of hashes, i.e. r###..

FmtStr(&'input [FmtStrFragment], u32)

Keyword(Keyword)

AttributeStart

Fields

is_inner: bool

is_tag: bool

LineComment(&'input str, Option<DocStyle>)

BlockComment(&'input str, Option<DocStyle>)

Quote(&'input Tokens)

QuotedType(QuotedTypeId)

InternedExpression(InternedExpressionKind)

InternedStatement(InternedStatementKind)

InternedLValue(InternedExpressionKind)

InternedUnresolvedTypeData(InternedUnresolvedTypeData)

InternedPattern(InternedPattern)

InternedCrate(CrateId)

Less

<

LessEqual

<=

Greater

GreaterEqual

=

Equal

==

NotEqual

!=

Plus

Minus

Star

Slash

/

Percent

%

Ampersand

&

DeprecatedVectorStart

&

At

@

Caret

^

ShiftLeft

<<

ShiftRight

Dot

.

DoubleDot

..

DoubleDotEqual

..=

LeftParen

(

RightParen

)

LeftBrace

{

RightBrace

}

LeftBracket

[

RightBracket

]

Arrow

->

FatArrow

=>

Pipe

|

Pound

Comma

,

Colon

:

DoubleColon

::

Semicolon

;

Bang

!

DollarSign

$

Assign

=

LogicalAnd

&&

EOF

Whitespace(&'input str)

UnquoteMarker(ExprId)

This is an implementation detail on how macros are implemented by quoting token streams. This token marks where an unquote operation is performed. The ExprId argument is the resolved variable which is being unquoted at this position in the token stream.

Invalid(char)

An invalid character is one that is not in noir’s language or grammar.

We don’t report invalid tokens in the source as errors until parsing to avoid reporting the error twice (once while lexing, again when it is encountered during parsing). Reporting during lexing then removing these from the token stream would not be equivalent as it would change the resulting parse.