noirc_frontend

Module modules

Functions

get_ancestor_module_reexport

Finds a visible reexport for any ancestor module of the given ModuleDefId,

get_parent_module

Returns the ModuleId a ModuleDefId is in.

module_def_id_is_visible

Returns true if the given ModuleDefId is visible from the current module, given its visibility.

module_def_id_relative_path

Returns the relative path to reach module_def_id named name starting from current_module_id.

module_def_id_to_reference_id

module_full_path

relative_module_full_path

Returns the fully-qualified path of the given ModuleDefId relative to current_module_id:

relative_module_id_path

Returns the path to reach an item inside target_module_id from inside current_module_id. Returns a relative path if possible.