noirc_frontend
Module monomorphization

Coming after elaboration, monomorphization is the last pass in Noir’s frontend. It accepts the type checked HIR as input and produces a monomorphized AST as output. This file implements the pass itself, while the AST is defined in the ast module.

Unlike the HIR, which is stored within the NodeInterner, the monomorphized AST is self-contained and does not need an external context struct. As a result, the NodeInterner can be safely discarded after monomorphization.

The entry point to this pass is the monomorphize function which, starting from a given function, will monomorphize the entire reachable program.

The monomorphized AST (mAST) has a few notable differences from the HIR:

It is self-contained without the need for an external context like the NodeInterner.

All generics are gone, they are specialized away by creating a new copy of each function for each combination of generic arguments it is used with.

All local lambdas are gone and closure environments are explicit. Closures are converted into normal tuples of (function, environment) and when the function is called, will forward the environment argument to the function as well. All functions are global and local lambdas no longer exist.

Most patterns are simplified away. For example, a tuple pattern in a function parameter: fn foo((a, b): (i32, i32)) will be translated into separate parameters: fn foo(a: i32, b: i32). Note that this transformation relies on our SSA construction pass to perform a similar translation for the callsites of functions. The monomorphized output alone does not change callsites, so it will produce mAST like the following:

fn main$f0() -> () {
    let tuple$l0 = (0, 1);
    foo$f1(tuple$l0);  // single tuple argument
}

fn foo$f1(a$l1: i32, b$l2: i32) -> () {}  // two separate parameters

Functions are represented as a pair of (constrained, unconstrained) versions of the same function. The variant to use is selected by Monomorphizer::extract_function to match the runtime when the function is later called.

At the end of monomorphization, a couple sub-passes are performed:

ownership: infers when values should be cloned or moved for unconstrained code. This is only relevant for arrays in unconstrained code which are implemented with copy on write semantics. An ast::Expression::Clone corresponds to an increment of the reference-count on a particular array rather than a deep clone. The deep clone itself will be performed by the Brillig runtime when mutating an array with a reference count greater than one.

proxies: wraps oracle functions in unconstrained function wrappers automatically. This is required in some corner cases when oracles are used as first-class functions.

Compared to monomorphization passes in other compilers, Noir’s is a bit odd in that it may still fail with an error message. An example of an error caught at this step would be an unconstrained lambda being passed into and called in constrained code. This is possible when a normal lambda is compiled in an unconstrained context and uses types, such as references, which shouldn’t leave the current context.

Modules

ast

debug 🔒

debug_types

errors

printer

This module implements printing of the monomorphized AST, for debugging purposes.

proxies

Implement a post-monomorphization pass where builtin/intrinsic/oracle functions used as values are wrapped in a proxy function which replaces them as a value and forwards calls to them.

visitor

Structs

LambdaContext 🔒

Monomorphizer

The context struct for the monomorphization pass.

Enums

TraitItem 🔒

Functions

append_printable_type_info_for_type

Extend the arguments to print (which is a bool to show if newline is needed and value to be printed itself) with a JSON serialized PrintableType to describe the value, and another bool to show if the print is using a format string, or a raw value.

append_printable_type_info_inner 🔒

bind_trait_impl_func_generics_to_trait_func_generics 🔒

Binds direct generics on a trait impl function to those on a corresponding trait function.

monomorphize

Starting from the given main function, monomorphize the entire program, replacing all references to type variables and NamedGenerics with concrete types, duplicating definitions as necessary to do so.

monomorphize_debug

A more general entry-point for the monomorphization pass containing an optional DebugInstrumenter which can be set to DebugInstrumenter::default in case it is not desired. If debugging is desired, additional function calls will be inserted to inspect values via debug functions.

perform_impl_bindings

Call sites are instantiated against the trait method, but when an impl is later selected, the corresponding method in the impl will have a different set of generics. perform_impl_bindings is needed to apply the generics from the trait method to the impl method. Without this, static method references to generic impls (e.g. Eq::eq for [T; N]) will fail to re-apply the correct type bindings during monomorphization.

perform_instantiation_bindings

resolve_trait_item 🔒

resolve_trait_item_impl 🔒

Resolve a trait item to a particular impl, returning the id of that impl or an error on failure.

undo_instantiation_bindings

unwrap_enum_type 🔒

Return this enum type’s variants or panic

unwrap_struct_type 🔒

Return this struct type’s fields or panic

unwrap_tuple_type 🔒

Return this tuple type’s fields or panic

Type Aliases

Functions 🔒

Using nested HashMaps here lets us avoid cloning HirTypes when calling .get()

HirType 🔒