noirc_frontend::monomorphization::debug_types
Struct DebugTypeTracker

pub struct DebugTypeTracker {
    source_variables: HashMap<SourceVarId, String>,
    source_field_names: HashMap<SourceFieldId, String>,
    source_to_debug_vars: HashMap<SourceVarId, DebugVarId>,
    variables: HashMap<DebugVarId, (SourceVarId, DebugTypeId)>,
    functions: HashMap<DebugFnId, DebugFunction>,
    types: HashMap<DebugTypeId, PrintableType>,
    types_reverse: HashMap<PrintableType, DebugTypeId>,
    next_var_id: u32,
    next_type_id: u32,
}

We keep a collection of the debug variables and their types in this structure. The source_var_id refers to the ID given by the debug instrumenter. This variable does not have a type yet and hence it can be instantiated for multiple types if itโs in the context of a generic variable. The var_id refers to the ID of the instantiated variable which will have a valid type.

Fields
source_variables: HashMap<SourceVarId, String>source_field_names: HashMap<SourceFieldId, String>source_to_debug_vars: HashMap<SourceVarId, DebugVarId>variables: HashMap<DebugVarId, (SourceVarId, DebugTypeId)>functions: HashMap<DebugFnId, DebugFunction>types: HashMap<DebugTypeId, PrintableType>types_reverse: HashMap<PrintableType, DebugTypeId>next_var_id: u32next_type_id: u32
Implementations

impl DebugTypeTracker

pub fn build_from_debug_instrumenter(instrumenter: &DebugInstrumenter) -> Self

pub fn extract_vars_and_types( &self, ) -> (DebugVariables, DebugFunctions, DebugTypes)

pub fn resolve_field_index( &self, field_id: SourceFieldId, cursor_type: &PrintableType, ) -> Option<usize>

fn insert_type(&mut self, the_type: &Type) -> DebugTypeId

pub fn insert_var( &mut self, source_var_id: SourceVarId, var_type: &Type, ) -> DebugVarId

pub fn get_var_id(&self, source_var_id: SourceVarId) -> Option<DebugVarId>

pub fn get_type(&self, source_var_id: SourceVarId) -> Option<&PrintableType>