noirc_frontend::monomorphization
Function bind_trait_impl_func_generics_to_trait_func_generics

fn bind_trait_impl_func_generics_to_trait_func_generics(
    interner: &mut NodeInterner,
    method_id: TraitItemId,
    impl_id: TraitImplId,
    bindings: &mut TypeBindings,
)

Binds direct generics on a trait impl function to those on a corresponding trait function.

Consider this code:

trait Bar {
    fn baz<let N: u32>();
}

impl Bar for Field {
    fn baz<let M: u32>() {
        let _ = M;
    }
}

fn foo<K: Bar>() {
    K::baz::<2>();
}

fn main() {
    foo::<Field>();
}

When type-checking K::baz::<2> the compiler will know that it corresponds to the method baz defined on the Bar trait, but won’t yet know which trait implementation it refers to. The instantiation bindings for this will have the type variable N bound to 2.

During monomorphization, on the specific call to foo::<Field>(), the compiler will know that K refers to Field, and that K::baz::<2> refers to the baz method on the Bar impl for Field. In that method the type variable M is unbound. Looking up its value will fail. However, we can bind M to N, so it eventually resolves to 2, which is what this method does.