noirc_frontend::monomorphization

Module proxies

Implement a post-monomorphization pass where builtin/intrinsic/oracle functions used as values are wrapped in a proxy function which replaces them as a value and forwards calls to them.

The reason this exists is because by the time we get to the defunctionalize SSA pass it would not be possible to deal with function pointers to foreign functions, because unlike for normal functions, we don’t have type information for them in the SSA. That is, we can’t tell what their parameter and return types are when they are just passed around as a value.

The type information is still available in the monomorphized AST though, so we can find all instances where a foreign function is referenced by an Ident without actually being the target of a Call, and replace them with a normal function, which will preserve the information we need create dispatch functions for them in the defunctionalize pass.

The pass also automatically wraps direct calls to oracle functions from constrained functions, which, after creating wrappers for function values, would only present an inconvenience for users if they have to keep creating wrappers themselves.

Structs

ForeignFunctionValue 🔒

When function values are passed around in the monomorphized AST, they appear as a pair (tuple) of a constrained and unconstrained function.

ProxyContext 🔒

Functions

is_foreign_func 🔒

Check if the definition is that of a function defined by a “name” rather than an ID.

is_func_pair 🔒

Check that the identifier is of a pair of constrained and unconstrained function types.

make_proxy 🔒

Create a proxy function definition for a foreign function based on an identifier that got replaced.