noirc_frontend::monomorphization
Struct Monomorphizer

pub struct Monomorphizer<'interner> {

    functions: FxHashMap<(FuncId, bool), FxHashMap<Type, FxHashMap<Vec<Type>, FuncId>>>,
    locals: FxHashMap<DefinitionId, LocalId>,
    globals: FxHashMap<GlobalId, GlobalId>,
    finished_globals: FxHashMap<GlobalId, (String, Type, Expression)>,
    queue: VecDeque<(FuncId, FuncId, TypeBindings, Option<TraitItemId>, bool, Location)>,
    finished_functions: BTreeMap<FuncId, Function>,
    interner: &'interner mut NodeInterner,
    lambda_envs_stack: Vec<LambdaContext>,
    next_local_id: u32,
    next_global_id: u32,
    next_function_id: u32,
    next_ident_id: u32,
    is_range_loop: bool,
    return_location: Option<Location>,
    debug_type_tracker: DebugTypeTracker,
    in_unconstrained_function: bool,
    force_unconstrained: bool,

}

The context struct for the monomorphization pass.

This struct holds the FIFO queue of functions to monomorphize, which is added to whenever a new (function, type) combination is encountered.

Fields
functions: FxHashMap<(FuncId, bool), FxHashMap<Type, FxHashMap<Vec<Type>, FuncId>>>

Functions are keyed by their unique ID, whether they’re unconstrained, their expected type, and any generics they have so that we can monomorphize a new version of the function for each type.

Keying by any turbofish generics that are specified is necessary for a case where we may have a trait generic that can be instantiated outside of a function parameter or return value.

locals: FxHashMap<DefinitionId, LocalId>

Unlike functions, locals are only keyed by their unique ID because they are never duplicated during monomorphization. Doing so would allow them to be used polymorphically but would also cause them to be re-evaluated which is a performance trap that would confuse users.

globals: FxHashMap<GlobalId, GlobalId>

Globals are keyed by their unique ID because they are never duplicated during monomorphization.

finished_globals: FxHashMap<GlobalId, (String, Type, Expression)>queue: VecDeque<(FuncId, FuncId, TypeBindings, Option<TraitItemId>, bool, Location)>

Queue of functions to monomorphize next each item in the queue is a tuple of: (old_id, new_monomorphized_id, any type bindings to apply, the trait method if old_id is from a trait impl, is_unconstrained, location)

finished_functions: BTreeMap<FuncId, Function>

When a function finishes being monomorphized, the monomorphized ast::Function is stored here along with its FuncId.

interner: &'interner mut NodeInterner

Used to reference existing definitions in the HIR

lambda_envs_stack: Vec<LambdaContext>next_local_id: u32next_global_id: u32next_function_id: u32next_ident_id: u32is_range_loop: boolreturn_location: Option<Location>debug_type_tracker: DebugTypeTrackerin_unconstrained_function: boolforce_unconstrained: bool

Set to true to force every function to be unconstrained. Note that this also changes the first-class function representation from a pair of (constrained, unconstrained) to (unconstrained, unconstrained)

Implementations

impl Monomorphizer<'_>

pub(super) fn patch_debug_instrumentation_call( &mut self, call: &HirCallExpression, function: &Expression, arguments: &mut [Expression], ) -> Result<(), MonomorphizationError>

Patch instrumentation calls inserted for debugging. This will record tracked variables and their types, and assign them an ID to use at runtime. This ID is different from the source ID assigned at instrumentation time because at that point we haven’t fully resolved the types for generic functions. So a single generic function may be instantiated multiple times with its tracked variables being of different types for each instance at runtime.

fn patch_debug_var_assign( &mut self, call: &HirCallExpression, arguments: &mut [Expression], ) -> Result<(), MonomorphizationError>

Update instrumentation code inserted on variable assignment. We need to register the variable instance, its type and replace the source_var_id with the ID of the registration. Multiple registrations of the same variable are possible if using generic functions, hence the temporary ID created when injecting the instrumentation code can map to multiple IDs at runtime.

fn patch_debug_var_drop( &mut self, call: &HirCallExpression, arguments: &mut [Expression], ) -> Result<(), MonomorphizationError>

Update instrumentation code for a variable being dropped out of scope. Given the source_var_id we search for the last assigned debug var_id and replace it instead.

fn patch_debug_member_assign( &mut self, call: &HirCallExpression, arguments: &mut [Expression], arity: usize, ) -> Result<(), MonomorphizationError>

Update instrumentation code inserted when assigning to a member of an existing variable. Same as above for replacing the source_var_id, but also we need to resolve the path and the type of the member being assigned. For this last part, we need to resolve the mapping from field names in structs to positions in the runtime tuple, since all structs are replaced by tuples during compilation.

fn intern_var_id(&mut self, var_id: DebugVarId, location: &Location) -> ExprId

impl<'interner> Monomorphizer<'interner>

pub fn new( interner: &'interner mut NodeInterner, debug_type_tracker: DebugTypeTracker, force_unconstrained: bool, ) -> Self

pub fn has_pending_jobs(&self) -> bool

pub fn process_next_job(&mut self) -> Result<bool, MonomorphizationError>

pub fn process_queue(&mut self) -> Result<(), MonomorphizationError>

pub fn peek_queue( &self, ) -> Option<&(FuncId, FuncId, TypeBindings, Option<TraitItemId>, bool, Location)>

pub fn finished_globals(&self) -> &HashMap<GlobalId, (String, Type, Expression)>

pub fn finished_functions(&self) -> &BTreeMap<FuncId, Function>

pub fn locals(&self) -> &HashMap<DefinitionId, LocalId>

pub fn return_location(&self) -> Option<Location>

pub fn interner(&self) -> &NodeInterner

pub fn into_program(self, function_sig: FunctionSignature) -> Program

pub(crate) fn next_local_id(&mut self) -> LocalId

fn next_function_id(&mut self) -> FuncId

fn next_global_id(&mut self) -> GlobalId

pub(crate) fn next_ident_id(&mut self) -> IdentId

fn lookup_local(&mut self, id: DefinitionId) -> Option<Definition>

fn lookup_function( &mut self, id: FuncId, expr_id: ExprId, typ: &Type, turbofish_generics: &[Type], trait_method: Option<TraitItemId>, ) -> Definition

fn define_local(&mut self, id: DefinitionId, new_id: LocalId)

fn define_function( &mut self, id: FuncId, typ: Type, turbofish_generics: Vec<Type>, is_unconstrained: bool, new_id: FuncId, )

Prerequisite: typ = typ.follow_bindings() and: turbofish_generics = vecmap(turbofish_generics, Type::follow_bindings)

pub fn compile_main( &mut self, main_id: FuncId, ) -> Result<FunctionSignature, MonomorphizationError>

pub fn function( &mut self, f: FuncId, id: FuncId, location: Location, ) -> Result<(), MonomorphizationError>

fn push_function(&mut self, id: FuncId, function: Function)

fn parameters( &mut self, params: &Parameters, ) -> Result<Vec<(LocalId, bool, String, Type, Visibility)>, MonomorphizationError>

Monomorphize each parameter, expanding tuple/struct patterns into multiple parameters and binding any generic types found.

fn parameter( &mut self, param: &HirPattern, typ: &Type, visibility: &Visibility, new_params: &mut Vec<(LocalId, bool, String, Type, Visibility)>, ) -> Result<(), MonomorphizationError>

pub(crate) fn expr( &mut self, expr: ExprId, ) -> Result<Expression, MonomorphizationError>

fn contains_reference(typ: &Type) -> bool

fn standard_array( &mut self, array: ExprId, array_elements: Vec<ExprId>, is_slice: bool, ) -> Result<Expression, MonomorphizationError>

fn repeated_array( &mut self, array: ExprId, repeated_element: ExprId, length: Type, is_slice: bool, ) -> Result<Expression, MonomorphizationError>

fn index( &mut self, id: ExprId, index: HirIndexExpression, ) -> Result<Expression, MonomorphizationError>

fn statement(&mut self, id: StmtId) -> Result<Expression, MonomorphizationError>

fn let_statement( &mut self, let_statement: HirLetStatement, ) -> Result<Expression, MonomorphizationError>

fn constructor( &mut self, constructor: HirConstructorExpression, id: ExprId, ) -> Result<Expression, MonomorphizationError>

fn enum_constructor( &mut self, constructor: HirEnumConstructorExpression, id: ExprId, ) -> Result<Expression, MonomorphizationError>

For an enum like: enum Foo { A(i32, u32), B(Field), C }

this will translate the call Foo::A(1, 2) into (0, (1, 2), (0,), ()) where the first field 0 is the tag value, the second is A, third is B, and fourth is C. Each variant that isn’t the desired variant has zeroed values filled in for its data.

fn block( &mut self, statement_ids: Vec<StmtId>, ) -> Result<Expression, MonomorphizationError>

fn unpack_pattern( &mut self, pattern: HirPattern, value: Expression, typ: &Type, ) -> Result<Expression, MonomorphizationError>

fn unpack_tuple_pattern( &mut self, value: Expression, fields: impl Iterator<Item = (HirPattern, Type)>, tuple_type: &Type, ) -> Result<Expression, MonomorphizationError>

fn lookup_captured_expr(&mut self, id: DefinitionId) -> Option<Expression>

Find a captured variable in the innermost closure, and construct an expression

fn lookup_captured_lvalue(&mut self, id: DefinitionId) -> Option<LValue>

Find a captured variable in the innermost closure construct a LValue

fn local_ident( &mut self, ident: &HirIdent, typ: &Type, ) -> Result<Option<Ident>, MonomorphizationError>

A local (ie non-function) ident only

fn ident( &mut self, ident: HirIdent, expr_id: ExprId, generics: Option<Vec<Type>>, use_current_runtime: bool, ) -> Result<Expression, MonomorphizationError>

fn function_reference( &mut self, mutable: bool, name: String, location: Location, func_id: FuncId, expr_id: ExprId, typ: &Type, generics: Option<Vec<Type>>, ) -> Result<Expression, MonomorphizationError>

Create a function value for the given function (not an actual &mut reference to a function). Note that this will refer to a single constrained or unconstrained version of the given function, it will never return a tuple of (constrained, unconstrained), but may return a tuple if the function is a closure.

fn numeric_generic( &self, value: Type, expected_type: Type, expr_type: Type, location: Location, ) -> Result<Expression, MonomorphizationError>

Monomorphize a numeric generic as a numeric constant. Expects arguments to correspond to let N: $expected_type = $value;

fn global_ident( &mut self, global_id: GlobalId, name: String, typ: &Type, location: Location, ) -> Result<Expression, MonomorphizationError>

pub fn convert_type( typ: &Type, location: Location, ) -> Result<Type, MonomorphizationError>

Convert a non-tuple/struct type to a monomorphized type

fn convert_type_helper( typ: &Type, location: Location, seen_types: &mut HashSet<Type>, ) -> Result<Type, MonomorphizationError>

fn check_type( typ: &Type, location: Location, ) -> Result<(), MonomorphizationError>

fn check_checked_cast( from: &Type, to: &Type, location: Location, ) -> Result<(), MonomorphizationError>

Check that the ‘from’ and to’ sides of a CheckedCast unify and that if the ‘to’ side evaluates to a field element, that the ‘from’ side evaluates to the same field element

fn is_closure_type(&self, t: &Type) -> bool

Returns true if a given function type is a closure. Note that this expects the argument to already be a function so that it can distinguish function values (pairs of (constrained, unconstrained)) from closures (function values where each element is itself a tuple), from normal tuples containing functions.

fn resolve_trait_item_expr( &mut self, expr_id: ExprId, function_type: Type, trait_item_id: TraitItemId, use_current_runtime: bool, ) -> Result<Expression, MonomorphizationError>

fn resolve_trait_method_expr( &mut self, func_id: FuncId, expr_id: ExprId, function_type: Type, trait_item_id: TraitItemId, ) -> Result<Expression, MonomorphizationError>

fn monomorphize_constrained_and_unconstrained<F>( &mut self, use_current_runtime: bool, f: F, ) -> Result<Expression, MonomorphizationError>
where F: Clone + FnOnce(&mut Self) -> Result<Expression, MonomorphizationError>,

Call the same function with self.in_unconstrained_fn = true and self.in_unconstrained_fn = false and return a pair of both results. This is most commonly used for functions which are represented as a pair of their constrained and unconstrained versions, with the appropriate version being selected later only once the function is called.

If use_current_runtime is true we only run the function once with the current runtime.

fn function_call( &mut self, call: HirCallExpression, id: ExprId, ) -> Result<Expression, MonomorphizationError>

fn append_printable_type_info( &mut self, hir_argument: &HirExpression, arguments: &mut Vec<Expression>, )

Adds a function argument that contains type metadata that is required to tell println how to convert values passed to an foreign call back to a human-readable string. The values passed to an foreign call will be a simple list of field elements, thus requiring extra metadata to correctly decode this list of elements.

The Noir compiler has a PrintableType that handles encoding/decoding a list of field elements to/from JSON. The type metadata attached in this method is the serialized PrintableType for the argument passed to the function. The caller that is running a Noir program should then deserialize the PrintableType, and accurately decode the list of field elements passed to the foreign call.

fn try_evaluate_call( &mut self, func: &Expression, expr_id: &ExprId, arguments: &[ExprId], argument_values: &[Expression], result_type: &Type, ) -> Result<Option<Expression>, MonomorphizationError>

Try to evaluate certain builtin functions (currently only ‘array_len’ and field modulus methods) at their call site. NOTE: Evaluating at the call site means we cannot track aliased functions. E.g. let f = std::array::len; f(arr) will fail to evaluate. To fix this we need to evaluate on the identifier instead, which requires us to evaluate to a Lambda value which isn’t in noir yet.

fn checked_transmute( &mut self, expr_id: ExprId, arguments: &[ExprId], argument_values: &[Expression], ) -> Result<Expression, MonomorphizationError>

fn modulus_slice_literal( &self, bytes: Vec<u8>, arr_elem_bits: IntegerBitSize, location: Location, ) -> Expression

fn queue_function( &mut self, id: FuncId, expr_id: ExprId, function_type: Type, turbofish_generics: Vec<Type>, trait_method: Option<TraitItemId>, ) -> FuncId

pub fn queue_function_with_bindings( &mut self, id: FuncId, expr_location: Location, bindings: HashMap<TypeVariableId, (TypeVariable, Kind, Type)>, function_type: Type, turbofish_generics: Vec<Type>, trait_method: Option<TraitItemId>, ) -> FuncId

pub fn follow_bindings(&self, bindings: &TypeBindings) -> TypeBindings

Follow any type variable links within the given TypeBindings to produce a new TypeBindings that won’t be changed when bindings are pushed or popped during {perform,undo}_monomorphization_bindings.

Without this, a monomorphized type may fail to propagate passed more than 2 function calls deep since it is possible for a previous link in the chain to unbind a type variable that was previously bound.

fn assign( &mut self, assign: HirAssignStatement, ) -> Result<Expression, MonomorphizationError>

fn lvalue(&mut self, lvalue: HirLValue) -> Result<LValue, MonomorphizationError>

fn lambda( &mut self, lambda: HirLambda, expr: ExprId, ) -> Result<Expression, MonomorphizationError>

fn lambda_no_capture( &mut self, lambda: HirLambda, expr: ExprId, ) -> Result<Expression, MonomorphizationError>

fn lambda_with_setup( &mut self, lambda: HirLambda, expr: ExprId, ) -> Result<(Expression, Expression), MonomorphizationError>

fn match_expr( &mut self, match_expr: HirMatch, expr_id: ExprId, ) -> Result<Expression, MonomorphizationError>

fn zeroed_value_of_type(&mut self, typ: &Type, location: Location) -> Expression

Implements std::unsafe_func::zeroed by returning an appropriate zeroed ast literal or collection node for the given type. Note that for functions there is no obvious zeroed value so this should be considered unsafe to use.

fn create_zeroed_function( &mut self, parameter_types: &[Type], ret_type: &Type, env_type: &Type, unconstrained: bool, location: Location, ) -> Expression

fn create_infix_operator_impl_call( &self, func: Expression, lhs: Expression, operator: HirBinaryOp, rhs: Expression, ret: Type, location: Location, ) -> Result<Expression, MonomorphizationError>

Call an infix operator overloading method for the given operator. This function handles the special cases some operators have which don’t map 1 to 1 onto their operator function. For example: != requires a negation on the result of its eq method, and the comparison operators each require a conversion from the Ordering result to a boolean.

fn create_prefix_operator_impl_call( &self, func: Expression, rhs: Expression, ret: Type, location: Location, ) -> Result<Expression, MonomorphizationError>

Call an operator overloading method for the given prefix operator.

fn is_unconstrained(&self, func_id: FuncId) -> bool

fn extract_function( &mut self, function: ExprId, ) -> Result<Expression, MonomorphizationError>