noirc_frontend::monomorphization
Type Alias HirType

type HirType = Type;

Aliased Type

enum HirType {

    FieldElement,
    Array(Box<Type>, Box<Type>),
    Vector(Box<Type>),
    Integer(Signedness, IntegerBitSize),
    Bool,
    String(Box<Type>),
    FmtString(Box<Type>, Box<Type>),
    Unit,
    Tuple(Vec<Type>),
    DataType(Shared<DataType>, Vec<Type>),
    Alias(Shared<TypeAlias>, Vec<Type>),
    TypeVariable(TypeVariable),
    TraitAsType(TraitId, Rc<String>, TraitGenerics),
    NamedGeneric(NamedGeneric),
    CheckedCast {
        from: Box<Type>,
        to: Box<Type>,
    },
    Function(Vec<Type>, Box<Type>, Box<Type>, bool),
    Reference(Box<Type>, bool),
    Forall(Vec<TypeVariable>, Box<Type>),
    Constant(SignedField, Kind),
    Quoted(QuotedType),
    InfixExpr(Box<Type>, BinaryTypeOperator, Box<Type>, bool),
    Error,

}

Variants

FieldElement

A primitive Field type

Array(Box<Type>, Box<Type>)

Array(N, E) is an array of N elements of type E. It is expected that N is either a type variable of some kind or a Type::Constant.

Vector(Box<Type>)

Vector(E) is a vector of elements of type E.

Integer(Signedness, IntegerBitSize)

A primitive integer type with the given sign and bit count. E.g. u32 would be Integer(Unsigned, ThirtyTwo)

Bool

The primitive bool type.

String(Box<Type>)

String(N) is an array of characters of length N. It is expected that N is either a type variable of some kind or a Type::Constant.

FmtString(Box<Type>, Box<Type>)

FmtString(N, Vec<E>) is an array of characters of length N that contains a list of fields specified inside the string by the following regular expression r“{([\S]+)}“ and where N is either a type variable of some kind or a Type::Constant

Unit

The unit type ().

Tuple(Vec<Type>)

A tuple type with the given list of fields in the order they appear in source code.

DataType(Shared<DataType>, Vec<Type>)

A user-defined struct or enum type. The Shared<DataType> field here refers to the shared definition for each instance of this struct or enum type. The Vec<Type> represents the generic arguments (if any) to this struct or enum type.

Alias(Shared<TypeAlias>, Vec<Type>)

A user-defined alias to another type. Similar to a struct, this carries a shared reference to the definition of the alias along with any generics that may have been applied to the alias.

TypeVariable(TypeVariable)

TypeVariables are stand-in variables for some type which is not yet known. They are not to be confused with NamedGenerics. While the latter mostly works as with normal types (ie. for two NamedGenerics T and U, T != U), TypeVariables will be automatically rebound as necessary to satisfy any calls to unify.

TypeVariables are often created when a generic function is instantiated. This is a process that replaces each NamedGeneric in a generic function with a TypeVariable. Doing this at each call site of a generic function is how they can be called with different argument types each time.

TraitAsType(TraitId, Rc<String>, TraitGenerics)

impl Trait when used in a type position. These are only matched based on the TraitId. The trait name parameter is only used for displaying error messages using the name of the trait.

NamedGeneric(NamedGeneric)

NamedGenerics are the ‘T’ or ‘U’ in a user-defined generic function like fn foo<T, U>(...) {}. Unlike TypeVariables, they cannot be bound over.

CheckedCast

A cast (to, from) that’s checked at monomorphization.

Simplifications on arithmetic generics are only allowed on the LHS.

Fields

from: Box<Type>

to: Box<Type>

Function(Vec<Type>, Box<Type>, Box<Type>, bool)

A functions with arguments, a return type and environment. the environment should be Unit by default, for closures it should contain a Tuple type with the captured variable types.

Reference(Box<Type>, bool)

&T

Forall(Vec<TypeVariable>, Box<Type>)

A type that’s generic over the given type variables. Storing both the TypeVariableId and TypeVariable isn’t necessary but it makes handling them both easier. The TypeVariableId should never be bound over during type checking, but during monomorphization it will be and thus needs the full TypeVariable link.

Constant(SignedField, Kind)

A type-level integer. Included to let

an Array’s size type variable bind to an integer without special checks to bind it to a non-type.

values to be used at the type level

Quoted(QuotedType)

The type of quoted code in macros. This is always a comptime-only type

InfixExpr(Box<Type>, BinaryTypeOperator, Box<Type>, bool)

An infix expression in the form lhs * rhs.

The inversion bool keeps track of whether this expression came from an expression like 4 = a / b which was transformed to a = 4 / b so that if at some point a infix expression b * (4 / b) is created, it could be simplified back to 4.

Error

The result of some type error. Remembering type errors as their own type variant lets us avoid issuing repeat type errors for the same item. For example, a lambda with an invalid type would otherwise issue a new error each time it is called if not for this variant.