noirc_frontend::node_interner

Module ids

Macros

into_index 🔒

Structs

DefinitionId

ExprId

FuncId

InternedExpressionKind

The ID of an crate::ast::ExpressionKind that’s been pushed into the NodeInterner.

InternedPattern

InternedStatementKind

The ID of a crate::ast::StatementKind that’s been pushed into the NodeInterner.

InternedUnresolvedTypeData

The ID of a crate::ast::UnresolvedTypeData that’s been pushed into the NodeInterner.

QuotedTypeId

StmtId

TraitAssociatedTypeId

TraitId

TraitImplId

TraitItemId

TypeAliasId

TypeId

A TypeId wraps a ModuleId, because types are represented by an anonymous module.

Enums

ReferenceId

A reference to a module, struct, trait, etc., mainly used by the LSP code to keep track of how symbols reference each other.