noirc_frontend::node_interner
Struct NodeInterner

pub struct NodeInterner {

    pub(crate) nodes: Arena<Node>,
    pub(crate) func_meta: FxHashMap<FuncId, FuncMeta>,
    function_definition_ids: FxHashMap<FuncId, DefinitionId>,
    function_modifiers: FxHashMap<FuncId, FunctionModifiers>,
    function_modules: FxHashMap<FuncId, ModuleId>,
    module_attributes: FxHashMap<ModuleId, ModuleAttributes>,
    dependency_graph: DiGraph<DependencyId, ()>,
    dependency_graph_indices: FxHashMap<DependencyId, NodeIndex>,
    pub(crate) id_to_location: FxHashMap<Index, Location>,
    definitions: Vec<DefinitionInfo>,
    id_to_type: FxHashMap<Index, Type>,
    definition_to_type: FxHashMap<DefinitionId, Type>,
    data_types: FxHashMap<TypeId, Shared<DataType>>,
    type_attributes: FxHashMap<TypeId, Vec<SecondaryAttribute>>,
    pub(crate) type_aliases: Vec<Shared<TypeAlias>>,
    pub(crate) trait_associated_types: Vec<TraitAssociatedType>,
    pub(crate) traits: FxHashMap<TraitId, Trait>,
    pub(crate) trait_implementations: FxHashMap<TraitImplId, Shared<TraitImpl>>,
    next_trait_implementation_id: usize,
    trait_impl_associated_types: FxHashMap<TraitImplId, Vec<NamedType>>,
    trait_impl_associated_constants: FxHashMap<TraitImplId, FxHashMap<String, (DefinitionId, Type)>>,
    trait_implementation_map: FxHashMap<TraitId, Vec<(Type, TraitImplKind)>>,
    selected_trait_implementations: FxHashMap<ExprId, TraitImplKind>,
    infix_operator_traits: FxHashMap<BinaryOpKind, TraitId>,
    prefix_operator_traits: FxHashMap<UnaryOp, TraitId>,
    ordering_type: Option<Type>,
    pub instantiation_bindings: FxHashMap<ExprId, TypeBindings>,
    field_indices: FxHashMap<ExprId, usize>,
    globals: Vec<GlobalInfo>,
    global_attributes: FxHashMap<GlobalId, Vec<SecondaryAttribute>>,
    next_type_variable_id: Cell<usize>,
    methods: FxHashMap<TypeMethodKey, FxHashMap<String, Methods>>,
    func_id_to_trait: FxHashMap<FuncId, (Type, TraitId)>,
    pub(crate) type_alias_ref: Vec<(TypeAliasId, Location)>,
    pub(crate) type_ref_locations: Vec<(Type, Location)>,
    quoted_types: Arena<Type>,
    interned_expression_kinds: Arena<ExpressionKind>,
    interned_statement_kinds: Arena<StatementKind>,
    interned_unresolved_type_data: Arena<UnresolvedTypeData>,
    interned_patterns: Arena<Pattern>,
    pub(crate) lsp_mode: bool,
    pub(crate) reference_graph: DiGraph<ReferenceId, ()>,
    pub(crate) reference_graph_indices: FxHashMap<ReferenceId, NodeIndex>,
    pub(crate) location_indices: LocationIndices,
    pub(crate) auto_import_names: FxHashMap<String, Vec<AutoImportEntry>>,
    pub(crate) comptime_scopes: Vec<FxHashMap<DefinitionId, Value>>,
    pub(crate) doc_comments: FxHashMap<ReferenceId, Vec<DocComment>>,
    pub reexports: FxHashMap<ModuleDefId, Vec<Reexport>>,
    pub primitive_docs: FxHashMap<String, Vec<DocComment>>,

}

The node interner is the central storage location of all nodes in Noir’s Hir (the various node types can be found in hir_def). The interner is also used to collect extra information about the Hir, such as the type of each node, information about each definition or struct, etc. Because it is used on the Hir, the NodeInterner is useful in passes where the Hir is used - name resolution, type checking, and monomorphization - and it is not useful afterward.

Fields
nodes: Arena<Node>func_meta: FxHashMap<FuncId, FuncMeta>function_definition_ids: FxHashMap<FuncId, DefinitionId>function_modifiers: FxHashMap<FuncId, FunctionModifiers>function_modules: FxHashMap<FuncId, ModuleId>module_attributes: FxHashMap<ModuleId, ModuleAttributes>dependency_graph: DiGraph<DependencyId, ()>

This graph tracks dependencies between different global definitions. This is used to ensure the absence of dependency cycles for globals and types.

dependency_graph_indices: FxHashMap<DependencyId, NodeIndex>

To keep track of where each DependencyId is in dependency_graph, we need this separate graph to map between the ids and indices.

id_to_location: FxHashMap<Index, Location>definitions: Vec<DefinitionInfo>id_to_type: FxHashMap<Index, Type>definition_to_type: FxHashMap<DefinitionId, Type>data_types: FxHashMap<TypeId, Shared<DataType>>type_attributes: FxHashMap<TypeId, Vec<SecondaryAttribute>>type_aliases: Vec<Shared<TypeAlias>>trait_associated_types: Vec<TraitAssociatedType>

Each trait associated type. These are tracked so that we can distinguish them from other types and know that, when directly referenced, they should also lead to an “ambiguous associated type” error.

traits: FxHashMap<TraitId, Trait>trait_implementations: FxHashMap<TraitImplId, Shared<TraitImpl>>next_trait_implementation_id: usizetrait_impl_associated_types: FxHashMap<TraitImplId, Vec<NamedType>>

The associated types for each trait impl. This is stored outside of the TraitImpl object since it is required before that object is created, when resolving the type signature of each method in the impl.

trait_impl_associated_constants: FxHashMap<TraitImplId, FxHashMap<String, (DefinitionId, Type)>>trait_implementation_map: FxHashMap<TraitId, Vec<(Type, TraitImplKind)>>

Trait implementations on each type. This is expected to always have the same length as self.trait_implementations.

For lack of a better name, this maps a trait id and type combination to a corresponding impl if one is available for the type. Due to generics, we cannot map from Type directly to impl, we need to iterate a Vec of all impls of that trait to see if any type may match. This can be further optimized later by splitting it up by type.

selected_trait_implementations: FxHashMap<ExprId, TraitImplKind>

When impls are found during type checking, we tag the function call’s Ident with the impl that was selected. For cases with where clauses, this may be an Assumed (but verified) impl. In this case the monomorphizer should have the context to get the concrete type of the object and select the correct impl itself.

infix_operator_traits: FxHashMap<BinaryOpKind, TraitId>

Holds the trait ids of the traits used for infix operator overloading

prefix_operator_traits: FxHashMap<UnaryOp, TraitId>

Holds the trait ids of the traits used for prefix operator overloading

ordering_type: Option<Type>

The Ordering type is a semi-builtin type that is the result of the comparison traits.

instantiation_bindings: FxHashMap<ExprId, TypeBindings>

Map from ExprId (referring to a Function/Method call) to its corresponding TypeBindings, filled out during type checking from instantiated variables. Used during monomorphization to map call site types back onto function parameter types, and undo this binding as needed.

field_indices: FxHashMap<ExprId, usize>

Remembers the field index a given HirMemberAccess expression was resolved to during type checking.

globals: Vec<GlobalInfo>global_attributes: FxHashMap<GlobalId, Vec<SecondaryAttribute>>next_type_variable_id: Cell<usize>methods: FxHashMap<TypeMethodKey, FxHashMap<String, Methods>>

A map from a type and method name to a function id for the method. This can resolve to potentially multiple methods if the same method name is specialized for different generics on the same type. E.g. for Struct<T>, we may have both impl Struct<u32> { fn foo(){} } and impl Struct<u8> { fn foo(){} }. If this happens, the returned Vec will have 2 entries and we’ll need to further disambiguate them by checking the type of each function.

func_id_to_trait: FxHashMap<FuncId, (Type, TraitId)>type_alias_ref: Vec<(TypeAliasId, Location)>

A list of all type aliases that are referenced in the program. Searched by LSP to resolve Locations of TypeAliass

type_ref_locations: Vec<(Type, Location)>

Stores the Location of a Type reference

quoted_types: Arena<Type>

In Noir’s metaprogramming, a noir type has the type Type. When these are spliced into quoted expressions, we preserve the original type by assigning it a unique id and creating a Token::QuotedType(id) from this id. We cannot create a token holding the actual type since types do not implement Send or Sync.

interned_expression_kinds: Arena<ExpressionKind>interned_statement_kinds: Arena<StatementKind>interned_unresolved_type_data: Arena<UnresolvedTypeData>interned_patterns: Arena<Pattern>lsp_mode: bool

Determines whether to run in LSP mode. In LSP mode references are tracked.

reference_graph: DiGraph<ReferenceId, ()>

Store the location of the references in the graph. Edges are directed from reference nodes to referenced nodes. For example:

let foo = 3;
//  referenced
//   ^
//   |
//   +------------+
let bar = foo;    |
//      reference |
//         v      |
//         |      |
//         +------+

reference_graph_indices: FxHashMap<ReferenceId, NodeIndex>

Tracks the index of the references in the graph

location_indices: LocationIndices

Store the location of the references in the graph

auto_import_names: FxHashMap<String, Vec<AutoImportEntry>>comptime_scopes: Vec<FxHashMap<DefinitionId, Value>>

Each value currently in scope in the comptime interpreter. Each element of the Vec represents a scope with every scope together making up all currently visible definitions. The first scope is always the global scope.

This is stored in the NodeInterner so that the Elaborator from each crate can share the same global values.

doc_comments: FxHashMap<ReferenceId, Vec<DocComment>>

Captures the documentation comments for each module, struct, trait, function, etc.

reexports: FxHashMap<ModuleDefId, Vec<Reexport>>

A map of ModuleDefId to each module that pub or pub(crate) exports it. This is used to offer importing the item via one of these exports if the item is not visible where it’s defined.

primitive_docs: FxHashMap<String, Vec<DocComment>>

Contains the docs comments of primitive types. These are defined in noir_stdlib/src/primitive_docs.nr using a tag attribute #['nargo_primitive_doc] on private modules.

Implementations

impl NodeInterner

pub fn reference_location(&self, reference: ReferenceId) -> Location

pub(crate) fn add_module_def_id_reference( &mut self, def_id: ModuleDefId, location: Location, is_self_type: bool, )

pub(crate) fn add_module_reference(&mut self, id: ModuleId, location: Location)

In LSP mode, take note that a ModuleId was referenced at a Location.

pub(crate) fn add_type_reference( &mut self, id: TypeId, location: Location, is_self_type: bool, )

In LSP mode, take note that a TypeId was referenced at a Location.

pub(crate) fn add_struct_member_reference( &mut self, id: TypeId, member_index: usize, location: Location, )

In LSP mode, take note that a specific field of a struct was referenced at a Location.

pub(crate) fn add_trait_reference( &mut self, id: TraitId, location: Location, is_self_type: bool, )

In LSP mode, take note that a TraitId was referenced at a Location.

pub(crate) fn add_trait_associated_type_reference( &mut self, id: TraitAssociatedTypeId, location: Location, )

In LSP mode, take note that a TraitAssociatedTypeId was referenced at a Location.

pub(crate) fn add_alias_reference( &mut self, id: TypeAliasId, location: Location, )

In LSP mode, take note that a TypeAliasId was referenced at a Location.

pub(crate) fn add_function_reference(&mut self, id: FuncId, location: Location)

In LSP mode, take note that a FuncId was referenced at a Location.

pub(crate) fn add_global_reference(&mut self, id: GlobalId, location: Location)

In LSP mode, take note that a GlobalId was referenced at a Location.

pub(crate) fn add_local_reference( &mut self, id: DefinitionId, location: Location, )

In LSP mode, take note that a DefinitionId was referenced at a Location.

pub(crate) fn add_reference( &mut self, referenced: ReferenceId, location: Location, is_self_type: bool, )

In LSP mode, take note that a ReferenceId was referenced at a Location.

pub(crate) fn add_definition_location( &mut self, referenced: ReferenceId, referenced_location: Location, )

pub(crate) fn get_or_insert_reference( &mut self, id: ReferenceId, ) -> PetGraphIndex

pub fn find_referenced_location( &self, reference_location: Location, ) -> Option<Location>

pub fn reference_at_location(&self, location: Location) -> Option<ReferenceId>

pub fn find_all_references( &self, location: Location, include_referenced: bool, include_self_type_name: bool, ) -> Option<Vec<Location>>

pub fn find_referenced(&self, location: Location) -> Option<ReferenceId>

fn find_all_references_for_index( &self, referenced_node_index: PetGraphIndex, include_referenced: bool, include_self_type_name: bool, ) -> Vec<Location>

fn referenced_index( &self, reference_index: PetGraphIndex, ) -> Option<PetGraphIndex>

pub(crate) fn register_module( &mut self, id: ModuleId, location: Location, visibility: ItemVisibility, name: String, )

pub(crate) fn register_global( &mut self, id: GlobalId, name: String, location: Location, visibility: ItemVisibility, )

pub(crate) fn register_type( &mut self, id: TypeId, name: String, location: Location, visibility: ItemVisibility, )

pub(crate) fn register_trait( &mut self, id: TraitId, name: String, location: Location, visibility: ItemVisibility, )

pub(crate) fn register_type_alias( &mut self, id: TypeAliasId, name: String, location: Location, visibility: ItemVisibility, )

pub(crate) fn register_function( &mut self, id: FuncId, func_def: &FunctionDefinition, )

pub fn register_name_for_auto_import( &mut self, name: String, module_def_id: ModuleDefId, visibility: ItemVisibility, defining_module: Option<ModuleId>, )

pub fn get_auto_import_names(&self) -> &HashMap<String, Vec<AutoImportEntry>>

impl NodeInterner

pub fn dependency_graph(&self) -> &DiGraph<DependencyId, ()>

Gets the dependency graph from the node interner.

pub fn add_type_dependency( &mut self, dependent: DependencyId, dependency: TypeId, )

Register that dependent depends on dependency. This is usually because dependent refers to dependency in one of its struct fields.

pub fn add_global_dependency( &mut self, dependent: DependencyId, dependency: GlobalId, )

Mark a DependencyId as being dependant on a GlobalId.

pub fn add_function_dependency( &mut self, dependent: DependencyId, dependency: FuncId, )

Mark a DependencyId as being dependant on a FuncId.

pub fn add_type_alias_dependency( &mut self, dependent: DependencyId, dependency: TypeAliasId, )

Mark a DependencyId as being dependant on a TypeAliasId.

pub fn add_trait_dependency( &mut self, dependent: DependencyId, dependency: TraitId, )

Mark a DependencyId as being dependant on a TraitId.

pub fn add_dependency( &mut self, dependent: DependencyId, dependency: DependencyId, )

pub fn get_or_insert_dependency(&mut self, id: DependencyId) -> PetGraphIndex

pub(crate) fn check_for_dependency_cycles(&self) -> Vec<CompilationError>

fn get_cycle_error_string( &self, scc: &[PetGraphIndex], start_index: usize, ) -> String

Build up a string starting from the given item containing each item in the dependency cycle. The final result will resemble foo -> bar -> baz -> foo, always going back to the element at the given start index.

impl NodeInterner

pub fn push_empty_fn(&mut self) -> FuncId

Intern an empty function.

pub fn update_fn(&mut self, func_id: FuncId, hir_func: HirFunction)

Updates the underlying interned Function.

This method is used as we eagerly intern empty functions to generate function identifiers and then we update at a later point in time.

pub fn find_function(&self, function_name: &str) -> Option<FuncId>

pub fn push_fn_meta(&mut self, func_data: FuncMeta, func_id: FuncId)

Interns a function’s metadata.

Note that the FuncId has been created already. See ModCollector for it’s usage.

pub fn push_function( &mut self, id: FuncId, function: &FunctionDefinition, module: ModuleId, location: Location, ) -> DefinitionId

pub fn push_function_definition( &mut self, func: FuncId, modifiers: FunctionModifiers, module: ModuleId, location: Location, ) -> DefinitionId

pub fn set_function_trait( &mut self, func: FuncId, self_type: Type, trait_id: TraitId, )

pub fn get_function_trait(&self, func: &FuncId) -> Option<(Type, TraitId)>

pub fn function_visibility(&self, func: FuncId) -> ItemVisibility

Returns the visibility of the given function.

The underlying function_visibilities map is populated during def collection, so this function can be called anytime afterward.

pub fn function_module(&self, func: FuncId) -> ModuleId

Returns the module this function was defined within

pub fn lookup_function_from_expr(&self, expr: &ExprId) -> Option<FuncId>

Returns the FuncId corresponding to the function referred to by expr_id

pub fn function(&self, func_id: &FuncId) -> HirFunction

Returns the interned HIR function corresponding to func_id

pub fn function_meta(&self, func_id: &FuncId) -> &FuncMeta

Returns the interned meta data corresponding to func_id

pub fn function_meta_mut(&mut self, func_id: &FuncId) -> &mut FuncMeta

pub fn try_function_meta(&self, func_id: &FuncId) -> Option<&FuncMeta>

pub fn function_ident(&self, func_id: &FuncId) -> Ident

pub fn function_name(&self, func_id: &FuncId) -> &str

pub fn function_modifiers(&self, func_id: &FuncId) -> &FunctionModifiers

pub fn function_modifiers_mut( &mut self, func_id: &FuncId, ) -> &mut FunctionModifiers

pub fn function_attributes(&self, func_id: &FuncId) -> &Attributes

impl NodeInterner

pub fn next_global_id(&mut self) -> GlobalId

fn push_global( &mut self, ident: Ident, local_id: LocalModuleId, crate_id: CrateId, let_statement: StmtId, file: FileId, attributes: Vec<SecondaryAttribute>, mutable: bool, comptime: bool, visibility: ItemVisibility, ) -> GlobalId

pub fn push_empty_global( &mut self, name: Ident, local_id: LocalModuleId, crate_id: CrateId, file: FileId, attributes: Vec<SecondaryAttribute>, mutable: bool, comptime: bool, visibility: ItemVisibility, ) -> GlobalId

Intern an empty global. Used for collecting globals before they’re defined

pub fn get_global(&self, global_id: GlobalId) -> &GlobalInfo

pub fn get_global_mut(&mut self, global_id: GlobalId) -> &mut GlobalInfo

pub fn get_global_definition(&self, global_id: GlobalId) -> &DefinitionInfo

pub fn get_global_definition_mut( &mut self, global_id: GlobalId, ) -> &mut DefinitionInfo

pub fn get_all_globals(&self) -> &[GlobalInfo]

pub fn get_global_let_statement( &self, global: GlobalId, ) -> Option<HirLetStatement>

Try to get the HirLetStatement which defines a given global value

impl NodeInterner

pub fn add_reexport( &mut self, module_def_id: ModuleDefId, module_id: ModuleId, name: Ident, visibility: ItemVisibility, )

pub fn get_reexports(&self, module_def_id: ModuleDefId) -> &[Reexport]

pub fn get_trait_reexports(&self, trait_id: TraitId) -> &[Reexport]

impl NodeInterner

pub fn next_trait_impl_id(&mut self) -> TraitImplId

Returns what the next trait impl id is expected to be.

pub fn trait_implementations(&self) -> &HashMap<TraitImplId, Shared<TraitImpl>>

Gets the trait implementations from the node interner.

pub fn try_get_trait_implementation( &self, id: TraitImplId, ) -> Option<Shared<TraitImpl>>

pub fn get_trait_implementation(&self, id: TraitImplId) -> Shared<TraitImpl>

pub fn get_trait_implementations_in_crate( &self, crate_id: CrateId, ) -> HashSet<TraitImplId>

pub fn add_assumed_trait_implementation( &mut self, object_type: Type, trait_id: TraitId, trait_generics: TraitGenerics, ) -> bool

Adds an “assumed” trait implementation to the currently known trait implementations. Unlike normal trait implementations, these are only assumed to exist. They often correspond to where clauses in functions where we assume there is some T: Eq even though we do not yet know T. For these cases, we store an impl here so that we assume they exist and can resolve them. They are then later verified when the function is called, and linked properly after being monomorphized to the correct variant.

Returns true on success, or false if there is already an overlapping impl in scope.

pub fn add_trait_implementation( &mut self, object_type: Type, trait_id: TraitId, impl_id: TraitImplId, impl_generics: GenericTypeVars, trait_impl: Shared<TraitImpl>, ) -> Result<(), Location>

Adds a trait implementation to the list of known implementations.

pub(crate) fn lookup_trait_implementation( &self, object_type: &Type, trait_id: TraitId, trait_generics: &[Type], trait_associated_types: &[NamedType], ) -> Result<(TraitImplKind, TypeBindings), ImplSearchErrorKind>

Given a ObjectType: TraitId pair, try to find an existing impl that satisfies the constraint. If an impl cannot be found, this will return a vector of each constraint in the path to get to the failing constraint. Usually this is just the single failing constraint, but when where clauses are involved, the failing constraint may be several constraints deep. In this case, all of the constraints are returned, starting with the failing one. If this list of failing constraints is empty, this means type annotations are required. Returns the list of instantiation bindings as well, which should be stored on the expression.

pub(crate) fn try_lookup_trait_implementation( &self, object_type: &Type, trait_id: TraitId, trait_generics: &[Type], trait_associated_types: &[NamedType], ) -> Result<(TraitImplKind, TypeBindings, TypeBindings), ImplSearchErrorKind>

Similar to lookup_trait_implementation but does not apply any type bindings on success. On error returns either:

1+ failing trait constraints, including the original. Each constraint after the first represents a where clause that was followed.

0 trait constraints indicating type annotations are needed to choose an impl.

fn lookup_trait_implementation_helper( &self, object_type: &Type, trait_id: TraitId, trait_generics: &[Type], trait_associated_types: &[NamedType], type_bindings: &mut TypeBindings, recursion_limit: u32, ) -> Result<(TraitImplKind, TypeBindings), ImplSearchErrorKind>

Returns the trait implementation if found along with the instantiation bindings for instantiating that trait impl. Note that this is separate from the passed-in TypeBindings which can be bound via Type::apply_type_bindings if needed. Instantiation bindings should be stored as such but not bound, lest the original named generics in trait impls get bound over.

On error returns either:

1+ failing trait constraints, including the original. Each constraint after the first represents a where clause that was followed.

0 trait constraints indicating type annotations are needed to choose an impl.

fn validate_where_clause( &self, where_clause: &[TraitConstraint], type_bindings: &mut TypeBindings, instantiation_bindings: &TypeBindings, recursion_limit: u32, ) -> Result<(), (TraitConstraint, ImplSearchErrorKind)>

Verifies that each constraint in the given where clause is valid. If an impl cannot be found for any constraint, the erroring constraint is returned.

pub(crate) fn trait_constraint_string( &self, object_type: &Type, trait_id: TraitId, trait_generics: &[Type], trait_associated_types: &[NamedType], ) -> String

pub fn remove_assumed_trait_implementations_for_trait( &mut self, trait_id: TraitId, )

Removes all TraitImplKind::Assumed from the list of known impls for the given trait

fn remove_assumed_trait_implementations_for_trait_and_parents( &mut self, trait_id: TraitId, starting_trait_id: TraitId, )

impl NodeInterner

pub fn push_stmt_full( &mut self, stmt: HirStatement, location: Location, ) -> StmtId

Intern a HIR statement with everything needed for it (location).

pub fn push_expr(&mut self, expr: HirExpression) -> PushedExpr

Interns a HIR expression, with the location and type information pushed as follow ups.

pub fn push_expr_full( &mut self, expr: HirExpression, location: Location, typ: Type, ) -> ExprId

Intern an expression with everything needed for it (location & type) instead of requiring they be pushed later.

pub fn push_expr_location(&mut self, expr_id: ExprId, location: Location)

Stores the span for an interned expression.

pub fn push_fn(&mut self, func: HirFunction) -> FuncId

Interns a HIR Function.

pub fn push_expr_type(&mut self, expr_id: ExprId, typ: Type)

Store the type for an interned expression

pub fn push_definition_type(&mut self, definition_id: DefinitionId, typ: Type)

Store the type for a definition

pub fn push_empty_trait( &mut self, type_id: TraitId, unresolved_trait: &UnresolvedTrait, generics: Generics, associated_types: Generics, associated_constant_ids: HashMap<String, DefinitionId>, )

pub fn new_type( &mut self, name: Ident, span: Span, attributes: Vec<SecondaryAttribute>, generics: Generics, visibility: ItemVisibility, krate: CrateId, local_id: LocalModuleId, file_id: FileId, ) -> TypeId

Creates a new struct or enum type with no fields or variants.

pub fn push_type_alias( &mut self, typ: &UnresolvedTypeAlias, generics: Generics, ) -> TypeAliasId

pub fn add_type_alias_ref(&mut self, type_id: TypeAliasId, location: Location)

Adds TypeAliasId and Location to the type_alias_ref vector So that we can later resolve Locations type aliases from the LSP requests

pub fn push_trait_associated_type( &mut self, trait_id: TraitId, name: Ident, ) -> TraitAssociatedTypeId

pub fn update_type(&mut self, type_id: TypeId, f: impl FnOnce(&mut DataType))

pub fn update_trait(&mut self, trait_id: TraitId, f: impl FnOnce(&mut Trait))

pub fn update_type_attributes( &mut self, type_id: TypeId, f: impl FnOnce(&mut Vec<SecondaryAttribute>), )

pub fn set_type_alias( &mut self, type_id: TypeAliasId, typ: Type, generics: Generics, num_expr: Option<UnresolvedTypeExpression>, )

pub fn update_statement( &mut self, stmt_id: &StmtId, f: impl FnOnce(&mut HirStatement), )

Returns the interned statement corresponding to stmt_id

pub fn update_expression( &mut self, expr_id: ExprId, f: impl FnOnce(&mut HirExpression), )

Updates the interned expression corresponding to expr_id

pub fn push_type_ref_location(&mut self, typ: &Type, location: Location)

In LSP mode, take note that the Type was referenced at a Location.

pub fn definition_count(&self) -> usize

pub fn push_definition( &mut self, name: String, mutable: bool, comptime: bool, definition: DefinitionKind, location: Location, ) -> DefinitionId

pub fn type_attributes(&self, struct_id: &TypeId) -> &Vec<SecondaryAttribute>

pub fn add_module_attributes( &mut self, module_id: ModuleId, attributes: ModuleAttributes, )

pub fn module_attributes(&self, module_id: ModuleId) -> &ModuleAttributes

pub fn try_module_attributes( &self, module_id: ModuleId, ) -> Option<&ModuleAttributes>

pub fn global_attributes(&self, global_id: &GlobalId) -> &[SecondaryAttribute]

pub fn statement(&self, stmt_id: &StmtId) -> HirStatement

Returns the interned statement corresponding to stmt_id

pub fn expression(&self, expr_id: &ExprId) -> HirExpression

Returns the interned expression corresponding to expr_id

pub fn definition(&self, id: DefinitionId) -> &DefinitionInfo

Retrieves the definition where the given id was defined. This will panic if given DefinitionId::dummy_id. Use try_definition for any call with a possibly undefined variable.

pub fn definition_mut(&mut self, id: DefinitionId) -> &mut DefinitionInfo

Retrieves the definition where the given id was defined. This will panic if given DefinitionId::dummy_id. Use try_definition for any call with a possibly undefined variable.

pub fn try_definition(&self, id: DefinitionId) -> Option<&DefinitionInfo>

Tries to retrieve the given id’s definition. This function should be used during name resolution or type checking when we cannot be sure all variables have corresponding definitions (in case of an error in the user’s code).

pub fn definition_name(&self, id: DefinitionId) -> &str

Returns the name of the definition

This is needed as the Environment needs to map variable names to witness indices

pub fn expr_span(&self, expr_id: &ExprId) -> Span

pub fn try_expr_span(&self, expr_id: &ExprId) -> Option<Span>

pub fn expr_location(&self, expr_id: &ExprId) -> Location

pub fn statement_span(&self, stmt_id: StmtId) -> Span

pub fn statement_location(&self, stmt_id: StmtId) -> Location

pub fn push_stmt_location(&mut self, id: StmtId, location: Location)

pub fn get_type(&self, id: TypeId) -> Shared<DataType>

pub fn get_type_methods(&self, typ: &Type) -> Option<&HashMap<String, Methods>>

pub fn get_trait(&self, id: TraitId) -> &Trait

pub fn get_trait_associated_type( &self, id: TraitAssociatedTypeId, ) -> &TraitAssociatedType

pub fn get_trait_mut(&mut self, id: TraitId) -> &mut Trait

pub fn try_get_trait(&self, id: TraitId) -> Option<&Trait>

pub fn get_type_alias(&self, id: TypeAliasId) -> Shared<TypeAlias>

pub fn id_type(&self, index: impl Into<Index>) -> Type

Returns the type of an item stored in the NodeInterner, or Type::Error if it was not found.

pub fn try_id_type(&self, index: impl Into<Index>) -> Option<&Type>

pub fn definition_type(&self, id: DefinitionId) -> Type

Returns the type of the definition, or Type::Error if it was not found.

pub fn id_type_substitute_trait_as_type(&self, def_id: DefinitionId) -> Type

Returns the type of the definition, unless it’s a function returning an impl Trait, in which case it looks up the type of its body and returns a new function type with the type fo the body substituted to its return type.

pub fn id_location(&self, index: impl Into<Index> + Copy) -> Location

Returns the span of an item stored in the Interner

pub fn try_id_location(&self, index: impl Into<Index>) -> Option<Location>

Returns the span of an item stored in the Interner, if present

pub fn replace_expr(&mut self, id: &ExprId, new: HirExpression)

Replaces the HirExpression at the given ExprId with a new HirExpression

pub fn replace_statement(&mut self, stmt_id: StmtId, hir_stmt: HirStatement)

Replaces the HirStatement at the given StmtId with a new HirStatement

pub fn next_type_variable_id(&self) -> TypeVariableId

pub fn next_type_variable(&self) -> Type

pub fn next_type_variable_with_kind(&self, kind: Kind) -> Type

pub fn store_instantiation_bindings( &mut self, expr_id: ExprId, instantiation_bindings: TypeBindings, )

Remember the TypeBindings used during the instantiation of an expression.

pub fn get_instantiation_bindings(&self, expr_id: ExprId) -> &TypeBindings

pub fn try_get_instantiation_bindings( &self, expr_id: ExprId, ) -> Option<&TypeBindings>

pub fn get_field_index(&self, expr_id: ExprId) -> usize

pub fn set_field_index(&mut self, expr_id: ExprId, index: usize)

pub fn function_definition_id(&self, function: FuncId) -> DefinitionId

Look up the DefinitionId of a FuncId.

Panics if it’s not found.

pub fn get_trait_item_id(&self, function_id: FuncId) -> Option<TraitItemId>

Returns the definition id and trait id for a given trait or impl function.

If this is an impl function, the DefinitionId inside the TraitItemId will still be that of the function in the parent trait.

pub fn add_method( &mut self, self_type: &Type, method_name: String, method_id: FuncId, trait_id: Option<TraitId>, ) -> Option<FuncId>

Adds a non-trait method to a type.

Returns Some(duplicate) if a matching method was already defined. Returns None otherwise.

pub fn lookup_direct_method( &self, typ: &Type, method_name: &str, check_self_param: bool, ) -> Option<FuncId>

Looks up a method that’s directly defined in the given type. If check_self_param is true, only a method that has a self parameter with a type that unifies with typ will be returned.

pub fn lookup_trait_methods( &self, typ: &Type, method_name: &str, has_self_arg: bool, ) -> Vec<(FuncId, TraitId)>

Looks up a methods that apply to the given type but are defined in traits.

pub fn lookup_generic_methods( &self, typ: &Type, method_name: &str, has_self_arg: bool, ) -> Vec<(FuncId, TraitId)>

Looks up methods at impls for all types T, e.g. impl<T> Foo for T

pub fn select_impl_for_expression( &mut self, ident_id: ExprId, trait_impl: TraitImplKind, )

Tags the given identifier with the selected trait_impl so that monomorphization can later recover which impl was selected, or alternatively see if it needs to decide which impl to select (because the impl was Assumed).

pub fn get_selected_impl_for_expression( &self, ident_id: ExprId, ) -> Option<TraitImplKind>

Retrieves the impl selected for a given ExprId during name resolution.

pub fn get_operator_trait_method(&self, operator: BinaryOpKind) -> TraitItemId

Retrieves the trait id for a given binary operator. All binary operators correspond to a trait - although multiple may correspond to the same trait (such as == and !=). self.infix_operator_traits is expected to be filled before name resolution, during definition collection.

pub fn get_prefix_operator_trait_method( &self, operator: &UnaryOp, ) -> Option<TraitItemId>

Retrieves the trait id for a given unary operator. Only some unary operators correspond to a trait: - and !, but for example * does not. self.prefix_operator_traits is expected to be filled before name resolution, during definition collection.

pub fn try_add_infix_operator_trait(&mut self, trait_id: TraitId)

Add the given trait as an operator trait if its name matches one of the operator trait names (Add, Sub, …).

pub fn try_add_prefix_operator_trait(&mut self, trait_id: TraitId)

Add the given trait as an operator trait if its name matches one of the prefix operator trait names (Not or Neg).

pub fn is_operator_trait(&self, trait_id: TraitId) -> bool

pub fn populate_dummy_operator_traits(&mut self)

This function is needed when creating a NodeInterner for testing so that calls to get_operator_trait do not panic when the stdlib isn’t present.

pub(crate) fn ordering_type(&self) -> Type

pub fn push_quoted_type(&mut self, typ: Type) -> QuotedTypeId

pub fn get_quoted_type(&self, id: QuotedTypeId) -> &Type

pub fn push_expression_kind( &mut self, expr: ExpressionKind, ) -> InternedExpressionKind

Intern a ExpressionKind.

pub fn get_expression_kind(&self, id: InternedExpressionKind) -> &ExpressionKind

Get an interned ExpressionKind by its InternedExpressionKind ID.

pub fn push_statement_kind( &mut self, statement: StatementKind, ) -> InternedStatementKind

Intern a StatementKind.

pub fn get_statement_kind(&self, id: InternedStatementKind) -> &StatementKind

Get an interned StatementKind by its InternedStatementKind ID.

pub fn push_lvalue(&mut self, lvalue: LValue) -> InternedExpressionKind

Intern an LValue by turning it into an Expression and interning its ExpressionKind.

pub fn get_lvalue( &self, id: InternedExpressionKind, location: Location, ) -> LValue

Get an interned LValue by its InternedExpressionKind ID.

pub fn push_pattern(&mut self, pattern: Pattern) -> InternedPattern

Intern a Pattern.

pub fn get_pattern(&self, id: InternedPattern) -> &Pattern

Get an interned Pattern by its InternedPattern ID.

pub fn push_unresolved_type_data( &mut self, typ: UnresolvedTypeData, ) -> InternedUnresolvedTypeData

Intern a UnresolvedTypeData.

pub fn get_unresolved_type_data( &self, id: InternedUnresolvedTypeData, ) -> &UnresolvedTypeData

pub fn get_infix_operator_type( &self, lhs: ExprId, operator: BinaryOpKind, operator_expr: ExprId, ) -> (Type, Type)

Returns the type of an operator (which is always a function), along with its return type.

pub fn get_prefix_operator_type( &self, operator_expr: ExprId, rhs: ExprId, ) -> (Type, Type)

Returns the type of a prefix operator (which is always a function), along with its return type.

pub fn is_in_lsp_mode(&self) -> bool

pub(crate) fn set_associated_types_for_impl( &mut self, impl_id: TraitImplId, associated_types: Vec<NamedType>, )

Sets the associated types for the given trait impl. Each type in NamedType will be wrapped in a Type::TypeVariable if it’s of kind Kind::Numeric.

pub fn get_associated_types_for_impl( &self, impl_id: TraitImplId, ) -> &[NamedType]

Returns the associated types for the given trait impl. The Type of each NamedType that is an associated constant is guaranteed to be a Type::TypeVariable.

pub fn find_associated_type_for_impl( &self, impl_id: TraitImplId, type_name: &str, ) -> Option<&Type>

pub fn get_trait_impl_associated_constant( &self, impl_id: TraitImplId, name: &str, ) -> Option<&(DefinitionId, Type)>

Returns the definition id for the associated constant of the given type variable.

pub fn trait_to_impl_bindings( &self, trait_id: TraitId, impl_id: TraitImplId, trait_impl_generics: &[Type], impl_self_type: Type, ) -> TypeBindings

Return a set of TypeBindings to bind types from the parent trait to those from the trait impl.

pub fn set_doc_comments( &mut self, id: ReferenceId, doc_comments: Vec<DocComment>, )

pub fn doc_comments(&self, id: ReferenceId) -> Option<&Vec<DocComment>>

pub fn get_expr_id_from_index(&self, index: impl Into<Index>) -> Option<ExprId>

pub fn get_meta_attribute_name(&self, meta: &MetaAttribute) -> Option<String>

impl NodeInterner

pub fn find_location_index( &self, location: Location, ) -> Option<impl Into<Index>>

Scans the interner for the item which is located at that Location

The Location may not necessarily point to the beginning of the item so we check if the location’s span is contained within the start or end of each items Location

pub fn type_at_location(&self, location: Location) -> Option<&Type>

Returns the Type of the expression that exists at the given location.

pub fn get_definition_location_from( &self, location: Location, return_type_location_instead: bool, ) -> Option<Location>

Returns the Location of the definition of the given Ident found at location. Returns None when definition is not found.

pub fn get_declaration_location_from( &self, location: Location, ) -> Option<Location>

fn resolve_location( &self, index: impl Into<Index>, return_type_location_instead: bool, ) -> Option<Location>

For a given [Index] we return Location to which we resolved to We currently return None for features not yet implemented TODO(#3659): LSP goto def should error when Ident at Location could not resolve

fn get_type_location_from_index( &self, index: impl Into<Index>, ) -> Option<Location>

fn resolve_expression_location( &self, expression: &HirExpression, return_type_location_instead: bool, ) -> Option<Location>

Resolves the Location of the definition for a given HirExpression

Note: current the code returns None because some expressions are not yet implemented.

fn resolve_struct_member_access( &self, expr_member_access: &HirMemberAccess, ) -> Option<Location>

Resolves the Location of the definition for a given crate::hir_def::expr::HirMemberAccess This is used to resolve the location of a struct member access. For example, in the expression foo.bar we want to resolve the location of bar to the location of the definition of bar in the struct foo.

fn try_resolve_trait_impl_location( &self, location: Location, ) -> Option<Location>

Attempts to resolve Location of Trait based on Location of TraitImpl This is used by LSP to resolve the location of a trait based on the location of a trait impl.

Example: impl Foo for Bar { … } -> trait Foo { … }

fn try_resolve_trait_method_declaration( &self, location: Location, ) -> Option<Location>

Attempts to resolve Location of Trait’s TraitFunction declaration based on the Location of a TraitFunction call.

This is used by LSP to resolve the location.

Example:

trait HasToField {
    fn to_field(self) -> Field;
       ^------------------------------\
}                                     |    
                                      |
fn main_func(x: u32) {                |
    assert(x.to_field() == 15);       |
              \......................./
}

pub(crate) fn try_resolve_type_ref( &self, location: Location, ) -> Option<Location>

Attempts to resolve Location of Type based on Location of reference in code

pub fn try_type_ref_at_location(&self, location: Location) -> Option<Type>

fn try_resolve_type_alias(&self, location: Location) -> Option<Location>