noirc_frontend::ownership

Function builtin_supports_clone_elision

pub fn builtin_supports_clone_elision(builtin: Builtin) -> bool

Return whether a call to the given builtin or low level function allows the Clone around its array arguments to be elided: the callee must neither modify the input nor return an alias of it that a later Brillig mutation could observe.

This mirrors Intrinsic::purity and Intrinsic::unsafe_for_clone_elision_in_brillig in noirc_evaluator, which cannot be used here directly since noirc_frontend cannot depend on noirc_evaluator. A test in noirc_evaluator (ssa::ssa_gen::tests) checks that this function agrees with them for every builtin.