noirc_frontend::ownership
Struct Context

struct Context {
    variables_to_move: FxHashMap<LocalId, Vec<IdentId>>,
}

Fields
variables_to_move: FxHashMap<LocalId, Vec<IdentId>>

This contains each instance of a variable we should move instead of cloning.

Implementations

impl Context

pub(super) fn find_last_uses_of_variables( function: &Function, ) -> HashMap<LocalId, Vec<IdentId>>

Traverse the given function and return the last use(s) of each local variable. A variable may have multiple last uses if it was last used within a conditional expression.

impl Context

fn should_move(&self, definition: LocalId, variable: IdentId) -> bool

fn handle_ownership_in_function(&mut self, function: &mut Function)

fn handle_expression(&mut self, expr: &mut Expression)

fn handle_reference_expression(&mut self, expr: &mut Expression)

Handle the rhs of a &expr unary expression. Variables and field accesses in these expressions are exempt from clones.

Note that this also matches on dereference operations to exempt their LHS from clones, but their LHS is always exempt from clones so this is unchanged.

Returns

A boolean representing whether or not the expression was borrowed by reference (false) or moved (true).

fn handle_extract_expression(&mut self, expr: &mut Expression)

fn handle_extract_expression_rec( &mut self, expr: &mut Expression, ) -> Option<(bool, Type)>

Traverse an expression comprised of only identifiers, tuple field extractions, and dereferences returning whether we should clone the result and the type of that result. Returns None if a different expression variant was found.

fn should_clone_ident(&self, ident: &Ident) -> bool

Whenever an ident is used it is always cloned unless it is the last use of the ident (not in a loop).

fn handle_ident(&self, expr: &mut Expression)

fn handle_literal(&mut self, literal: &mut Literal)

fn handle_unary(&mut self, expr: &mut Expression)

fn handle_binary(&mut self, binary: &mut Binary)

fn handle_index(&mut self, index_expr: &mut Expression)

fn handle_cast(&mut self, cast: &mut Cast)

fn handle_for(&mut self, for_expr: &mut For)

fn handle_while(&mut self, while_expr: &mut While)

fn handle_if(&mut self, if_expr: &mut If)

fn handle_match(&mut self, match_expr: &mut Match)

fn handle_tuple(&mut self, elements: &mut [Expression])

fn handle_call(&mut self, call: &mut Call)

fn handle_let(&mut self, let_expr: &mut Let)

fn handle_constrain( &mut self, boolean: &mut Expression, msg: &mut Option<Box<(Expression, Type)>>, )

fn handle_assign(&mut self, assign: &mut Assign)

fn handle_lvalue(&mut self, lvalue: &mut LValue)