noirc_frontend::parser
Struct Parser

pub struct Parser<'a> {
    pub(crate) errors: Vec<ParserError>,
    tokens: TokenStream<'a>,
    token: LocatedToken,
    next_token: LocatedToken,
    current_token_location: Location,
    previous_token_location: Location,
    current_token_comments: String,
    next_token_comments: String,
    statement_comments: Option<String>,
}

Fields
errors: Vec<ParserError>tokens: TokenStream<'a>token: LocatedTokennext_token: LocatedTokencurrent_token_location: Locationprevious_token_location: Locationcurrent_token_comments: Stringnext_token_comments: Stringstatement_comments: Option<String>

The current statement’s comments. This is used to eventually know if an unsafe { ... } expression is commented in its containing statement. For example:

// Safety: test
let x = unsafe { call() };

Implementations

impl Parser<'_>

pub(crate) fn parse_arguments(&mut self) -> Option<Vec<Expression>>

Arguments = ‘(’ ArgumentsList? ‘)’

ArgumentsList = Expression ( ‘,’ Expression )? ‘,’?

pub(super) fn parse_call_arguments(&mut self) -> Option<CallArguments>

CallArguments = ‘!’? Arguments

impl Parser<'_>

pub(super) fn parse_inner_attribute(&mut self) -> Option<SecondaryAttribute>

InnerAttribute = ‘#![’ SecondaryAttribute ‘]’

pub(super) fn parse_attributes(&mut self) -> Vec<(Attribute, Location)>

Attributes = Attribute*

pub(crate) fn parse_attribute(&mut self) -> Option<(Attribute, Location)>

Attribute = ‘#[’ (FunctionAttribute | SecondaryAttribute) ‘]’

FunctionAttribute = ‘builtin’ ‘(’ AttributeValue ‘)’ | ‘fold’ | ‘foreign’ ‘(’ AttributeValue ‘)’ | ‘inline_always’ | ‘no_predicates’ | ‘oracle’ ‘(’ AttributeValue ‘)’ | ‘recursive’ | ‘test’ | ‘test’ ‘(’ ‘should_fail’ ‘)’ | ‘test’ ‘(’ ‘should_fail_with’ ‘=’ string ‘)’ | ‘fuzz’ | ‘fuzz’ ‘(’ ‘only_fail_with’ ‘=’ string ‘)’ | ‘fuzz’ ‘(’ ‘should_fail’ ‘)’ | ‘fuzz’ ‘(’ ‘should_fail_with’ ‘=’ string ‘)’

SecondaryAttribute = ‘abi’ ‘(’ AttributeValue ‘)’ | ‘allow’ ‘(’ AttributeValue ‘)’ | ‘deprecated’ | ‘deprecated’ ‘(’ string ‘)’ | ‘contract_library_method’ | ‘export’ | ‘field’ ‘(’ AttributeValue ‘)’ | ‘use_callers_scope’ | ‘varargs’ | MetaAttribute

MetaAttribute = Path Arguments?

AttributeValue = Path | integer

pub(super) fn validate_secondary_attributes( &mut self, attributes: Vec<(Attribute, Location)>, ) -> Vec<SecondaryAttribute>

fn parse_tag_attribute(&mut self, start_location: Location) -> Attribute

fn parse_non_tag_attribute(&mut self, start_location: Location) -> Attribute

fn parse_meta_attribute( &mut self, name: MetaAttributeName, start_location: Location, ) -> Attribute

fn parse_ident_attribute_other_than_test_and_fuzz( &mut self, ident: &Ident, start_location: Location, ) -> Attribute

fn parse_deprecated_attribute( &mut self, ident: &Ident, arguments: Vec<Expression>, ) -> SecondaryAttributeKind

fn parse_test_attribute(&mut self, start_location: Location) -> Attribute

fn parse_fuzz_attribute(&mut self, start_location: Location) -> Attribute

fn parse_must_use_attribute(&mut self, start_location: Location) -> Attribute

fn parse_single_name_attribute<F>( &mut self, ident: &Ident, arguments: Vec<Expression>, start_location: Location, f: F, ) -> Attribute
where F: FnOnce(String) -> Attribute,

fn parse_no_args_attribute( &mut self, ident: &Ident, arguments: Vec<Expression>, attribute: Attribute, ) -> Attribute

fn skip_until_right_bracket(&mut self, issue_error: bool)

impl Parser<'_>

pub(super) fn parse_inner_doc_comments(&mut self) -> Vec<DocComment>

InnerDocComments = inner_doc_comment*

fn parse_inner_doc_comment(&mut self) -> Option<DocComment>

pub(super) fn parse_outer_doc_comments(&mut self) -> Vec<DocComment>

OuterDocComments = OuterDocComment*

pub(super) fn parse_outer_doc_comment(&mut self) -> Option<DocComment>

OuterDocComment = outer_doc_comment

pub(super) fn warn_on_outer_doc_comments(&mut self)

Skips any outer doc comments but produces a warning saying that they don’t document anything.

pub(super) fn error_on_outer_doc_comments_on_parameter(&mut self)

Skips any outer doc comments but produces an error saying that they can’t be applied to parameters

fn skip_doc_comments_with_reason(&mut self, reason: ParserErrorReason)

impl Parser<'_>

pub(crate) fn parse_enum( &mut self, attributes: Vec<(Attribute, Location)>, visibility: ItemVisibility, start_location: Location, ) -> NoirEnumeration

Enum = ‘enum’ identifier Generics ‘{’ EnumVariant* ‘}’

EnumField = OuterDocComments identifier ‘:’ Type

fn parse_enum_variant(&mut self) -> Option<Documented<EnumVariant>>

fn empty_enum( &self, name: Ident, attributes: Vec<SecondaryAttribute>, visibility: ItemVisibility, generics: UnresolvedGenerics, start_location: Location, ) -> NoirEnumeration

impl Parser<'_>

pub(crate) fn parse_expression_or_error(&mut self) -> Expression

pub(crate) fn parse_expression(&mut self) -> Option<Expression>

Expression = EqualOrNotEqualExpression

pub(crate) fn parse_expression_except_constructor_or_error( &mut self, ) -> Expression

When parsing if conditions we don’t allow constructors. For example if foo { 1 } shouldn’t have foo { 1 } as the condition, but foo instead. The same goes with for: for x in foo { 1 } should have foo as the collection, not foo { 1 }.

ExpressionExceptConstructor = “Expression except ConstructorException”

pub(crate) fn parse_expression_or_error_impl( &mut self, allow_constructors: bool, ) -> Expression

fn parse_expression_impl( &mut self, allow_constructors: bool, ) -> Option<Expression>

pub(super) fn parse_term( &mut self, allow_constructors: bool, ) -> Option<Expression>

Term = UnaryExpression | CastExpression

fn parse_unary(&mut self, allow_constructors: bool) -> Option<Expression>

UnaryExpression = UnaryOp* Atom

fn parse_unary_op(&mut self) -> Option<UnaryOp>

UnaryOp = ‘&’ ‘mut’ | ‘-’ | ‘!’ | ‘*’

fn parse_atom(&mut self, allow_constructors: bool) -> Option<Expression>

Atom = Quark AtomRhs*

fn parse_atom_rhs( &mut self, atom: Expression, start_location: Location, ) -> (Expression, bool)

AtomRhs = CallExpression | MemberAccessOrMethodCallExpression | IndexExpression

pub(super) fn parse_member_accesses_or_method_calls_after_expression( &mut self, atom: Expression, start_location: Location, ) -> Expression

fn parse_call( &mut self, atom: Expression, start_location: Location, ) -> (Expression, bool)

CallExpression = Quark CallArguments

fn parse_member_access_or_method_call( &mut self, atom: Expression, start_location: Location, ) -> (Expression, bool)

MemberAccessOrMethodCallExpression = MemberAccessExpression | MethodCallExpression

MemberAccessExpression = Quark ‘.’ identifier

MethodCallExpression = Quark ‘.’ identifier CallArguments

fn parse_member_access_field_name(&mut self) -> Option<Ident>

fn parse_cast( &mut self, atom: Expression, start_location: Location, ) -> (Expression, bool)

CastExpression = UnaryExpression ‘as’ Type

fn parse_index( &mut self, atom: Expression, start_location: Location, ) -> (Expression, bool)

IndexExpression = Quark ‘[’ Expression ‘]’

fn parse_generics_after_member_access_field_name( &mut self, ) -> Option<Vec<UnresolvedType>>

fn parse_quark(&mut self, allow_constructors: bool) -> Option<Expression>

Quark = Literal | ParenthesesExpression | UnsafeExpression | PathExpression | IfExpression | Lambda | ComptimeExpression | UnquoteExpression | TypePathExpression | NamelessTypePathExpression | AsTraitPath | ResolvedExpression | InternedExpression | InternedStatementExpression

fn parse_quark_kind( &mut self, allow_constructors: bool, ) -> Option<ExpressionKind>

fn parse_nameless_type_path_or_as_trait_path_type_expression( &mut self, ) -> Option<ExpressionKind>

NamelessTypePathExpression = ‘<’ Type ‘>’ ‘::’ identifier ( ‘::’ GenericTypeArgs )?

fn parse_resolved_expr(&mut self) -> Option<ExpressionKind>

ResolvedExpression = unquote_marker

fn parse_interned_expr(&mut self) -> Option<ExpressionKind>

InternedExpression = interned_expr

fn parse_interned_statement_expr(&mut self) -> Option<ExpressionKind>

InternedStatementExpression = interned_statement

fn parse_unsafe_expr(&mut self) -> Option<ExpressionKind>

UnsafeExpression = ‘unsafe’ Block

fn parse_path_expr( &mut self, allow_constructors: bool, ) -> Option<ExpressionKind>

PathExpression = VariableExpression | ConstructorExpression

VariableExpression = Path

fn parse_constructor(&mut self, typ: UnresolvedType) -> ExpressionKind

ConstructorExpression = Type ‘{’ ConstructorFields? ‘}’

ConstructorFields = ConstructorField ( ‘,’ ConstructorField )* ‘,’?

ConstructorField = identifier ( ‘:’ Expression )?

fn parse_constructor_field(&mut self) -> Option<(Ident, Expression)>

pub(super) fn parse_if_expr(&mut self) -> Option<ExpressionKind>

IfExpression = ‘if’ ExpressionExceptConstructor Block ( ‘else’ ( Block | IfExpression ) )?

pub(super) fn parse_match_expr(&mut self) -> Option<ExpressionKind>

MatchExpression = ‘match’ ExpressionExceptConstructor ‘{’ MatchRule* ‘}’

fn parse_match_rule(&mut self) -> Option<(Expression, Expression)>

MatchRule = Expression ‘=>’ (Block ‘,’?) | (Expression ‘,’)

fn parse_comptime_expr(&mut self) -> Option<ExpressionKind>

ComptimeExpression = ‘comptime’ Block

fn parse_unquote_expr(&mut self) -> Option<ExpressionKind>

UnquoteExpression = ‘$’ identifier | ‘$’ ‘(’ Expression ‘)’

fn parse_type_path_expr(&mut self) -> Option<ExpressionKind>

TypePathExpression = PrimitiveType ‘::’ identifier ( ‘::’ GenericTypeArgs )?

fn parse_type_path_expr_for_type(&mut self, typ: UnresolvedType) -> TypePath

fn parse_literal(&mut self) -> Option<ExpressionKind>

Literal = bool | int | str | rawstr | fmtstr | QuoteExpression | ArrayExpression | VectorExpression | BlockExpression | ConstrainExpression

QuoteExpression = ‘quote’ ‘{’ token* ‘}’

ArrayExpression = ArrayLiteral

BlockExpression = Block

fn parse_array_literal(&mut self) -> Option<ArrayLiteralOrError>

ArrayLiteral = StandardArrayLiteral | RepeatedArrayLiteral

StandardArrayLiteral = ‘[’ ArrayElements? ‘]’

ArrayElements = Expression ( ‘,’ Expression )? ‘,’?

RepeatedArrayLiteral = ‘[’ Expression ‘;’ TypeExpression ‘]’

fn parse_vector_literal(&mut self) -> Option<ArrayLiteralOrError>

VectorExpression = ‘&’ ArrayLiteral

fn parse_parentheses_expression(&mut self) -> Option<ExpressionKind>

ParenthesesExpression = UnitLiteral | ParenthesizedExpression | TupleExpression

UnitLiteral = ‘(’ ‘)’

ParenthesizedExpression = ‘(’ Expression ‘)’

TupleExpression = ‘(’ Expression ( ‘,’ Expression )+ ‘,’? ‘)’

pub(super) fn parse_expression_in_list(&mut self) -> Option<Expression>

pub(super) fn parse_constrain_expression( &mut self, ) -> Option<ConstrainExpression>

ConstrainExpression = ‘constrain’ Expression | ‘assert’ Arguments | ‘assert_eq’ Arguments

fn parse_constrain_kind(&mut self) -> Option<ConstrainKind>

pub(super) fn parse_block(&mut self) -> Option<BlockExpression>

Block = ‘{’ Statement* ‘}’

fn parse_statement_in_block( &mut self, ) -> Option<(Statement, (Option<Token>, Location))>

fn check_statements_require_semicolon( &mut self, statements: Vec<(Statement, (Option<Token>, Location))>, ) -> Vec<Statement>

pub(super) fn push_expected_expression(&mut self)

impl Parser<'_>

pub(crate) fn parse_function( &mut self, attributes: Vec<(Attribute, Location)>, visibility: ItemVisibility, is_comptime: bool, is_unconstrained: bool, allow_self: bool, ) -> NoirFunction

Function = ‘fn’ identifier Generics FunctionParameters ( ‘->’ Visibility Type )? WhereClause ( Block | ‘;’ )

pub(crate) fn parse_function_definition( &mut self, attributes: Vec<(Attribute, Location)>, visibility: ItemVisibility, is_comptime: bool, is_unconstrained: bool, allow_self: bool, ) -> FunctionDefinition

pub(super) fn parse_function_definition_with_optional_body( &mut self, allow_optional_body: bool, allow_self: bool, ) -> FunctionDefinitionWithOptionalBody

fn parse_function_parameters(&mut self, allow_self: bool) -> Option<Vec<Param>>

FunctionParameters = ‘(’ FunctionParametersList? ‘)’

FunctionParametersList = FunctionParameter ( ‘,’ FunctionParameter )* ‘,’?

FunctionParameter = Visibility PatternOrSelf ‘:’ Type

fn parse_function_parameter(&mut self, allow_self: bool) -> Option<Param>

fn pattern_param(&mut self, pattern: Pattern, start_location: Location) -> Param

fn self_pattern_param(&mut self, self_pattern: SelfPattern) -> Param

fn parse_visibility(&mut self) -> Visibility

Visibility = ‘pub’ | ‘return_data’ | ‘call_data’ ‘(’ int ‘)’ | nothing

fn validate_attributes( &mut self, attributes: Vec<(Attribute, Location)>, ) -> Attributes

impl Parser<'_>

pub(super) fn parse_generics_disallowing_trait_bounds( &mut self, ) -> UnresolvedGenerics

pub(super) fn parse_generics_allowing_trait_bounds( &mut self, ) -> UnresolvedGenerics

fn parse_generics(&mut self, allow_trait_bounds: bool) -> UnresolvedGenerics

Generics = ( ‘<’ GenericsList? ‘>’ )?

GenericsList = Generic ( ‘,’ Generic )* ‘,’?

fn parse_generic_in_list( &mut self, allow_trait_bounds: bool, ) -> Option<UnresolvedGeneric>

fn parse_generic( &mut self, allow_trait_bounds: bool, ) -> Option<UnresolvedGeneric>

Generic = VariableGeneric | NumericGeneric | ResolvedGeneric

fn parse_variable_generic( &mut self, allow_trait_bounds: bool, ) -> Option<UnresolvedGeneric>

VariableGeneric = identifier ( ‘:’ TraitBounds ) ?

fn parse_ident_or_quoted(&mut self) -> Option<IdentOrQuotedType>

fn parse_numeric_generic(&mut self) -> Option<UnresolvedGeneric>

NumericGeneric = ‘let’ identifier ‘:’ Type

pub(super) fn parse_generic_type_args(&mut self) -> GenericTypeArgs

GenericTypeArgs = ( ‘<’ GenericTypeArgsList? ‘>’ )

GenericTypeArgsList = GenericTypeArg ( ‘,’ GenericTypeArg )* ‘,’?

GenericTypeArg = NamedTypeArg | OrderedTypeArg

NamedTypeArg = identifier ‘=’ Type

OrderedTypeArg = TypeOrTypeExpression

fn parse_generic_type_arg(&mut self) -> Option<GenericTypeArg>

impl Parser<'_>

pub(crate) fn parse_global( &mut self, attributes: Vec<(Attribute, Location)>, comptime: bool, mutable: bool, ) -> LetStatement

Global = ‘global’ identifier OptionalTypeAnnotation ‘=’ Expression ‘;’

impl Parser<'_>

pub(crate) fn parse_impl(&mut self) -> Impl

Impl = TypeImpl | TraitImpl

fn parse_type_impl( &mut self, object_type: UnresolvedType, type_location: Location, generics: Vec<UnresolvedGeneric>, ) -> TypeImpl

TypeImpl = ‘impl’ Generics Type TypeImplBody

fn parse_type_impl_body(&mut self) -> Vec<(Documented<NoirFunction>, Location)>

TypeImplBody = ‘{’ TypeImplItem* ‘}’

TypeImplItem = OuterDocComments Attributes Modifiers Function

fn parse_type_impl_method( &mut self, ) -> Option<(Documented<NoirFunction>, Location)>

fn parse_trait_impl( &mut self, impl_generics: Vec<UnresolvedGeneric>, trait: UnresolvedType, ) -> NoirTraitImpl

TraitImpl = ‘impl’ Generics Type ‘for’ Type TraitImplBody

fn parse_trait_impl_body(&mut self) -> Vec<Documented<TraitImplItem>>

TraitImplBody = ‘{’ TraitImplItem* ‘}’

fn parse_trait_impl_item(&mut self) -> Option<Documented<TraitImplItem>>

fn parse_trait_impl_item_kind(&mut self) -> Option<TraitImplItemKind>

TraitImplItem = TraitImplType | TraitImplConstant | TraitImplFunction

fn parse_trait_impl_type(&mut self) -> Option<TraitImplItemKind>

TraitImplType = ‘type’ identifier ( ‘:’ Type )? ‘;’

fn parse_trait_impl_constant(&mut self) -> Option<TraitImplItemKind>

TraitImplConstant = ‘let’ identifier OptionalTypeAnnotation ‘;’

fn parse_trait_impl_function(&mut self) -> Option<TraitImplItemKind>

TraitImplFunction = Attributes Modifiers Function

impl Parser<'_>

pub(super) fn parse_equal_or_not_equal( &mut self, allow_constructors: bool, ) -> Option<Expression>

EqualOrNotEqualExpression = OrExpression ( ( ‘==’ | ‘!=’ ) OrExpression )*

pub(super) fn parse_or( &mut self, allow_constructors: bool, ) -> Option<Expression>

OrExpression = AndExpression ( ‘|’ AndExpression )*

pub(super) fn parse_and( &mut self, allow_constructors: bool, ) -> Option<Expression>

AndExpression = XorExpression ( ‘&’ XorExpression )*

pub(super) fn parse_xor( &mut self, allow_constructors: bool, ) -> Option<Expression>

XorExpression = LessOrGreaterExpression ( ‘^’ LessOrGreaterExpression )*

pub(super) fn parse_less_or_greater( &mut self, allow_constructors: bool, ) -> Option<Expression>

LessOrGreaterExpression = ShiftExpression ( ( ‘<’ | ‘<=’ | ‘>’ | ‘>=’ ) ShiftExpression )*

pub(super) fn parse_shift( &mut self, allow_constructors: bool, ) -> Option<Expression>

ShiftExpression = AddOrSubtractExpression ( ( ‘<<’ | ‘>’ ‘>’ ) AddOrSubtractExpression )*

pub(super) fn parse_add_or_subtract( &mut self, allow_constructors: bool, ) -> Option<Expression>

AddOrSubtractExpression = MultiplyOrDivideOrModuloExpression ( ( ‘+’ | ‘-’ ) MultiplyOrDivideOrModuloExpression )*

pub(super) fn parse_multiply_or_divide_or_modulo( &mut self, allow_constructors: bool, ) -> Option<Expression>

MultiplyOrDivideOrModuloExpression = Term ( ( ‘’ | ‘/’ | ‘%’ ) Term )

fn new_infix_expression( &self, lhs: Expression, operator: Located<BinaryOpKind>, rhs: Expression, start_location: Location, ) -> Expression

impl<'a> Parser<'a>

pub(crate) fn parse_top_level_items(&mut self) -> Vec<Item>

pub(crate) fn parse_module_items(&mut self, nested: bool) -> Vec<Item>

fn parse_module_item_in_list(&mut self, nested: bool) -> Vec<Item>

pub(super) fn parse_item_in_list<T, F>( &mut self, label: ParsingRuleLabel, f: F, ) -> Option<T>
where F: FnMut(&mut Parser<'a>) -> Option<T>,

Parses an item inside an impl or trait, with good recovery:

If we run into EOF, we error that we expect a ‘}’

If we can’t parse an item and we don’t end up in ‘}’, error but try with the next token

fn parse_item(&mut self) -> Vec<Item>

Item = ( Attribute | OuterDocComments )* ItemKind

fn parse_item_kind( &mut self, attributes: Vec<(Attribute, Location)>, ) -> Vec<ItemKind>

This method returns one ‘ItemKind’ in the majority of cases. The current exception is when parsing a trait alias, which returns both the trait and the impl.

ItemKind = InnerAttribute | Attributes Modifiers ( Use | ModOrContract | Struct | Enum | Impl | Trait | Global | TypeAlias | Function )

fn eat_mod_or_contract(&mut self) -> Option<bool>

impl Parser<'_>

pub fn parse_item_visibility(&mut self) -> ItemVisibility

ItemVisibility = ‘pub’ // ItemVisibility::Public | ‘pub’ ‘(’ ‘crate’ ‘)’ // ItemVisibility::PublicCrate | nothing // ItemVisibility::Private

impl Parser<'_>

pub(super) fn parse_lambda(&mut self) -> Option<ExpressionKind>

Lambda = ( ‘unconstrained’ )? ‘|’ LambdaParameters? ‘|’ ( ‘->’ Type )? Expression

LambdaParameters = LambdaParameter ( ‘,’ LambdaParameter )? ‘,’?

LambdaParameter = Pattern OptionalTypeAnnotation

fn parse_lambda_parameters(&mut self) -> Vec<(Pattern, Option<UnresolvedType>)>

fn parse_lambda_parameter( &mut self, ) -> Option<(Pattern, Option<UnresolvedType>)>

impl Parser<'_>

pub(crate) fn parse_modifiers(&mut self, allow_mutable: bool) -> Modifiers

Modifiers = ItemVisibility ‘unconstrained’? ‘comptime’? ‘mut’?

NOTE: we also allow unconstrained before the visibility for backwards compatibility. The formatter will put it after the visibility.

impl Parser<'_>

pub(super) fn parse_mod_or_contract( &mut self, ident: Ident, attributes: Vec<(Attribute, Location)>, is_contract: bool, visibility: ItemVisibility, ) -> ItemKind

ModOrContract = ( ‘mod’ | ‘contract’ ) identifier ( ‘{’ Module ‘}’ | ‘;’ )

impl<'a> Parser<'a>

pub(super) fn parse_many<T, F>( &mut self, items: &'static str, separated_by: SeparatedBy, f: F, ) -> Vec<T>
where F: FnMut(&mut Parser<'a>) -> Option<T>,

Parses a list of items separated by a token, optionally ending when another token is found. The given function f must parse one item (eventually parsing many, if separators are found). If no item is parsed, f must report an error and return None.

pub(super) fn parse_many_to_many<T, F>( &mut self, items: &'static str, separated_by: SeparatedBy, f: F, ) -> Vec<T>
where F: FnMut(&mut Parser<'a>) -> Vec<T>,

parse_many, where the given function f may return multiple results

pub(super) fn parse_many_return_trailing_separator_if_any<T, F>( &mut self, items: &'static str, separated_by: SeparatedBy, f: F, ) -> (Vec<T>, bool)
where F: FnMut(&mut Parser<'a>) -> Option<T>,

Same as parse_many, but returns a bool indicating whether a trailing separator was found.

fn parse_many_to_many_return_trailing_separator_if_any<T, F>( &mut self, items: &'static str, separated_by: SeparatedBy, f: F, ) -> (Vec<T>, bool)
where F: FnMut(&mut Parser<'a>) -> Vec<T>,

Same as parse_many, but returns a bool indicating whether a trailing separator was found.

impl Parser<'_>

pub(crate) fn parse_path(&mut self) -> Option<Path>

Tries to parse a Path. Note that crate::, super::, etc., are not valid paths on their own.

Path = PathKind identifier Turbofish? ( ‘::’ identifier Turbofish? )*

Turbofish = ‘::’ PathGenerics

pub(crate) fn parse_path_no_turbofish_or_error(&mut self) -> Path

pub fn parse_path_no_turbofish(&mut self) -> Option<Path>

PathNoTurbofish = PathKind identifier ( ‘::’ identifier )*

pub(super) fn parse_path_impl( &mut self, allow_turbofish: bool, allow_trailing_double_colon: bool, ) -> Option<Path>

pub(super) fn parse_optional_path_after_kind( &mut self, kind: PathKind, allow_turbofish: bool, allow_trailing_double_colon: bool, start_location: Location, ) -> Option<Path>

pub(super) fn parse_path_after_kind( &mut self, kind: PathKind, allow_turbofish: bool, allow_trailing_double_colon: bool, start_location: Location, ) -> Path

Parses a path assuming the path’s kind (plain, crate::, super::, etc.) was already parsed. Note that this method always returns a Path, even if it ends up being just crate:: or an empty path.

pub(super) fn parse_path_generics( &mut self, on_named_arg_error: ParserErrorReason, ) -> Option<Vec<UnresolvedType>>

PathGenerics = GenericTypeArgs

pub(super) fn parse_path_kind(&mut self) -> PathKind

PathKind | ‘crate’ ‘::’ | ‘dep’ ‘::’ | ‘super’ ‘::’ | nothing

pub(super) fn parse_as_trait_path(&mut self) -> Option<AsTraitPath>

AsTraitPath = ‘<’ Type ‘as’ PathNoTurbofish GenericTypeArgs ‘>’ ‘::’ identifier

pub(super) fn parse_as_trait_path_for_type_after_as_keyword( &mut self, typ: UnresolvedType, ) -> AsTraitPath

impl Parser<'_>

pub(crate) fn parse_pattern_or_error(&mut self) -> Pattern

pub(crate) fn parse_pattern(&mut self) -> Option<Pattern>

Pattern = ‘mut’ PatternNoMut

pub(crate) fn parse_pattern_or_self(&mut self) -> Option<PatternOrSelf>

PatternOrSelf = Pattern | SelfPattern

fn next_is_colon(&self) -> bool

pub(crate) fn parse_pattern_after_modifiers( &mut self, mutable: bool, start_location: Location, ) -> Option<Pattern>

fn parse_pattern_no_mut(&mut self) -> Option<Pattern>

PatternNoMut = InternedPattern | ParenthesizedPattern | TuplePattern | StructPattern | IdentifierPattern

IdentifierPattern = identifier

fn parse_interned_pattern(&mut self) -> Option<Pattern>

InternedPattern = interned_pattern

fn parse_parenthesized_or_tuple_pattern(&mut self) -> Option<Pattern>

ParenthesizedPattern = ‘(’ Pattern ‘)’ TuplePattern = ‘(’ PatternList? ‘)’

PatternList = Pattern ( ‘,’ Pattern )* ‘,’?

fn parse_tuple_pattern_element(&mut self) -> Option<Pattern>

fn parse_struct_pattern( &mut self, path: Path, start_location: Location, ) -> Pattern

StructPattern = Path ‘{’ StructPatternFields? ‘}’

StructPatternFields = StructPatternField ( ‘,’ StructPatternField )? ‘,’?

StructPatternField = identifier ( ‘:’ Pattern )?

fn parse_struct_pattern_field(&mut self) -> Option<(Ident, Pattern)>

impl Parser<'_>

pub(crate) fn parse_statement_or_error(&mut self) -> Statement

pub(crate) fn parse_statement( &mut self, ) -> Option<(Statement, (Option<Token>, Location))>

Statement = Attributes StatementKind ‘;’?

fn parse_statement_kind( &mut self, attributes: Vec<(Attribute, Location)>, ) -> Option<StatementKind>

StatementKind = BreakStatement | ContinueStatement | ReturnStatement | LetStatement | ComptimeStatement | ForStatement | LoopStatement | WhileStatement | IfStatement | BlockStatement | AssignStatement | ExpressionStatement

BreakStatement = ‘break’

ContinueStatement = ‘continue’

ReturnStatement = ‘return’ Expression?

IfStatement = IfExpression

BlockStatement = Block

AssignStatement = Expression ‘=’ Expression

ExpressionStatement = Expression

fn parse_block_like(&mut self) -> Option<ExpressionKind>

Parses an expression that looks like a block (ends with ‘}’): { ... }, if { ... } and match { ... }.

fn next_is_op_assign(&mut self) -> Option<BinaryOp>

fn parse_for(&mut self) -> Option<ForLoopStatement>

ForStatement = ‘for’ identifier ‘in’ ForRange Block

fn parse_loop(&mut self) -> Option<LoopStatement>

LoopStatement = ‘loop’ Block

fn parse_while(&mut self) -> Option<WhileStatement>

WhileStatement = ‘while’ ExpressionExceptConstructor Block

fn parse_for_range(&mut self) -> ForRange

ForRange = ExpressionExceptConstructor | ExpressionExceptConstructor ‘..’ ExpressionExceptConstructor

fn empty_for_loop( &mut self, identifier: Ident, start_location: Location, ) -> ForLoopStatement

fn parse_comptime_statement( &mut self, attributes: Vec<(Attribute, Location)>, ) -> Option<StatementKind>

ComptimeStatement = ComptimeBlock | ComptimeLet | ComptimeFor

ComptimeBlock = ‘comptime’ Block

ComptimeLet = ‘comptime’ LetStatement

ComptimeFor = ‘comptime’ ForStatement

fn parse_comptime_statement_kind( &mut self, attributes: Vec<(Attribute, Location)>, ) -> Option<StatementKind>

fn parse_let_statement( &mut self, attributes: Vec<(Attribute, Location)>, ) -> Option<LetStatement>

LetStatement = ‘let’ pattern OptionalTypeAnnotation ‘=’ Expression

impl Parser<'_>

pub(crate) fn parse_statement_or_expression_or_lvalue( &mut self, ) -> StatementOrExpressionOrLValue

Parses either a statement, an expression or an LValue. Returns StatementKind::Error if none can be parsed, recording an error if so.

This method is only used in Quoted::as_expr.

impl Parser<'_>

pub(crate) fn parse_struct( &mut self, attributes: Vec<(Attribute, Location)>, visibility: ItemVisibility, start_location: Location, ) -> NoirStruct

Struct = ‘struct’ identifier Generics ‘{’ StructField* ‘}’

StructField = OuterDocComments identifier ‘:’ Type

fn parse_struct_field(&mut self) -> Option<Documented<StructField>>

fn empty_struct( &self, name: Ident, attributes: Vec<SecondaryAttribute>, visibility: ItemVisibility, generics: UnresolvedGenerics, start_location: Location, ) -> NoirStruct

impl Parser<'_>

pub(crate) fn parse_trait( &mut self, attributes: Vec<(Attribute, Location)>, visibility: ItemVisibility, start_location: Location, ) -> (NoirTrait, Option<NoirTraitImpl>)

Trait = ‘trait’ identifier Generics ( ‘:’ TraitBounds )? WhereClause TraitBody | ‘trait’ identifier Generics ‘=’ TraitBounds WhereClause ‘;’

fn parse_trait_body(&mut self) -> Vec<Documented<TraitItem>>

TraitBody = ‘{’ ( OuterDocComments TraitItem )* ‘}’

fn parse_trait_item_in_list(&mut self) -> Option<Documented<TraitItem>>

fn parse_trait_item(&mut self) -> Option<TraitItem>

TraitItem = TraitType | TraitConstant | TraitFunction

fn parse_trait_type(&mut self) -> Option<TraitItem>

TraitType = ‘type’ identifier ( ‘:’ TraitBounds ) ‘;’

fn parse_trait_constant(&mut self) -> Option<TraitItem>

TraitConstant = ‘let’ identifier ‘:’ Type ( ‘=’ Expression )? ‘;’

fn parse_trait_function(&mut self) -> Option<TraitItem>

TraitFunction = Modifiers Function

impl Parser<'_>

pub(crate) fn parse_type_alias( &mut self, visibility: ItemVisibility, start_location: Location, ) -> TypeAlias

TypeAlias = ‘type’ identifier Generics ‘=’ Type ‘;’

impl Parser<'_>

pub(crate) fn parse_type_expression( &mut self, ) -> Result<UnresolvedTypeExpression, ParserError>

TypeExpression= AddOrSubtractTypeExpression

fn parse_add_or_subtract_type_expression( &mut self, ) -> Option<UnresolvedTypeExpression>

AddOrSubtractTypeExpression = MultiplyOrDivideOrModuloTypeExpression ( ( ‘+’ | ‘-’ ) MultiplyOrDivideOrModuloTypeExpression )*

fn parse_add_or_subtract_type_expression_after_lhs( &mut self, lhs: UnresolvedTypeExpression, start_location: Location, ) -> UnresolvedTypeExpression

fn parse_multiply_or_divide_or_modulo_type_expression( &mut self, ) -> Option<UnresolvedTypeExpression>

MultiplyOrDivideOrModuloTypeExpression = TermTypeExpression ( ( ‘’ | ‘/’ | ‘%’ ) TermTypeExpression )

fn parse_multiply_or_divide_or_modulo_type_expression_after_lhs( &mut self, lhs: UnresolvedTypeExpression, start_location: Location, ) -> UnresolvedTypeExpression

fn parse_term_type_expression(&mut self) -> Option<UnresolvedTypeExpression>

TermTypeExpression = ’- TermTypeExpression | AtomTypeExpression

fn parse_atom_type_expression(&mut self) -> Option<UnresolvedTypeExpression>

AtomTypeExpression = ConstantTypeExpression | VariableTypeExpression | AsTraitPathTypeExpression | ParenthesizedTypeExpression

fn parse_constant_type_expression(&mut self) -> Option<UnresolvedTypeExpression>

ConstantTypeExpression = int

fn parse_variable_type_expression(&mut self) -> Option<UnresolvedTypeExpression>

VariableTypeExpression = Path

fn parse_parenthesized_type_expression( &mut self, ) -> Option<UnresolvedTypeExpression>

ParenthesizedTypeExpression = ‘(’ TypeExpression ‘)’

pub(crate) fn parse_type_or_type_expression(&mut self) -> Option<UnresolvedType>

TypeOrTypeExpression = Type | TypeExpression

fn parse_add_or_subtract_type_or_type_expression( &mut self, ) -> Option<UnresolvedType>

fn parse_multiply_or_divide_or_modulo_type_or_type_expression( &mut self, ) -> Option<UnresolvedType>

fn parse_term_type_or_type_expression(&mut self) -> Option<UnresolvedType>

fn parse_atom_type_or_type_expression(&mut self) -> Option<UnresolvedType>

fn parse_parenthesized_type_or_type_expression( &mut self, ) -> Option<UnresolvedType>

fn expected_type_expression_after_this( &mut self, ) -> Result<UnresolvedTypeExpression, ParserError>

impl Parser<'_>

pub(crate) fn parse_type_or_error(&mut self) -> UnresolvedType

pub(crate) fn parse_type_or_error_without_generics(&mut self) -> UnresolvedType

pub(crate) fn parse_type_or_error_impl( &mut self, allow_generics: bool, ) -> UnresolvedType

pub(crate) fn parse_type_or_error_with_recovery( &mut self, stop_tokens: &[Token], ) -> UnresolvedType

Tries to parse a type. If the current token doesn’t denote a type and it’s not one of stop_tokens, try to parse a type starting from the next token (and so on).

pub(crate) fn parse_type(&mut self) -> Option<UnresolvedType>

pub(crate) fn parse_type_allowing_generics( &mut self, allow_generics: bool, ) -> Option<UnresolvedType>

fn parse_unresolved_type_data( &mut self, allow_generics: bool, ) -> Option<UnresolvedTypeData>

pub(super) fn parse_primitive_type(&mut self) -> Option<UnresolvedTypeData>

fn parse_function_type(&mut self) -> Option<UnresolvedTypeData>

fn parse_parameter(&mut self) -> Option<UnresolvedType>

fn parse_trait_as_type(&mut self) -> Option<UnresolvedTypeData>

fn parse_as_trait_path_type(&mut self) -> Option<UnresolvedTypeData>

pub(super) fn parse_resolved_type(&mut self) -> Option<UnresolvedTypeData>

pub(super) fn parse_interned_type(&mut self) -> Option<UnresolvedTypeData>

fn parse_reference_type(&mut self) -> Option<UnresolvedTypeData>

fn parse_array_or_vector_type(&mut self) -> Option<UnresolvedTypeData>

fn parse_parentheses_type(&mut self) -> Option<UnresolvedTypeData>

pub(super) fn parse_type_in_list(&mut self) -> Option<UnresolvedType>

pub(super) fn parse_optional_type_annotation( &mut self, ) -> Option<UnresolvedType>

OptionalTypeAnnotation = ( ‘:’ Type )?

fn error_type_at_previous_token_end(&mut self) -> UnresolvedType

impl Parser<'_>

pub(super) fn parse_use_tree(&mut self) -> UseTree

Use = ‘use’ PathKind PathNoTurbofish UseTree

UseTree = PathNoTurbofish ( ‘::’ ‘{’ UseTreeList? ‘}’ )?

UseTreeList = UseTree (‘,’ UseTree)* ‘,’?

pub(super) fn parse_use_tree_without_kind( &mut self, start_location: Location, kind: PathKind, nested: bool, ) -> UseTree

fn parse_use_tree_in_list(&mut self) -> Option<UseTree>

pub(super) fn parse_path_use_tree_end( &mut self, prefix: Path, nested: bool, start_location: Location, ) -> UseTree

impl Parser<'_>

pub(super) fn parse_where_clause(&mut self) -> Vec<UnresolvedTraitConstraint>

WhereClause = ‘where’ WhereClauseItems?

WhereClauseItems = WhereClauseItem ( ‘,’ WhereClauseItem )* ‘,’?

WhereClauseItem = Type ‘:’ TraitBounds

fn parse_single_where_clause( &mut self, ) -> Option<(UnresolvedType, Vec<TraitBound>)>

pub(super) fn parse_trait_bounds(&mut self) -> Vec<TraitBound>

TraitBounds = TraitBound ( ‘+’ TraitBound )? ‘+’?

fn parse_trait_bound_in_list(&mut self) -> Option<TraitBound>

pub(crate) fn parse_trait_bound_or_error(&mut self) -> TraitBound

pub(crate) fn parse_trait_bound(&mut self) -> Option<TraitBound>

TraitBound = PathNoTurbofish GenericTypeArgs

impl<'a> Parser<'a>

pub fn for_lexer(lexer: Lexer<'a>) -> Self

pub fn for_tokens(tokens: Tokens) -> Self

pub fn for_str(str: &'a str, file_id: FileId) -> Self

pub fn for_str_with_dummy_file(str: &'a str) -> Self

fn new(tokens: TokenStream<'a>) -> Self

pub(crate) fn parse_program(&mut self) -> ParsedModule

Program = Module

pub(crate) fn parse_module(&mut self, nested: bool) -> ParsedModule

Module = InnerDocComments Item*

pub fn parse_result<T, F>( self, parsing_function: F, ) -> Result<(T, Vec<ParserError>), Vec<ParserError>>
where F: FnOnce(&mut Parser<'a>) -> T,

Invokes parsing_function (parsing_function must be some parse_* method of the parser) and returns the result (and any warnings) if the parser has no errors, and if the parser consumed all tokens. Otherwise returns the list of errors.

fn bump(&mut self) -> LocatedToken

Bumps this parser by one token. Returns the token that was previously the “current” token.

fn read_two_first_tokens(&mut self)

fn read_token_internal(&mut self) -> (LocatedToken, String)

fn eat_kind(&mut self, kind: TokenKind) -> Option<LocatedToken>

fn eat_keyword(&mut self, keyword: Keyword) -> bool

fn eat_ident(&mut self) -> Option<Ident>

fn eat_self(&mut self) -> bool

fn eat_int(&mut self) -> Option<(FieldElement, Option<IntegerTypeSuffix>)>

fn eat_bool(&mut self) -> Option<bool>

fn eat_str(&mut self) -> Option<String>

fn eat_raw_str(&mut self) -> Option<(String, u8)>

fn eat_fmt_str(&mut self) -> Option<(Vec<FmtStrFragment>, u32)>

fn eat_quote(&mut self) -> Option<Tokens>

fn eat_unquote_marker(&mut self) -> Option<ExprId>

fn eat_attribute_start(&mut self) -> Option<bool>

fn eat_inner_attribute_start(&mut self) -> Option<bool>

fn eat_comma(&mut self) -> bool

fn eat_semicolon(&mut self) -> bool

fn eat_semicolons(&mut self) -> bool

fn eat_semicolon_or_error(&mut self)

fn eat_colon(&mut self) -> bool

fn eat_double_colon(&mut self) -> bool

fn eat_left_paren(&mut self) -> bool

fn eat_right_paren(&mut self) -> bool

fn eat_left_brace(&mut self) -> bool

fn eat_left_bracket(&mut self) -> bool

fn eat_right_bracket(&mut self) -> bool

fn eat_less(&mut self) -> bool

fn eat_assign(&mut self) -> bool

fn eat_dot(&mut self) -> bool

fn eat_pipe(&mut self) -> bool

fn eat(&mut self, token: Token) -> bool

fn eat_keyword_or_error(&mut self, keyword: Keyword)

fn eat_or_error(&mut self, token: Token)

fn at(&self, token: Token) -> bool

fn at_keyword(&self, keyword: Keyword) -> bool

fn at_whitespace(&self) -> bool

fn set_lexer_skip_whitespaces_flag(&mut self, flag: bool)

fn next_is(&self, token: Token) -> bool

fn at_eof(&self) -> bool

fn location_since(&self, start_location: Location) -> Location

fn location_at_previous_token_end(&self) -> Location

fn unknown_ident_at_previous_token_end(&self) -> Ident

fn expected_identifier(&mut self)

fn expected_string(&mut self)

fn expected_token(&mut self, token: Token)

fn expected_one_of_tokens(&mut self, tokens: &[Token])

fn expected_label(&mut self, label: ParsingRuleLabel)

fn expected_token_separating_items( &mut self, token: Token, items: &'static str, location: Location, )

fn expected_mut_after_ampersand(&mut self)

fn modifiers_not_followed_by_an_item(&mut self, modifiers: Modifiers)

fn visibility_not_followed_by_an_item(&mut self, modifiers: Modifiers)

fn unconstrained_not_followed_by_an_item(&mut self, modifiers: Modifiers)

fn comptime_not_followed_by_an_item(&mut self, modifiers: Modifiers)

fn comptime_mutable_and_unconstrained_not_applicable( &mut self, modifiers: Modifiers, )

fn mutable_not_applicable(&mut self, modifiers: Modifiers)

fn comptime_not_applicable(&mut self, modifiers: Modifiers)

fn unconstrained_not_applicable(&mut self, modifiers: Modifiers)

fn push_error(&mut self, reason: ParserErrorReason, location: Location)