noirc_frontend

Module shared

The shared module contains simple types which are using in multiple of Noir’s IRs.

This is done to avoid each IR from needing to have its own definition of elementary types while avoiding one IR being embedded within another.

Modules

foreign_calls 🔒

signedness 🔒

visibility 🔒

Enums

ForeignCall

This enumeration represents the Brillig foreign calls that appear in the Noir stdlib, and are natively supported by nargo.

Signedness

Visibility

Represents whether a parameter or return value is public or known only to the prover.