noirc_frontend::shared

Enum Builtin

pub enum Builtin {

    ApplyRangeConstraint,
    ArrayAsStrUnchecked,
    ArrayLen,
    ArrayRefcount,
    AsVector,
    AsWitness,
    AssertConstant,
    BlackBoxHint,
    DerivePedersenGenerators,
    FieldLessThan,
    IsUnconstrained,
    StaticAssert,
    StrAsBytes,
    ToBeBits,
    ToBeRadix,
    ToLeBits,
    ToLeRadix,
    VectorInsert,
    VectorPopBack,
    VectorPopFront,
    VectorPushBack,
    VectorPushFront,
    VectorRefcount,
    VectorRemove,
    CheckedTransmute,
    ModulusBeBits,
    ModulusBeBytes,
    ModulusLeBits,
    ModulusLeBytes,
    ModulusNumBits,
    Zeroed,
    CtstringAppend,
    CtstringEq,
    CtstringHash,
    ExprAsArray,
    ExprAsAssert,
    ExprAsAssertEq,
    ExprAsAssign,
    ExprAsBinaryOp,
    ExprAsBlock,
    ExprAsBool,
    ExprAsCast,
    ExprAsComptime,
    ExprAsConstructor,
    ExprAsFor,
    ExprAsForRange,
    ExprAsFunctionCall,
    ExprAsIf,
    ExprAsIndex,
    ExprAsInteger,
    ExprAsLambda,
    ExprAsLet,
    ExprAsMemberAccess,
    ExprAsMethodCall,
    ExprAsRepeatedElementArray,
    ExprAsRepeatedElementVector,
    ExprAsTuple,
    ExprAsUnaryOp,
    ExprAsUnsafe,
    ExprAsVector,
    ExprHasSemicolon,
    ExprIsBreak,
    ExprIsContinue,
    ExprResolve,
    FmtstrAsCtstring,
    FmtstrQuotedContents,
    FreshTypeVariable,
    FunctionDefAsTypedExpr,
    FunctionDefBody,
    FunctionDefDisable,
    FunctionDefEq,
    FunctionDefHasBuiltinAttribute,
    FunctionDefHasNamedAttribute,
    FunctionDefHash,
    FunctionDefIsUnconstrained,
    FunctionDefLocation,
    FunctionDefModule,
    FunctionDefName,
    FunctionDefNamedAttributeArgs,
    FunctionDefParameters,
    FunctionDefReturnType,
    FunctionDefVisibility,
    IssueError,
    IssueWarning,
    LocationEq,
    LocationHash,
    ModuleChildModules,
    ModuleEq,
    ModuleFunctions,
    ModuleHasBuiltinAttribute,
    ModuleHasNamedAttribute,
    ModuleHash,
    ModuleIsContract,
    ModuleLocation,
    ModuleName,
    ModuleNamedAttributeArgs,
    ModuleParent,
    ModuleStructs,
    QuotedAsExpr,
    QuotedAsModule,
    QuotedAsTraitConstraint,
    QuotedAsType,
    QuotedEq,
    QuotedHash,
    QuotedLocation,
    QuotedTokens,
    StrAsCtstring,
    TraitConstraintEq,
    TraitConstraintHash,
    TraitDefAsTraitConstraint,
    TraitDefEq,
    TraitDefHash,
    TraitDefLocation,
    TraitImplMethods,
    TraitImplTraitGenericArgs,
    TypeAsArray,
    TypeAsConstant,
    TypeAsDataType,
    TypeAsInteger,
    TypeAsMutableReference,
    TypeAsStr,
    TypeAsTuple,
    TypeAsVector,
    TypeDefAddAbi,
    TypeDefAsType,
    TypeDefAsTypeWithGenerics,
    TypeDefEq,
    TypeDefFields,
    TypeDefFieldsAsWritten,
    TypeDefGenerics,
    TypeDefHasBuiltinAttribute,
    TypeDefHasNamedAttribute,
    TypeDefHash,
    TypeDefLocation,
    TypeDefModule,
    TypeDefName,
    TypeDefNamedAttributeArgs,
    TypeEq,
    TypeGetTraitImpl,
    TypeHash,
    TypeImplements,
    TypeIsBool,
    TypeIsField,
    TypeIsUnit,
    TypeOf,
    TypedExprAsFunctionDefinition,
    TypedExprGetType,
    TypedExprLocation,
    Poseidon2ConfigStateSize,
    BlackBox(BlackBoxFunc),

}

A function the compiler implements, declared in the standard library with the #[builtin(name)] or #[foreign(name)] attribute.

The two attribute kinds share a single flat namespace: by the time a call is dispatched, builtin and foreign (“low level”) functions are treated identically, so this enum covers both. Oracle (#[oracle]) names are deliberately not here: they are an open, user-extensible namespace and stay strings.

Attribute parsing keeps the name as a free-form string (so malformed or unknown names still parse and error later, with a real diagnostic); the string is resolved to this enum where the function is actually consumed: the comptime interpreter, the monomorphizer, and (via Definition::Builtin/Definition::LowLevel) SSA generation.

Variants

ApplyRangeConstraint

ArrayAsStrUnchecked

ArrayLen

ArrayRefcount

AsVector

AsWitness

AssertConstant

BlackBoxHint

DerivePedersenGenerators

FieldLessThan

IsUnconstrained

StaticAssert

StrAsBytes

ToBeBits

ToBeRadix

ToLeBits

ToLeRadix

VectorInsert

VectorPopBack

VectorPopFront

VectorPushBack

VectorPushFront

VectorRefcount

VectorRemove

CheckedTransmute

ModulusBeBits

ModulusBeBytes

ModulusLeBits

ModulusLeBytes

ModulusNumBits

Zeroed

CtstringAppend

CtstringEq

CtstringHash

ExprAsArray

ExprAsAssert

ExprAsAssertEq

ExprAsAssign

ExprAsBinaryOp

ExprAsBlock

ExprAsBool

ExprAsCast

ExprAsComptime

ExprAsConstructor

ExprAsFor

ExprAsForRange

ExprAsFunctionCall

ExprAsIf

ExprAsIndex

ExprAsInteger

ExprAsLambda

ExprAsLet

ExprAsMemberAccess

ExprAsMethodCall

ExprAsRepeatedElementArray

ExprAsRepeatedElementVector

ExprAsTuple

ExprAsUnaryOp

ExprAsUnsafe

ExprAsVector

ExprHasSemicolon

ExprIsBreak

ExprIsContinue

ExprResolve

FmtstrAsCtstring

FmtstrQuotedContents

FreshTypeVariable

FunctionDefAsTypedExpr

FunctionDefBody

FunctionDefDisable

FunctionDefEq

FunctionDefHasBuiltinAttribute

FunctionDefHasNamedAttribute

FunctionDefHash

FunctionDefIsUnconstrained

FunctionDefLocation

FunctionDefModule

FunctionDefName

FunctionDefNamedAttributeArgs

FunctionDefParameters

FunctionDefReturnType

FunctionDefVisibility

IssueError

IssueWarning

LocationEq

LocationHash

ModuleChildModules

ModuleEq

ModuleFunctions

ModuleHasBuiltinAttribute

ModuleHasNamedAttribute

ModuleHash

ModuleIsContract

ModuleLocation

ModuleName

ModuleNamedAttributeArgs

ModuleParent

ModuleStructs

QuotedAsExpr

QuotedAsModule

QuotedAsTraitConstraint

QuotedAsType

QuotedEq

QuotedHash

QuotedLocation

QuotedTokens

StrAsCtstring

TraitConstraintEq

TraitConstraintHash

TraitDefAsTraitConstraint

TraitDefEq

TraitDefHash

TraitDefLocation

TraitImplMethods

TraitImplTraitGenericArgs

TypeAsArray

TypeAsConstant

TypeAsDataType

TypeAsInteger

TypeAsMutableReference

TypeAsStr

TypeAsTuple

TypeAsVector

TypeDefAddAbi

TypeDefAsType

TypeDefAsTypeWithGenerics

TypeDefEq

TypeDefFields

TypeDefFieldsAsWritten

TypeDefGenerics

TypeDefHasBuiltinAttribute

TypeDefHasNamedAttribute

TypeDefHash

TypeDefLocation

TypeDefModule

TypeDefName

TypeDefNamedAttributeArgs

TypeEq

TypeGetTraitImpl

TypeHash

TypeImplements

TypeIsBool

TypeIsField

TypeIsUnit

TypeOf

TypedExprAsFunctionDefinition

TypedExprGetType

TypedExprLocation

Poseidon2ConfigStateSize

BlackBox(BlackBoxFunc)

A foreign (#[foreign]) function implemented as an ACIR black box function (hashes, curve operations, signature verification). Excluded from the strum derives because of its payload; Self::lookup and Self::name handle it explicitly.

Implementations

impl Builtin

pub fn lookup(name: &str) -> Option<Self>

Resolve a #[builtin(name)] / #[foreign(name)] attribute name. Returns None for names the compiler does not implement.

pub fn name(&self) -> &str

The name this builtin is declared with, i.e. the inverse of Self::lookup.