noirc_frontend

Module test_utils

This crate represents utility methods which can be useful for testing in other crates which also desire to compile the frontend.

This module is split out from the tests module and has an additional test_utils feature as a module configured only for tests will not be accessible in other crates. A crate that needs to use the methods in this module should add the noirc_frontend crate as a dev dependency with the test_utils feature activated.

Modules

stdlib_src

These snippets can be used in conjunction with get_program_with_stdlib.

Structs

GetProgramOptions

Enums

Expect

Functions

get_monomorphized

Compile and monomorphize a program.

get_monomorphized_with_options

Compile and monomorphize a program.

get_monomorphized_with_stdlib

Helper to monomorphize code which needs some parts of the stdlib repeated for the test.

get_program

Compile a program.

get_program_with_options 🔒

Compile a program.

has_parser_error 🔒

remove_experimental_warnings 🔒