noirc_frontend::ast
Module statement

Structs

AsTraitPath

A special kind of path in the form <MyType as Trait>::ident. Note that this path must consist of exactly two segments.

AssignStatement

ForBounds

ForLoopStatement

Ident

ImportStatement

LetStatement

LoopStatement

ModuleDeclaration

Path

PathSegment

Statement

TypePath

A special kind of path in the form Type::ident::<turbofish> Unlike normal paths, the type here can be a primitive type or interned type.

UnsafeExpression

UseTree

WhileStatement

Enums

ForRange

LValue

Represents an Ast form that can be assigned to

PathKind

Pattern

StatementKind

Ast node for statements in noir. Statements are always within a block { } of some kind and are terminated via a Semicolon, except if the statement ends in a block, such as a Statement::Expression containing an if expression.

UseTreeKind

Constants

ERROR_IDENT

This is used when an identifier fails to parse in the parser. Instead of failing the parse, we can often recover using this as the default value instead. Further passes like name resolution should also check for this ident to avoid issuing multiple errors for an identifier that already failed to parse.

WILDCARD_TYPE

This is used to represent an UnresolvedTypeData::Unspecified in a Path