noirc_frontend::ast
Module traits

Structs

NoirTrait

AST node for trait definitions: trait name<generics> { ... items ... }

NoirTraitImpl

Ast node for an implementation of a trait for a particular type impl trait_name<trait_generics> for object_type where where_clauses { ... items ... }

TraitBound

Represents a single trait bound, such as TraitX or TraitY<U, V>

TraitImplItem

TypeImpl

Ast node for an impl of a concrete type impl object_type<generics> { ... methods ... }

UnresolvedTraitConstraint

Represents a simple trait constraint such as where Foo: TraitY<U, V> Complex trait constraints such as where Foo: Display + TraitX + TraitY<U, V> are converted in the parser to a series of simple constraints: Foo: Display Foo: TraitX Foo: TraitY<U, V>

Enums

TraitImplItemKind

TraitItem

Any declaration inside the body of a trait that a user is required to specify when implementing the trait.

Functions

desugar_generic_trait_bounds 🔒

Moves trait bounds from generics into where clauses. For example:

desugar_generic_trait_bounds_and_reorder_where_clause 🔒

Does both desugar_generic_trait_bounds and reorder_where_clause.

reorder_where_clause 🔒

Reorders a where clause in-place so that simpler constraints come before more complex ones. The resulting where clause will have constraints in this order: