noirc_frontend::ast::visitor
Trait Visitor

pub trait Visitor {

    // Provided methods
    fn visit_parsed_module(&mut self, _: &ParsedModule) -> bool { ... }
    fn visit_item(&mut self, _: &Item) -> bool { ... }
    fn visit_parsed_submodule(&mut self, _: &ParsedSubModule, _: Span) -> bool { ... }
    fn visit_noir_function(&mut self, _: &NoirFunction, _: Span) -> bool { ... }
    fn visit_noir_trait_impl(&mut self, _: &NoirTraitImpl, _: Span) -> bool { ... }
    fn visit_type_impl(&mut self, _: &TypeImpl, _: Span) -> bool { ... }
    fn visit_trait_impl_item(&mut self, _: &TraitImplItem) -> bool { ... }
    fn visit_trait_impl_item_kind(
        &mut self,
        _: &TraitImplItemKind,
        _span: Span,
    ) -> bool { ... }
    fn visit_trait_impl_item_function(
        &mut self,
        _: &NoirFunction,
        _span: Span,
    ) -> bool { ... }
    fn visit_trait_impl_item_constant(
        &mut self,
        _name: &Ident,
        _typ: Option<&UnresolvedType>,
        _expression: &Expression,
        _span: Span,
    ) -> bool { ... }
    fn visit_trait_impl_item_type(
        &mut self,
        _name: &Ident,
        _alias: Option<&UnresolvedType>,
        _span: Span,
    ) -> bool { ... }
    fn visit_noir_trait(&mut self, _: &NoirTrait, _: Span) -> bool { ... }
    fn visit_trait_item(&mut self, _: &TraitItem) -> bool { ... }
    fn visit_trait_item_function(
        &mut self,
        _name: &Ident,
        _generics: &UnresolvedGenerics,
        _parameters: &[(Ident, UnresolvedType)],
        _return_type: &FunctionReturnType,
        _where_clause: &[UnresolvedTraitConstraint],
        _body: &Option<BlockExpression>,
    ) -> bool { ... }
    fn visit_trait_item_constant(
        &mut self,
        _name: &Ident,
        _typ: Option<&UnresolvedType>,
    ) -> bool { ... }
    fn visit_trait_item_type(&mut self, _: &Ident, _: &[TraitBound]) -> bool { ... }
    fn visit_use_tree(&mut self, _: &UseTree) -> bool { ... }
    fn visit_use_tree_path(
        &mut self,
        _: &UseTree,
        _ident: &Ident,
        _alias: &Option<Ident>,
    ) { ... }
    fn visit_use_tree_list(&mut self, _: &UseTree, _: &[UseTree]) -> bool { ... }
    fn visit_noir_struct(&mut self, _: &NoirStruct, _: Span) -> bool { ... }
    fn visit_noir_enum(&mut self, _: &NoirEnumeration, _: Span) -> bool { ... }
    fn visit_noir_type_alias(&mut self, _: &TypeAlias, _: Span) -> bool { ... }
    fn visit_module_declaration(&mut self, _: &ModuleDeclaration, _: Span) { ... }
    fn visit_expression(&mut self, _: &Expression) -> bool { ... }
    fn visit_literal(&mut self, _: &Literal, _: Span) -> bool { ... }
    fn visit_literal_array(&mut self, _: &ArrayLiteral, _: Span) -> bool { ... }
    fn visit_literal_slice(&mut self, _: &ArrayLiteral, _: Span) -> bool { ... }
    fn visit_literal_bool(&mut self, _: bool, _: Span) { ... }
    fn visit_literal_integer(
        &mut self,
        _value: SignedField,
        _suffix: Option<IntegerTypeSuffix>,
        _: Span,
    ) { ... }
    fn visit_literal_str(&mut self, _: &str, _: Span) { ... }
    fn visit_literal_raw_str(&mut self, _: &str, _: u8, _: Span) { ... }
    fn visit_literal_fmt_str(
        &mut self,
        _: &[FmtStrFragment],
        _length: u32,
        _: Span,
    ) { ... }
    fn visit_literal_unit(&mut self, _: Span) { ... }
    fn visit_block_expression(
        &mut self,
        _: &BlockExpression,
        _: Option<Span>,
    ) -> bool { ... }
    fn visit_prefix_expression(&mut self, _: &PrefixExpression, _: Span) -> bool { ... }
    fn visit_index_expression(&mut self, _: &IndexExpression, _: Span) -> bool { ... }
    fn visit_call_expression(&mut self, _: &CallExpression, _: Span) -> bool { ... }
    fn visit_method_call_expression(
        &mut self,
        _: &MethodCallExpression,
        _: Span,
    ) -> bool { ... }
    fn visit_constructor_expression(
        &mut self,
        _: &ConstructorExpression,
        _: Span,
    ) -> bool { ... }
    fn visit_member_access_expression(
        &mut self,
        _: &MemberAccessExpression,
        _: Span,
    ) -> bool { ... }
    fn visit_cast_expression(&mut self, _: &CastExpression, _: Span) -> bool { ... }
    fn visit_infix_expression(&mut self, _: &InfixExpression, _: Span) -> bool { ... }
    fn visit_if_expression(&mut self, _: &IfExpression, _: Span) -> bool { ... }
    fn visit_match_expression(&mut self, _: &MatchExpression, _: Span) -> bool { ... }
    fn visit_tuple(&mut self, _: &[Expression], _: Span) -> bool { ... }
    fn visit_parenthesized(&mut self, _: &Expression, _: Span) -> bool { ... }
    fn visit_unquote(&mut self, _: &Expression, _: Span) -> bool { ... }
    fn visit_comptime_expression(
        &mut self,
        _: &BlockExpression,
        _: Span,
    ) -> bool { ... }
    fn visit_unsafe_expression(&mut self, _: &UnsafeExpression, _: Span) -> bool { ... }
    fn visit_variable(&mut self, _: &Path, _: Span) -> bool { ... }
    fn visit_quote(&mut self, _: &Tokens) { ... }
    fn visit_resolved_expression(&mut self, _expr_id: ExprId) { ... }
    fn visit_interned_expression(&mut self, _id: InternedExpressionKind) { ... }
    fn visit_error_expression(&mut self) { ... }
    fn visit_lambda(&mut self, _: &Lambda, _: Span) -> bool { ... }
    fn visit_array_literal(&mut self, _: &ArrayLiteral, _: Span) -> bool { ... }
    fn visit_array_literal_standard(
        &mut self,
        _: &[Expression],
        _: Span,
    ) -> bool { ... }
    fn visit_array_literal_repeated(
        &mut self,
        _repeated_element: &Expression,
        _length: &Expression,
        _: Span,
    ) -> bool { ... }
    fn visit_statement(&mut self, _: &Statement) -> bool { ... }
    fn visit_import(
        &mut self,
        _: &UseTree,
        _: Span,
        _visibility: ItemVisibility,
    ) -> bool { ... }
    fn visit_global(&mut self, _: &LetStatement, _: Span) -> bool { ... }
    fn visit_let_statement(&mut self, _: &LetStatement) -> bool { ... }
    fn visit_constrain_statement(&mut self, _: &ConstrainExpression) -> bool { ... }
    fn visit_assign_statement(&mut self, _: &AssignStatement) -> bool { ... }
    fn visit_for_loop_statement(&mut self, _: &ForLoopStatement) -> bool { ... }
    fn visit_loop_statement(&mut self, _: &Expression) -> bool { ... }
    fn visit_while_statement(
        &mut self,
        _condition: &Expression,
        _body: &Expression,
    ) -> bool { ... }
    fn visit_comptime_statement(&mut self, _: &Statement) -> bool { ... }
    fn visit_break(&mut self) { ... }
    fn visit_continue(&mut self) { ... }
    fn visit_interned_statement(&mut self, _: InternedStatementKind) { ... }
    fn visit_error_statement(&mut self) { ... }
    fn visit_lvalue(&mut self, _: &LValue) -> bool { ... }
    fn visit_lvalue_path(&mut self, _: &Path) { ... }
    fn visit_lvalue_member_access(
        &mut self,
        _object: &LValue,
        _field_name: &Ident,
        _span: Span,
    ) -> bool { ... }
    fn visit_lvalue_index(
        &mut self,
        _array: &LValue,
        _index: &Expression,
        _span: Span,
    ) -> bool { ... }
    fn visit_lvalue_dereference(
        &mut self,
        _lvalue: &LValue,
        _span: Span,
    ) -> bool { ... }
    fn visit_lvalue_interned(
        &mut self,
        _id: InternedExpressionKind,
        _span: Span,
    ) { ... }
    fn visit_for_range(&mut self, _: &ForRange) -> bool { ... }
    fn visit_as_trait_path(&mut self, _: &AsTraitPath, _: Span) -> bool { ... }
    fn visit_type_path(&mut self, _: &TypePath, _: Span) -> bool { ... }
    fn visit_unresolved_type(&mut self, _: &UnresolvedType) -> bool { ... }
    fn visit_array_type(
        &mut self,
        _: &UnresolvedTypeExpression,
        _: &UnresolvedType,
        _: Span,
    ) -> bool { ... }
    fn visit_slice_type(&mut self, _: &UnresolvedType, _: Span) -> bool { ... }
    fn visit_parenthesized_type(&mut self, _: &UnresolvedType, _: Span) -> bool { ... }
    fn visit_named_type(
        &mut self,
        _: &Path,
        _: &GenericTypeArgs,
        _: Span,
    ) -> bool { ... }
    fn visit_trait_as_type(
        &mut self,
        _: &Path,
        _: &GenericTypeArgs,
        _: Span,
    ) -> bool { ... }
    fn visit_reference_type(
        &mut self,
        _: &UnresolvedType,
        _mutable: bool,
        _: Span,
    ) -> bool { ... }
    fn visit_tuple_type(&mut self, _: &[UnresolvedType], _: Span) -> bool { ... }
    fn visit_function_type(
        &mut self,
        _args: &[UnresolvedType],
        _ret: &UnresolvedType,
        _env: &UnresolvedType,
        _unconstrained: bool,
        _span: Span,
    ) -> bool { ... }
    fn visit_as_trait_path_type(&mut self, _: &AsTraitPath, _: Span) -> bool { ... }
    fn visit_expression_type(
        &mut self,
        _: &UnresolvedTypeExpression,
        _: Span,
    ) -> bool { ... }
    fn visit_unspecified_type(&mut self, _: Span) { ... }
    fn visit_unit_type(&mut self, _: Span) { ... }
    fn visit_resolved_type(&mut self, _: QuotedTypeId, _: Location) { ... }
    fn visit_interned_type(&mut self, _: InternedUnresolvedTypeData, _: Span) { ... }
    fn visit_error_type(&mut self, _: Span) { ... }
    fn visit_path(&mut self, _: &Path) { ... }
    fn visit_generic_type_args(&mut self, _: &GenericTypeArgs) -> bool { ... }
    fn visit_unresolved_generic(&mut self, _: &UnresolvedGeneric) -> bool { ... }
    fn visit_function_return_type(&mut self, _: &FunctionReturnType) -> bool { ... }
    fn visit_trait_bound(&mut self, _: &TraitBound) -> bool { ... }
    fn visit_unresolved_trait_constraint(
        &mut self,
        _: &UnresolvedTraitConstraint,
    ) -> bool { ... }
    fn visit_unresolved_type_expression(
        &mut self,
        _: &UnresolvedTypeExpression,
    ) -> bool { ... }
    fn visit_variable_type_expression(&mut self, _: &Path) -> bool { ... }
    fn visit_constant_type_expression(
        &mut self,
        _value: SignedField,
        _suffix: Option<IntegerTypeSuffix>,
        _span: Span,
    ) { ... }
    fn visit_binary_type_expression(
        &mut self,
        _lhs: &UnresolvedTypeExpression,
        _op: BinaryTypeOperator,
        _rhs: &UnresolvedTypeExpression,
        _span: Span,
    ) -> bool { ... }
    fn visit_as_trait_path_type_expression(
        &mut self,
        _as_trait_path: &AsTraitPath,
    ) -> bool { ... }
    fn visit_pattern(&mut self, _: &Pattern) -> bool { ... }
    fn visit_identifier_pattern(&mut self, _: &Ident) { ... }
    fn visit_mutable_pattern(
        &mut self,
        _: &Pattern,
        _: Span,
        _is_synthesized: bool,
    ) -> bool { ... }
    fn visit_tuple_pattern(&mut self, _: &[Pattern], _: Span) -> bool { ... }
    fn visit_struct_pattern(
        &mut self,
        _: &Path,
        _: &[(Ident, Pattern)],
        _: Span,
    ) -> bool { ... }
    fn visit_parenthesized_pattern(&mut self, _: &Pattern, _: Span) -> bool { ... }
    fn visit_interned_pattern(&mut self, _: &InternedPattern, _: Span) { ... }
    fn visit_secondary_attribute(
        &mut self,
        _: &SecondaryAttribute,
        _target: AttributeTarget,
    ) -> bool { ... }
    fn visit_secondary_attribute_kind(
        &mut self,
        _: &SecondaryAttributeKind,
        _target: AttributeTarget,
        _span: Span,
    ) -> bool { ... }
    fn visit_meta_attribute(
        &mut self,
        _: &MetaAttribute,
        _target: AttributeTarget,
        _span: Span,
    ) -> bool { ... }

}

Implements the Visitor pattern for Noir’s AST.

In this implementation, methods must return a bool:

true means children must be visited

false means children must not be visited, either because the visitor implementation will visit children of interest manually, or because no children are of interest

Provided Methods

fn visit_parsed_module(&mut self, _: &ParsedModule) -> bool

fn visit_item(&mut self, _: &Item) -> bool

fn visit_parsed_submodule(&mut self, _: &ParsedSubModule, _: Span) -> bool

fn visit_noir_function(&mut self, _: &NoirFunction, _: Span) -> bool

fn visit_noir_trait_impl(&mut self, _: &NoirTraitImpl, _: Span) -> bool

fn visit_type_impl(&mut self, _: &TypeImpl, _: Span) -> bool

fn visit_trait_impl_item(&mut self, _: &TraitImplItem) -> bool

fn visit_trait_impl_item_kind( &mut self, _: &TraitImplItemKind, _span: Span, ) -> bool

fn visit_trait_impl_item_function( &mut self, _: &NoirFunction, _span: Span, ) -> bool

fn visit_trait_impl_item_constant( &mut self, _name: &Ident, _typ: Option<&UnresolvedType>, _expression: &Expression, _span: Span, ) -> bool

fn visit_trait_impl_item_type( &mut self, _name: &Ident, _alias: Option<&UnresolvedType>, _span: Span, ) -> bool

fn visit_noir_trait(&mut self, _: &NoirTrait, _: Span) -> bool

fn visit_trait_item(&mut self, _: &TraitItem) -> bool

fn visit_trait_item_function( &mut self, _name: &Ident, _generics: &UnresolvedGenerics, _parameters: &[(Ident, UnresolvedType)], _return_type: &FunctionReturnType, _where_clause: &[UnresolvedTraitConstraint], _body: &Option<BlockExpression>, ) -> bool

fn visit_trait_item_constant( &mut self, _name: &Ident, _typ: Option<&UnresolvedType>, ) -> bool

fn visit_trait_item_type(&mut self, _: &Ident, _: &[TraitBound]) -> bool

fn visit_use_tree(&mut self, _: &UseTree) -> bool

fn visit_use_tree_path( &mut self, _: &UseTree, _ident: &Ident, _alias: &Option<Ident>, )

fn visit_use_tree_list(&mut self, _: &UseTree, _: &[UseTree]) -> bool

fn visit_noir_struct(&mut self, _: &NoirStruct, _: Span) -> bool

fn visit_noir_enum(&mut self, _: &NoirEnumeration, _: Span) -> bool

fn visit_noir_type_alias(&mut self, _: &TypeAlias, _: Span) -> bool

fn visit_module_declaration(&mut self, _: &ModuleDeclaration, _: Span)

fn visit_expression(&mut self, _: &Expression) -> bool

fn visit_literal(&mut self, _: &Literal, _: Span) -> bool

fn visit_literal_array(&mut self, _: &ArrayLiteral, _: Span) -> bool

fn visit_literal_slice(&mut self, _: &ArrayLiteral, _: Span) -> bool

fn visit_literal_bool(&mut self, _: bool, _: Span)

fn visit_literal_integer( &mut self, _value: SignedField, _suffix: Option<IntegerTypeSuffix>, _: Span, )

fn visit_literal_str(&mut self, _: &str, _: Span)

fn visit_literal_raw_str(&mut self, _: &str, _: u8, _: Span)

fn visit_literal_fmt_str(&mut self, _: &[FmtStrFragment], _length: u32, _: Span)

fn visit_literal_unit(&mut self, _: Span)

fn visit_block_expression( &mut self, _: &BlockExpression, _: Option<Span>, ) -> bool

fn visit_prefix_expression(&mut self, _: &PrefixExpression, _: Span) -> bool

fn visit_index_expression(&mut self, _: &IndexExpression, _: Span) -> bool

fn visit_call_expression(&mut self, _: &CallExpression, _: Span) -> bool

fn visit_method_call_expression( &mut self, _: &MethodCallExpression, _: Span, ) -> bool

fn visit_constructor_expression( &mut self, _: &ConstructorExpression, _: Span, ) -> bool

fn visit_member_access_expression( &mut self, _: &MemberAccessExpression, _: Span, ) -> bool

fn visit_cast_expression(&mut self, _: &CastExpression, _: Span) -> bool

fn visit_infix_expression(&mut self, _: &InfixExpression, _: Span) -> bool

fn visit_if_expression(&mut self, _: &IfExpression, _: Span) -> bool

fn visit_match_expression(&mut self, _: &MatchExpression, _: Span) -> bool

fn visit_tuple(&mut self, _: &[Expression], _: Span) -> bool

fn visit_parenthesized(&mut self, _: &Expression, _: Span) -> bool

fn visit_unquote(&mut self, _: &Expression, _: Span) -> bool

fn visit_comptime_expression(&mut self, _: &BlockExpression, _: Span) -> bool

fn visit_unsafe_expression(&mut self, _: &UnsafeExpression, _: Span) -> bool

fn visit_variable(&mut self, _: &Path, _: Span) -> bool

fn visit_quote(&mut self, _: &Tokens)

fn visit_resolved_expression(&mut self, _expr_id: ExprId)

fn visit_interned_expression(&mut self, _id: InternedExpressionKind)

fn visit_error_expression(&mut self)

fn visit_lambda(&mut self, _: &Lambda, _: Span) -> bool

fn visit_array_literal(&mut self, _: &ArrayLiteral, _: Span) -> bool

fn visit_array_literal_standard(&mut self, _: &[Expression], _: Span) -> bool

fn visit_array_literal_repeated( &mut self, _repeated_element: &Expression, _length: &Expression, _: Span, ) -> bool

fn visit_statement(&mut self, _: &Statement) -> bool

fn visit_import( &mut self, _: &UseTree, _: Span, _visibility: ItemVisibility, ) -> bool

fn visit_global(&mut self, _: &LetStatement, _: Span) -> bool

fn visit_let_statement(&mut self, _: &LetStatement) -> bool

fn visit_constrain_statement(&mut self, _: &ConstrainExpression) -> bool

fn visit_assign_statement(&mut self, _: &AssignStatement) -> bool

fn visit_for_loop_statement(&mut self, _: &ForLoopStatement) -> bool

fn visit_loop_statement(&mut self, _: &Expression) -> bool

fn visit_while_statement( &mut self, _condition: &Expression, _body: &Expression, ) -> bool

fn visit_comptime_statement(&mut self, _: &Statement) -> bool

fn visit_break(&mut self)

fn visit_continue(&mut self)

fn visit_interned_statement(&mut self, _: InternedStatementKind)

fn visit_error_statement(&mut self)

fn visit_lvalue(&mut self, _: &LValue) -> bool

fn visit_lvalue_path(&mut self, _: &Path)

fn visit_lvalue_member_access( &mut self, _object: &LValue, _field_name: &Ident, _span: Span, ) -> bool

fn visit_lvalue_index( &mut self, _array: &LValue, _index: &Expression, _span: Span, ) -> bool

fn visit_lvalue_dereference(&mut self, _lvalue: &LValue, _span: Span) -> bool

fn visit_lvalue_interned(&mut self, _id: InternedExpressionKind, _span: Span)

fn visit_for_range(&mut self, _: &ForRange) -> bool

fn visit_as_trait_path(&mut self, _: &AsTraitPath, _: Span) -> bool

fn visit_type_path(&mut self, _: &TypePath, _: Span) -> bool

fn visit_unresolved_type(&mut self, _: &UnresolvedType) -> bool

fn visit_array_type( &mut self, _: &UnresolvedTypeExpression, _: &UnresolvedType, _: Span, ) -> bool

fn visit_slice_type(&mut self, _: &UnresolvedType, _: Span) -> bool

fn visit_parenthesized_type(&mut self, _: &UnresolvedType, _: Span) -> bool

fn visit_named_type(&mut self, _: &Path, _: &GenericTypeArgs, _: Span) -> bool

fn visit_trait_as_type( &mut self, _: &Path, _: &GenericTypeArgs, _: Span, ) -> bool

fn visit_reference_type( &mut self, _: &UnresolvedType, _mutable: bool, _: Span, ) -> bool

fn visit_tuple_type(&mut self, _: &[UnresolvedType], _: Span) -> bool

fn visit_function_type( &mut self, _args: &[UnresolvedType], _ret: &UnresolvedType, _env: &UnresolvedType, _unconstrained: bool, _span: Span, ) -> bool

fn visit_as_trait_path_type(&mut self, _: &AsTraitPath, _: Span) -> bool

fn visit_expression_type( &mut self, _: &UnresolvedTypeExpression, _: Span, ) -> bool

fn visit_unspecified_type(&mut self, _: Span)

fn visit_unit_type(&mut self, _: Span)

fn visit_resolved_type(&mut self, _: QuotedTypeId, _: Location)

fn visit_interned_type(&mut self, _: InternedUnresolvedTypeData, _: Span)

fn visit_error_type(&mut self, _: Span)

fn visit_path(&mut self, _: &Path)

fn visit_generic_type_args(&mut self, _: &GenericTypeArgs) -> bool

fn visit_unresolved_generic(&mut self, _: &UnresolvedGeneric) -> bool

fn visit_function_return_type(&mut self, _: &FunctionReturnType) -> bool

fn visit_trait_bound(&mut self, _: &TraitBound) -> bool

fn visit_unresolved_trait_constraint( &mut self, _: &UnresolvedTraitConstraint, ) -> bool

fn visit_unresolved_type_expression( &mut self, _: &UnresolvedTypeExpression, ) -> bool

fn visit_variable_type_expression(&mut self, _: &Path) -> bool

fn visit_constant_type_expression( &mut self, _value: SignedField, _suffix: Option<IntegerTypeSuffix>, _span: Span, )

fn visit_binary_type_expression( &mut self, _lhs: &UnresolvedTypeExpression, _op: BinaryTypeOperator, _rhs: &UnresolvedTypeExpression, _span: Span, ) -> bool

fn visit_as_trait_path_type_expression( &mut self, _as_trait_path: &AsTraitPath, ) -> bool

fn visit_pattern(&mut self, _: &Pattern) -> bool

fn visit_identifier_pattern(&mut self, _: &Ident)

fn visit_mutable_pattern( &mut self, _: &Pattern, _: Span, _is_synthesized: bool, ) -> bool

fn visit_tuple_pattern(&mut self, _: &[Pattern], _: Span) -> bool

fn visit_struct_pattern( &mut self, _: &Path, _: &[(Ident, Pattern)], _: Span, ) -> bool

fn visit_parenthesized_pattern(&mut self, _: &Pattern, _: Span) -> bool

fn visit_interned_pattern(&mut self, _: &InternedPattern, _: Span)

fn visit_secondary_attribute( &mut self, _: &SecondaryAttribute, _target: AttributeTarget, ) -> bool

fn visit_secondary_attribute_kind( &mut self, _: &SecondaryAttributeKind, _target: AttributeTarget, _span: Span, ) -> bool

fn visit_meta_attribute( &mut self, _: &MetaAttribute, _target: AttributeTarget, _span: Span, ) -> bool