noirc_frontend::elaborator
Module lints

Lint checks for function attributes, visibility, and usage restrictions.

Functions

can_return_without_recursing 🔒

Check if an expression will end up calling a specific function.

can_return_without_recursing_match 🔒

check_integer_literal_fits_its_type 🔒

Checks if an ExprId, which has to be an integer literal, fits in its type.

databus_on_non_entry_point 🔒

call_data and return_data visibility modifiers are only allowed on entry point functions.

deprecated_function 🔒

func_meta_name_ident 🔒

inlining_attributes 🔒

Inline attributes are only relevant for constrained functions as all unconstrained functions are not inlined and so associated attributes are disallowed.

low_level_function_outside_stdlib 🔒

Attempting to define new low level (#[builtin] or #[foreign]) functions outside of the stdlib is disallowed.

missing_pub 🔒

pub is required on return types for entry point functions

oracle_called_from_constrained_function 🔒

Oracle functions may not be called by constrained functions directly.

oracle_not_marked_unconstrained 🔒

Oracle definitions (functions with the #[oracle] attribute) must be marked as unconstrained.

oracle_returns_multiple_slices 🔒

Oracle functions cannot return more than 1 slice in their output.

unbounded_recursion 🔒

Check that a recursive function can return without endlessly calling itself.

unconstrained_function_args 🔒

Check that we are not passing a mutable reference from a constrained runtime to an unconstrained runtime.

unconstrained_function_return 🔒

Check that that a type returned from an unconstrained to a constrained runtime is safe:

unnecessary_pub_argument 🔒

Only arguments to entrypoint functions may have a non-private visibility modifier applied to them.

unnecessary_pub_return 🔒

Only entrypoint functions require a pub visibility modifier applied to their return types.