noirc_frontend::elaborator
Struct Elaborator

pub struct Elaborator<'context> {

    scopes: ScopeForest<String, ResolverMeta>,
    pub(crate) errors: Vec<CompilationError>,
    pub(crate) interner: &'context mut NodeInterner,
    pub(crate) def_maps: &'context mut DefMaps,
    pub(crate) usage_tracker: &'context mut UsageTracker,
    pub(crate) crate_graph: &'context CrateGraph,
    pub(crate) interpreter_output: &'context Option<Rc<RefCell<dyn Write>>>,
    unsafe_block_status: UnsafeBlockStatus,
    current_loop: Option<Loop>,
    generics: Vec<ResolvedGeneric>,
    lambda_stack: Vec<LambdaContext>,
    self_type: Option<Type>,
    current_item: Option<DependencyId>,
    current_trait_impl: Option<TraitImplId>,
    current_trait: Option<TraitId>,
    resolving_ids: BTreeSet<TypeId>,
    trait_bounds: Vec<TraitConstraint>,
    function_context: Vec<FunctionContext>,
    local_module: LocalModuleId,
    in_comptime_context: bool,
    crate_id: CrateId,
    unresolved_globals: BTreeMap<GlobalId, UnresolvedGlobal>,
    pub(crate) interpreter_call_stack: Vector<Location>,
    silence_field_visibility_errors: usize,
    options: GenericOptions<'context, FileId>,
    pub(crate) elaborate_reasons: Vector<ElaborateReason>,

}

Fields
scopes: ScopeForest<String, ResolverMeta>errors: Vec<CompilationError>interner: &'context mut NodeInternerdef_maps: &'context mut DefMapsusage_tracker: &'context mut UsageTrackercrate_graph: &'context CrateGraphinterpreter_output: &'context Option<Rc<RefCell<dyn Write>>>unsafe_block_status: UnsafeBlockStatuscurrent_loop: Option<Loop>generics: Vec<ResolvedGeneric>

Contains a mapping of the current struct or functions’s generics to unique type variables if we’re resolving a struct. Empty otherwise. This is a Vec rather than a map to preserve the order a functions generics were declared in.

lambda_stack: Vec<LambdaContext>

When resolving lambda expressions, we need to keep track of the variables that are captured. We do this in order to create the hidden environment parameter for the lambda function.

self_type: Option<Type>

Set to the current type if we’re resolving an impl

current_item: Option<DependencyId>

The current dependency item we’re resolving. Used to link items to their dependencies in the dependency graph

current_trait_impl: Option<TraitImplId>

If we’re currently resolving methods within a trait impl, this will be set to the corresponding trait impl ID.

current_trait: Option<TraitId>

The trait we’re currently resolving, if we are resolving one.

resolving_ids: BTreeSet<TypeId>

In-resolution names

This needs to be a set because we can have multiple in-resolution names when resolving structs that are declared in reverse order of their dependencies, such as in the following case:

struct Wrapper {
    value: Wrapped
}
struct Wrapped {
}

trait_bounds: Vec<TraitConstraint>

Each constraint in the where clause of the function currently being resolved.

function_context: Vec<FunctionContext>

This is a stack of function contexts. Most of the time, for each function we expect this to be of length one, containing each type variable and trait constraint used in the function. This is also pushed to when a comptime {} block is used within the function. Since it can force us to resolve that block’s trait constraints earlier so that they are resolved when the interpreter is run before the enclosing function is finished elaborating. When this happens, we need to resolve any type variables that were made within this block as well so that we can solve these traits.

local_module: LocalModuleId

The current module this elaborator is in. Initially empty, it is set whenever a new top-level item is resolved.

in_comptime_context: bool

True if we’re elaborating a comptime item such as a comptime function, block, global, or attribute.

crate_id: CrateIdunresolved_globals: BTreeMap<GlobalId, UnresolvedGlobal>

These are the globals that have yet to be elaborated. This map is used to lazily evaluate these globals if they’re encountered before they are elaborated (e.g. in a function’s type or another global’s RHS).

interpreter_call_stack: Vector<Location>silence_field_visibility_errors: usize

If greater than 0, field visibility errors won’t be reported. This is used when elaborating a comptime expression that is a struct constructor like Foo { inner: 5 }: in that case we already elaborated the code that led to that comptime value and any visibility errors were already reported.

options: GenericOptions<'context, FileId>

Options from the nargo cli

elaborate_reasons: Vector<ElaborateReason>

Sometimes items are elaborated because a function attribute ran and generated items. The Elaborator keeps track of these reasons so that when an error is produced it will be wrapped in another error that will include this reason.

Implementations

impl<'context> Elaborator<'context>

pub fn elaborate_item_from_comptime_in_function<'a, T>( &'a mut self, current_function: Option<FuncId>, reason: Option<ElaborateReason>, f: impl FnOnce(&mut Elaborator<'a>) -> T, ) -> T

Elaborate an expression from the middle of a comptime scope. When this happens we require additional information to know what variables should be in scope.

pub fn elaborate_item_from_comptime_in_module<'a, T>( &'a mut self, module: ModuleId, reason: Option<ElaborateReason>, f: impl FnOnce(&mut Elaborator<'a>) -> T, ) -> T

fn elaborate_item_from_comptime<'a, T>( &'a mut self, reason: Option<ElaborateReason>, f: impl FnOnce(&mut Elaborator<'a>) -> T, setup: impl FnOnce(&mut Elaborator<'a>), ) -> T

fn populate_scope_from_comptime_scopes(&mut self)

fn collect_comptime_attributes_on_item( &mut self, attributes: &[SecondaryAttribute], item: Value, attribute_context: AttributeContext, attributes_to_run: &mut Vec<(FuncId, Value, Vec<Expression>, AttributeContext, Location)>, )

fn collect_comptime_attribute_on_item( &mut self, attribute: &SecondaryAttribute, item: &Value, attribute_context: AttributeContext, attributes_to_run: &mut Vec<(FuncId, Value, Vec<Expression>, AttributeContext, Location)>, )

fn collect_comptime_attribute_name_on_item( &mut self, attribute: &MetaAttribute, location: Location, item: Value, attribute_context: AttributeContext, attributes_to_run: &mut Vec<(FuncId, Value, Vec<Expression>, AttributeContext, Location)>, ) -> Result<(), CompilationError>

Resolve an attribute to the function it refers to and add it to attributes_to_run

fn run_attribute( &mut self, attribute_context: AttributeContext, function: FuncId, arguments: Vec<Expression>, item: Value, location: Location, generated_items: &mut CollectedItems, ) -> Result<(), CompilationError>

fn handle_attribute_arguments( interpreter: &mut Interpreter<'_, '_>, item: &Value, function: FuncId, arguments: Vec<Expression>, location: Location, ) -> Result<Vec<(Value, Location)>, CompilationError>

fn add_items( &mut self, items: Vec<Item>, generated_items: &mut CollectedItems, location: Location, )

pub(crate) fn add_item( &mut self, item: Item, generated_items: &mut CollectedItems, location: Location, )

pub fn setup_interpreter<'local>( &'local mut self, ) -> Interpreter<'local, 'context>

pub(super) fn debug_comptime<T: Display, F: FnMut(&mut NodeInterner) -> T>( &mut self, location: Location, expr_f: F, )

pub(super) fn run_attributes( &mut self, traits: &BTreeMap<TraitId, UnresolvedTrait>, types: &BTreeMap<TypeId, UnresolvedStruct>, functions: &[UnresolvedFunctions], module_attributes: &[ModuleAttribute], )

Run all the attributes on each item in the crate in source-order. Source-order is defined as running all child modules before their parent modules are run. Child modules of a parent are run in order of their mod foo; declarations in the parent.

fn sort_attributes_by_run_order( &self, attributes: &mut Vec<(FuncId, Value, Vec<Expression>, AttributeContext, Location)>, )

fn collect_attributes_on_modules( &mut self, module_attributes: &[ModuleAttribute], attributes_to_run: &mut Vec<(FuncId, Value, Vec<Expression>, AttributeContext, Location)>, )

fn collect_attributes_on_functions( &mut self, function_sets: &[UnresolvedFunctions], attributes_to_run: &mut Vec<(FuncId, Value, Vec<Expression>, AttributeContext, Location)>, )

pub(super) fn elaborate_in_comptime_context<T>( &mut self, f: impl FnOnce(&mut Self) -> T, ) -> T

Perform the given function in a comptime context. This will set the in_comptime_context flag on self as well as push a new function context to resolve any trait constraints early.

pub(super) fn in_comptime_context(&self) -> bool

True if we’re currently within a comptime block, function, or global

pub(crate) fn with_elaborate_reason<F, T>( &mut self, reason: ElaborateReason, f: F, ) -> T
where F: FnOnce(&mut Elaborator<'_>) -> T,

fn wrap_errors_in_macro_error( &self, errors: Vec<CompilationError>, ) -> Vec<CompilationError>

fn wrap_error_in_macro_error(&self, error: CompilationError) -> CompilationError

impl Elaborator<'_>

pub(super) fn define_enum_variant_constructor( &mut self, enum_: &NoirEnumeration, type_id: TypeId, variant: &EnumVariant, variant_arg_types: Option<Vec<Type>>, variant_index: usize, datatype: &Shared<DataType>, self_type: &Type, self_type_unresolved: UnresolvedType, )

Defines the value of an enum variant that we resolve an enum variant expression to. E.g. Foo::Bar in Foo::Bar(baz).

If the variant requires arguments we should define a function, otherwise we define a polymorphic global containing the tag value.

fn define_enum_variant_global( &mut self, enum_: &NoirEnumeration, type_id: TypeId, variant: &EnumVariant, variant_index: usize, datatype: &Shared<DataType>, self_type: &Type, )

fn define_enum_variant_function( &mut self, enum_: &NoirEnumeration, type_id: TypeId, variant: &EnumVariant, variant_arg_types: Vec<Type>, variant_index: usize, datatype: &Shared<DataType>, self_type: &Type, self_type_unresolved: UnresolvedType, )

fn make_enum_variant_constructor( &mut self, self_type: &Shared<DataType>, variant_index: usize, parameters: &Parameters, location: Location, ) -> ExprId

fn make_enum_variant_parameters( &mut self, parameter_types: Vec<Type>, location: Location, ) -> Parameters

pub(super) fn elaborate_match_rules( &mut self, variable_to_match: DefinitionId, rules: Vec<(Expression, Expression)>, ) -> (Vec<Row>, Type)

To elaborate the rules of a match we need to go through the pattern first to define all the variables within, then compile the corresponding branch. For each branch we do this way we’ll need to keep a distinct scope so that branches cannot access the pattern variables from other branches.

Returns (rows, result type) where rows is a pattern matrix used to compile the match into a decision tree.

fn expression_to_pattern( &mut self, expression: Expression, expected_type: &Type, variables_defined: &mut Vec<Ident>, ) -> Pattern

Convert an expression into a Pattern, defining any variables within.

fn define_pattern_variable( &mut self, name: Ident, expected_type: &Type, variables_defined: &mut Vec<Ident>, ) -> Pattern

fn constructor_to_pattern( &mut self, constructor: ConstructorExpression, variables_defined: &mut Vec<Ident>, ) -> Pattern

fn expression_to_constructor( &mut self, name: Expression, args: Vec<Expression>, expected_type: &Type, variables_defined: &mut Vec<Ident>, ) -> Pattern

fn path_resolution_to_constructor( &mut self, resolution: PathResolutionItem, name: Option<TypedPathSegment>, args: Vec<Expression>, expected_type: &Type, location: Location, variables_defined: &mut Vec<Ident>, ) -> Pattern

Convert a PathResolutionItem - usually an enum variant or global - to a Constructor. If name is Some, we’ll define a Pattern::Binding instead of erroring if the item doesn’t resolve to a variant or global. This would shadow an existing value such as a free function. Generally this is desired unless the variable was a path with multiple components such as foo::bar which should always be treated as a path to an existing item.

fn global_constant_to_integer_constructor( &mut self, constant: Value, expected_type: &Type, location: Location, ) -> Pattern

fn struct_name_and_field_types( &mut self, typ: &Type, location: Location, ) -> Option<(Ident, Vec<(String, Type, ItemVisibility)>)>

pub(super) fn elaborate_match_rows( &mut self, rows: Vec<Row>, type_matched_on: &Type, location: Location, ) -> HirMatch

Compiles the rows of a match expression, outputting a decision tree for the match.

This is an adaptation of https://github.com/yorickpeterse/pattern-matching-in-rust/tree/main/jacobs2021 which is an implementation of https://julesjacobs.com/notes/patternmatching/patternmatching.pdf

impl Elaborator<'_>

pub(crate) fn elaborate_expression( &mut self, expr: Expression, ) -> (ExprId, Type)

pub(crate) fn elaborate_expression_with_target_type( &mut self, expr: Expression, target_type: Option<&Type>, ) -> (ExprId, Type)

fn elaborate_interned_statement_as_expr( &mut self, id: InternedStatementKind, location: Location, ) -> (ExprId, Type)

pub(super) fn elaborate_block( &mut self, block: BlockExpression, target_type: Option<&Type>, ) -> (HirExpression, Type)

fn elaborate_block_expression( &mut self, block: BlockExpression, target_type: Option<&Type>, ) -> (HirBlockExpression, Type)

fn elaborate_unsafe_block( &mut self, unsafe_expression: UnsafeExpression, target_type: Option<&Type>, ) -> (HirExpression, Type)

fn elaborate_literal( &mut self, literal: Literal, location: Location, ) -> (HirExpression, Type)

fn integer_suffix_type(&mut self, suffix: Option<IntegerTypeSuffix>) -> Type

fn elaborate_array_literal( &mut self, array_literal: ArrayLiteral, location: Location, is_array: bool, ) -> (HirExpression, Type)

fn elaborate_fmt_string( &mut self, fragments: Vec<FmtStrFragment>, length: u32, ) -> (HirExpression, Type)

fn elaborate_prefix( &mut self, prefix: PrefixExpression, location: Location, ) -> (ExprId, Type)

pub(super) fn check_can_mutate(&mut self, expr_id: ExprId, location: Location)

pub(super) fn check_can_mutate_lambda_capture( &mut self, id: DefinitionId, name: String, location: Location, )

fn elaborate_index( &mut self, index_expr: IndexExpression, ) -> (HirExpression, Type)

fn elaborate_call( &mut self, call: CallExpression, location: Location, ) -> (HirExpression, Type)

fn elaborate_method_call( &mut self, method_call: MethodCallExpression, location: Location, ) -> (HirExpression, Type)

pub(super) fn elaborate_constrain( &mut self, expr: ConstrainExpression, ) -> (HirExpression, Type)

fn elaborate_expression_with_type( &mut self, arg: Expression, typ: Option<&Type>, ) -> (ExprId, Type)

Elaborates an expression knowing that it has to match a given type.

fn check_method_call_visibility( &mut self, func_id: FuncId, object_type: &Type, name: &Ident, )

fn elaborate_constructor( &mut self, constructor: ConstructorExpression, ) -> (HirExpression, Type)

fn elaborate_constructor_with_type( &mut self, typ: Type, fields: Vec<(Ident, Expression)>, location: Location, last_segment: Option<TypedPathSegment>, ) -> (HirExpression, Type)

pub(super) fn mark_struct_as_constructed( &mut self, struct_type: Shared<DataType>, )

fn resolve_constructor_expr_fields( &mut self, struct_type: Shared<DataType>, field_types: Vec<(String, ItemVisibility, Type)>, fields: Vec<(Ident, Expression)>, location: Location, ) -> Vec<(Ident, ExprId)>

Resolve all the fields of a struct constructor expression. Ensures all fields are present, none are repeated, and all are part of the struct.

fn elaborate_member_access( &mut self, access: MemberAccessExpression, location: Location, ) -> (ExprId, Type)

pub fn intern_expr(&mut self, expr: HirExpression, location: Location) -> ExprId

fn elaborate_cast( &mut self, cast: CastExpression, location: Location, ) -> (HirExpression, Type)

fn elaborate_infix( &mut self, infix: InfixExpression, location: Location, ) -> (ExprId, Type)

fn handle_operand_type_rules_result( &mut self, result: Result<(Type, bool), TypeCheckError>, operand_type: &Type, trait_method_id: Option<TraitItemId>, expr_id: ExprId, location: Location, ) -> Type

fn elaborate_if( &mut self, if_expr: IfExpression, target_type: Option<&Type>, ) -> (HirExpression, Type)

fn elaborate_match( &mut self, match_expr: MatchExpression, location: Location, ) -> (HirExpression, Type)

fn wrap_in_let(&mut self, expr_id: ExprId, typ: Type) -> (StmtId, DefinitionId)

fn elaborate_tuple( &mut self, tuple: Vec<Expression>, target_type: Option<&Type>, ) -> (HirExpression, Type)

fn elaborate_lambda_with_target_type( &mut self, lambda: Lambda, target_type: Option<&Type>, ) -> (HirExpression, Type)

fn elaborate_lambda_with_parameter_type_hints( &mut self, lambda: Lambda, parameters_type_hints: Option<&Vec<Type>>, ) -> (HirExpression, Type)

For elaborating a lambda we might get parameters_type_hints. These come from a potential call that has this lambda as the argument. The parameter type hints will be the types of the function type corresponding to the lambda argument.

fn elaborate_quote( &mut self, tokens: Tokens, location: Location, ) -> (HirExpression, Type)

fn elaborate_comptime_block( &mut self, block: BlockExpression, location: Location, target_type: Option<&Type>, ) -> (ExprId, Type)

pub fn inline_comptime_value( &mut self, value: Result<Value, InterpreterError>, location: Location, ) -> (ExprId, Type)

fn try_get_comptime_function( &mut self, func: ExprId, location: Location, ) -> Result<Option<FuncId>, ResolverError>

fn call_macro( &mut self, func: ExprId, arguments: Vec<ExprId>, location: Location, return_type: Type, ) -> Option<(HirExpression, Type)>

Call a macro function and inlines its code at the call site. This will also perform a type check to ensure that the return type is an Expr value.

fn elaborate_as_trait_path(&mut self, path: AsTraitPath) -> (ExprId, Type)

impl Elaborator<'_>

pub(super) fn push_defaultable_type_variable(&mut self, typ: Type)

Push a type variable into the current FunctionContext to be defaulted if needed at the end of the earlier of either the current function or the current comptime scope.

pub(super) fn push_required_type_variable( &mut self, type_variable_id: TypeVariableId, typ: Type, kind: BindableTypeVariableKind, location: Location, )

Push a type variable (its ID and type) as a required type variable: it must be bound after type-checking the current function.

pub(super) fn push_trait_constraint( &mut self, constraint: TraitConstraint, expr_id: ExprId, select_impl: bool, )

Push a trait constraint into the current FunctionContext to be solved if needed at the end of the earlier of either the current function or the current comptime scope.

pub fn push_integer_literal_expr_id(&mut self, literal_expr_id: ExprId)

Push an ExprId that corresponds to an integer literal. At the end of the current function we’ll check that they fit in their type’s range.

fn get_function_context_mut(&mut self) -> &mut FunctionContext

pub(super) fn push_function_context(&mut self)

pub(super) fn check_and_pop_function_context(&mut self)

Defaults all type variables used in this function context then solves all still-unsolved trait constraints in this context.

fn check_defaultable_type_variables(&self, type_variables: Vec<Type>)

fn check_integer_literal_fit_their_type(&mut self, expr_ids: Vec<ExprId>)

fn check_trait_constraints( &mut self, trait_constraints: Vec<(TraitConstraint, ExprId, bool)>, )

fn check_required_type_variables( &mut self, type_variables: Vec<RequiredTypeVariable>, )

impl Elaborator<'_>

pub(super) fn resolve_path_or_error( &mut self, path: TypedPath, target: PathResolutionTarget, ) -> Result<PathResolutionItem, ResolverError>

pub(super) fn use_path_or_error( &mut self, path: TypedPath, target: PathResolutionTarget, ) -> Result<PathResolutionItem, ResolverError>

pub(super) fn resolve_path_or_error_inner( &mut self, path: TypedPath, target: PathResolutionTarget, mode: PathResolutionMode, ) -> Result<PathResolutionItem, ResolverError>

pub(super) fn resolve_path_as_type( &mut self, path: TypedPath, ) -> Result<PathResolution, PathResolutionError>

pub(super) fn use_path_as_type( &mut self, path: TypedPath, ) -> Result<PathResolution, PathResolutionError>

pub(super) fn resolve_path_inner( &mut self, path: TypedPath, target: PathResolutionTarget, mode: PathResolutionMode, ) -> Result<PathResolution, PathResolutionError>

Resolves a path in the current module. If the referenced name can’t be found, Err will be returned. If it can be found, Ok will be returned with a potential list of errors if, for example, one of the segments is not accessible from the current module (e.g. because it’s private).

fn resolve_path_in_module( &mut self, path: TypedPath, importing_module: ModuleId, intermediate_item: IntermediatePathResolutionItem, target: PathResolutionTarget, mode: PathResolutionMode, ) -> Result<PathResolution, PathResolutionError>

Resolves a path in current_module. importing_module is the module where the lookup originally started.

fn resolve_name_in_module( &mut self, path: TypedPath, starting_module: ModuleId, importing_module: ModuleId, intermediate_item: IntermediatePathResolutionItem, target: PathResolutionTarget, mode: PathResolutionMode, ) -> Result<PathResolution, PathResolutionError>

Resolves a Path assuming we are inside starting_module. importing_module is the module where the lookup originally started.

fn per_ns_item_to_path_resolution_item( &mut self, path: TypedPath, importing_module: ModuleId, intermediate_item: IntermediatePathResolutionItem, current_module_id: ModuleId, errors: &mut Vec<PathResolutionError>, module_def_id: ModuleDefId, visibility: ItemVisibility, ) -> PathResolutionItem

fn self_type_module_id(&self) -> Option<ModuleId>

fn resolve_method( &self, importing_module_id: ModuleId, current_module: &ModuleData, ident: &Ident, ) -> MethodLookupResult

fn resolve_primitive_type_or_function( &mut self, path: TypedPath, importing_module_id: ModuleId, ) -> Option<Result<PathResolution, PathResolutionError>>

impl Elaborator<'_>

pub(super) fn elaborate_pattern( &mut self, pattern: Pattern, expected_type: Type, definition_kind: DefinitionKind, warn_if_unused: bool, ) -> HirPattern

pub fn elaborate_pattern_and_store_ids( &mut self, pattern: Pattern, expected_type: Type, definition_kind: DefinitionKind, created_ids: &mut Vec<HirIdent>, warn_if_unused: bool, ) -> HirPattern

Equivalent to elaborate_pattern, this version just also adds any new DefinitionIds that were created to the given Vec.

fn elaborate_pattern_mut( &mut self, pattern: Pattern, expected_type: Type, definition: DefinitionKind, mutable: Option<Location>, new_definitions: &mut Vec<HirIdent>, warn_if_unused: bool, ) -> HirPattern

fn elaborate_struct_pattern( &mut self, name: TypedPath, fields: Vec<(Ident, Pattern)>, location: Location, expected_type: Type, definition: DefinitionKind, mutable: Option<Location>, new_definitions: &mut Vec<HirIdent>, ) -> HirPattern

fn resolve_constructor_pattern_fields( &mut self, struct_type: Shared<DataType>, fields: Vec<(Ident, Pattern)>, location: Location, expected_type: Type, definition: DefinitionKind, mutable: Option<Location>, new_definitions: &mut Vec<HirIdent>, ) -> Vec<(Ident, HirPattern)>

Resolve all the fields of a struct constructor expression. Ensures all fields are present, none are repeated, and all are part of the struct.

pub(super) fn add_variable_decl( &mut self, name: Ident, mutable: bool, allow_shadowing: bool, warn_if_unused: bool, definition: DefinitionKind, ) -> HirIdent

pub fn add_existing_variable_to_scope( &mut self, name: String, ident: HirIdent, warn_if_unused: bool, )

pub fn add_global_variable_decl( &mut self, name: Ident, global_id: GlobalId, ) -> HirIdent

pub(super) fn use_variable( &mut self, name: &Ident, ) -> Result<(HirIdent, usize), ResolverError>

Lookup and use the specified variable. This will increment its use counter by one and return the variable if found. If the variable is not found, an error is returned.

pub(super) fn resolve_function_turbofish_generics( &mut self, func_id: &FuncId, resolved_turbofish: Option<Vec<Located<Type>>>, location: Location, ) -> Option<Vec<Type>>

Resolve generics using the expected kinds of the function we are calling

pub(super) fn resolve_struct_turbofish_generics( &mut self, struct_type: &DataType, generics: Vec<Type>, resolved_turbofish: Option<Vec<Located<Type>>>, location: Location, ) -> Vec<Type>

pub(super) fn resolve_trait_turbofish_generics( &mut self, trait_name: &str, trait_generic_kinds: Vec<Kind>, generics: Vec<Type>, resolved_turbofish: Option<Vec<Located<Type>>>, location: Location, ) -> Vec<Type>

pub(super) fn resolve_alias_turbofish_generics( &mut self, type_alias: &TypeAlias, generics: Vec<Type>, resolved_turbofish: Option<Vec<Located<Type>>>, location: Location, ) -> Vec<Type>

pub(super) fn resolve_item_turbofish_generics( &mut self, item_kind: &'static str, item_name: &str, item_generic_kinds: Vec<Kind>, generics: Vec<Type>, resolved_turbofish: Option<Vec<Located<Type>>>, location: Location, ) -> Vec<Type>

pub(super) fn resolve_turbofish_generics( &mut self, kinds: Vec<Kind>, turbofish_generics: Vec<Located<Type>>, ) -> Vec<Type>

pub(super) fn elaborate_variable(&mut self, variable: Path) -> (ExprId, Type)

fn elaborate_variable_as_self_method_or_associated_constant( &mut self, variable: &TypedPath, ) -> Option<(ExprId, Type)>

Checks whether variable is Self::method_name or Self::AssociatedConstant when we are inside a trait impl and Self resolves to a primitive type.

In the first case we elaborate this as if it were a TypePath (for example, if Self is u32 then we consider this the same as u32::method_name). A regular path lookup won’t work here for the same reason TypePath exists.

In the second case we solve the associated constant by looking up its value, later turning it into a literal.

pub(crate) fn validate_path(&mut self, path: Path) -> TypedPath

fn validate_path_segment(&mut self, segment: PathSegment) -> TypedPathSegment

fn constant_integer( &mut self, numeric_type: Type, value: SignedField, location: Location, ) -> (ExprId, Type)

fn resolve_item_turbofish_and_self_type( &mut self, item: PathResolutionItem, ) -> (Vec<Type>, Option<Type>)

Solve any generics that are part of the path before the function, for example:

foo::Bar::<i32>::baz   

Solve <i32> above

pub(super) fn resolve_struct_id_turbofish_generics( &mut self, struct_id: TypeId, turbofish: Option<Turbofish>, ) -> Vec<Type>

pub(super) fn resolve_type_alias_id_turbofish_generics( &mut self, type_alias_id: TypeAliasId, generics: Option<Turbofish>, ) -> Vec<Type>

fn resolve_variable( &mut self, path: TypedPath, ) -> (HirIdent, Option<PathResolutionItem>)

pub(crate) fn handle_hir_ident( &mut self, hir_ident: &HirIdent, var_scope_index: usize, location: Location, )

pub(crate) fn type_check_variable( &mut self, ident: HirIdent, expr_id: ExprId, generics: Option<Vec<Type>>, ) -> Type

pub(crate) fn type_check_variable_with_bindings( &mut self, ident: HirIdent, expr_id: ExprId, generics: Option<Vec<Type>>, bindings: TypeBindings, push_required_type_variables: bool, ) -> Type

fn instantiate( &mut self, typ: Type, bindings: TypeBindings, turbofish_generics: Option<Vec<Type>>, function_generic_count: usize, location: Location, ) -> (Type, TypeBindings)

pub(crate) fn get_ident_from_path( &mut self, path: TypedPath, ) -> ((HirIdent, usize), Option<PathResolutionItem>)

pub(super) fn elaborate_type_path(&mut self, path: TypePath) -> (ExprId, Type)

fn elaborate_type_path_impl( &mut self, typ: Type, ident: Ident, turbofish: Option<GenericTypeArgs>, typ_location: Location, ) -> (ExprId, Type)

impl Elaborator<'_>

pub(crate) fn instantiate_primitive_type( &mut self, primitive_type: PrimitiveType, args: GenericTypeArgs, location: Location, ) -> Type

pub(crate) fn instantiate_primitive_type_with_turbofish( &mut self, primitive_type: PrimitiveType, turbofish: Option<Turbofish>, ) -> Type

impl Elaborator<'_>

pub fn module_id(&self) -> ModuleId

pub fn replace_module(&mut self, new_module: ModuleId) -> ModuleId

pub(super) fn get_type(&self, type_id: TypeId) -> Shared<DataType>

pub(super) fn get_trait_mut(&mut self, trait_id: TraitId) -> &mut Trait

pub(super) fn resolve_local_variable( &mut self, hir_ident: HirIdent, var_scope_index: usize, )

pub(super) fn lookup_global( &mut self, path: TypedPath, ) -> Result<(DefinitionId, PathResolutionItem), ResolverError>

pub fn push_scope(&mut self)

pub fn pop_scope(&mut self)

pub fn check_for_unused_variables_in_scope_tree( &mut self, scope_decls: ScopeTree<String, ResolverMeta>, )

fn check_for_unused_variables_in_local_scope( decl_map: Scope<String, ResolverMeta>, unused_vars: &mut Vec<HirIdent>, )

pub(crate) fn lookup_trait_or_error( &mut self, path: TypedPath, ) -> Option<&mut Trait>

Lookup a given trait by name/path.

pub(super) fn lookup_type_or_error(&mut self, path: TypedPath) -> Option<Type>

Looks up a given type by name. This will also instantiate any struct types found.

pub(super) fn lookup_type_alias( &mut self, path: TypedPath, mode: PathResolutionMode, ) -> Option<Shared<TypeAlias>>

impl Elaborator<'_>

fn elaborate_statement_value( &mut self, statement: Statement, ) -> (HirStatement, Type)

fn elaborate_statement_value_with_target_type( &mut self, statement: Statement, target_type: Option<&Type>, ) -> (HirStatement, Type)

pub(crate) fn elaborate_statement( &mut self, statement: Statement, ) -> (StmtId, Type)

pub(crate) fn elaborate_statement_with_target_type( &mut self, statement: Statement, target_type: Option<&Type>, ) -> (StmtId, Type)

pub(super) fn elaborate_local_let( &mut self, let_stmt: LetStatement, ) -> (HirStatement, Type)

pub(super) fn elaborate_let( &mut self, let_stmt: LetStatement, global_id: Option<GlobalId>, ) -> (HirStatement, Type)

Elaborate a local or global let statement. If this is a global let, the DefinitionId of the global is specified so that elaborate_pattern can create a Global definition kind with the correct ID instead of a local one with a fresh ID.

pub(super) fn elaborate_assign( &mut self, assign: AssignStatement, ) -> (HirStatement, Type)

pub(super) fn elaborate_for( &mut self, for_loop: ForLoopStatement, ) -> (HirStatement, Type)

pub(super) fn elaborate_loop( &mut self, block: Expression, location: Location, ) -> (HirStatement, Type)

pub(super) fn elaborate_while( &mut self, while_: WhileStatement, ) -> (HirStatement, Type)

fn elaborate_jump( &mut self, is_break: bool, location: Location, ) -> (HirStatement, Type)

fn get_lvalue_error_info( &self, lvalue: &HirLValue, ) -> (DefinitionId, String, Location)

fn elaborate_lvalue( &mut self, lvalue: LValue, ) -> (HirLValue, Type, bool, Vec<StmtId>)

Elaborates an lvalue returning:

The HirLValue equivalent of the given lvalue

The type being assigned to

Whether the underlying variable is mutable

A vector of new statements which need to prefix the resulting assign statement. This hoists out any sub-expressions to simplify sequencing of side-effects.

fn fresh_definition_for_lvalue_index( &mut self, expr: ExprId, typ: Type, location: Location, ) -> Option<(StmtId, ExprId)>

pub(super) fn check_field_access( &mut self, lhs_type: &Type, field_name: &str, location: Location, dereference_lhs: Option<impl FnMut(&mut Self, Type, Type)>, ) -> Option<(Type, usize)>

Type checks a field access, adding dereference operators as necessary

pub(super) fn check_struct_field_visibility( &mut self, struct_type: &DataType, field_name: &str, visibility: ItemVisibility, location: Location, )

fn elaborate_comptime_statement( &mut self, statement: Statement, ) -> (HirStatement, Type)

impl Elaborator<'_>

pub(super) fn collect_trait_impl_methods( &mut self, trait_id: TraitId, trait_impl: &mut UnresolvedTraitImpl, trait_impl_where_clause: &[TraitConstraint], )

fn check_where_clause_against_trait( &mut self, func_id: &FuncId, method: &TraitFunction, trait_impl_where_clause: &[TraitConstraint], trait_impl_generics: &[Type], trait_id: TraitId, impl_id: TraitImplId, )

Issue an error if the impl is stricter than the trait.

Example

trait MyTrait { }
trait Foo<T> {
    fn foo<U>();
}
impl<A> Foo<A> for () {
    // Error issued here as `foo` does not have the `MyTrait` constraint
    fn foo<B>() where B: MyTrait {}
}

fn check_trait_impl_crate_coherence( &mut self, trait_id: TraitId, trait_impl: &UnresolvedTraitImpl, )

pub(super) fn take_unresolved_associated_types( &mut self, trait_impl: &mut UnresolvedTraitImpl, ) -> Vec<(Ident, UnresolvedType, Kind)>

impl Elaborator<'_>

pub fn collect_traits( &mut self, traits: &mut BTreeMap<TraitId, UnresolvedTrait>, )

pub fn collect_trait_methods( &mut self, traits: &mut BTreeMap<TraitId, UnresolvedTrait>, )

fn resolve_trait_bounds( &mut self, bounds: &[TraitBound], ) -> Vec<ResolvedTraitBound>

fn resolve_trait_methods( &mut self, trait_id: TraitId, unresolved_trait: &UnresolvedTrait, ) -> Vec<TraitFunction>

pub fn resolve_trait_function( &mut self, trait_id: TraitId, name: &Ident, is_unconstrained: bool, generics: &UnresolvedGenerics, parameters: &[(Ident, UnresolvedType)], return_type: &FunctionReturnType, where_clause: Vec<UnresolvedTraitConstraint>, body: &Option<BlockExpression>, trait_visibility: ItemVisibility, func_id: FuncId, )

impl Elaborator<'_>

pub(crate) fn resolve_type(&mut self, typ: UnresolvedType) -> Type

pub(crate) fn use_type(&mut self, typ: UnresolvedType) -> Type

pub(crate) fn use_type_with_kind( &mut self, typ: UnresolvedType, kind: &Kind, ) -> Type

fn resolve_type_inner( &mut self, typ: UnresolvedType, kind: &Kind, mode: PathResolutionMode, ) -> Type

Translates an UnresolvedType to a Type with a TypeKind::Normal

pub(crate) fn resolve_type_with_kind( &mut self, typ: UnresolvedType, kind: &Kind, ) -> Type

fn resolve_type_with_kind_inner( &mut self, typ: UnresolvedType, kind: &Kind, mode: PathResolutionMode, ) -> Type

Translates an UnresolvedType into a Type and appends any freshly created TypeVariables created to new_variables.

pub fn find_generic(&self, target_name: &str) -> Option<&ResolvedGeneric>

fn lookup_associated_type_on_self(&self, path: &TypedPath) -> Option<Type>

fn resolve_named_type( &mut self, path: TypedPath, args: GenericTypeArgs, mode: PathResolutionMode, ) -> Type

fn lookup_type_variable( &mut self, path: &TypedPath, args: &GenericTypeArgs, ) -> Option<Type>

fn resolve_trait_as_type( &mut self, path: TypedPath, args: GenericTypeArgs, mode: PathResolutionMode, ) -> Type

pub(super) fn resolve_trait_args_from_trait_impl( &mut self, args: GenericTypeArgs, item: TraitId, location: Location, ) -> (Vec<Type>, Vec<NamedType>)

Identical to resolve_type_args but does not allow associated types to be elided since trait impls must specify them.

pub(super) fn use_type_args( &mut self, args: GenericTypeArgs, item: impl Generic, location: Location, ) -> (Vec<Type>, Vec<NamedType>)

pub(super) fn resolve_type_args_inner( &mut self, args: GenericTypeArgs, item: impl Generic, location: Location, mode: PathResolutionMode, ) -> (Vec<Type>, Vec<NamedType>)

pub(super) fn resolve_type_or_trait_args_inner( &mut self, args: GenericTypeArgs, item: impl Generic, location: Location, allow_implicit_named_args: bool, mode: PathResolutionMode, ) -> (Vec<Type>, Vec<NamedType>)

fn resolve_associated_type_args( &mut self, args: Vec<(Ident, UnresolvedType)>, item: impl Generic, location: Location, allow_implicit_named_args: bool, mode: PathResolutionMode, ) -> Vec<NamedType>

fn lookup_generic_or_global_type( &mut self, path: &TypedPath, mode: PathResolutionMode, ) -> Option<Type>

pub(super) fn convert_expression_type( &mut self, length: UnresolvedTypeExpression, expected_kind: &Kind, location: Location, ) -> Type

pub(super) fn check_kind( &mut self, typ: Type, expected_kind: &Kind, location: Location, ) -> Type

fn resolve_as_trait_path(&mut self, path: AsTraitPath) -> Type

fn get_associated_type_from_trait_impl( &mut self, path: AsTraitPath, impl_kind: TraitImplKind, ) -> Type

fn resolve_trait_static_method_by_self( &mut self, path: &TypedPath, ) -> Option<TraitPathResolution>

fn resolve_trait_static_method( &mut self, path: &TypedPath, ) -> Option<TraitPathResolution>

fn resolve_trait_method_by_named_generic( &mut self, path: &TypedPath, ) -> Option<TraitPathResolution>

fn resolve_type_trait_method( &mut self, path: &TypedPath, ) -> Option<TraitPathResolution>

This resolves a method in the form Type::method where method is a trait method

pub(super) fn resolve_trait_generic_path( &mut self, path: &TypedPath, ) -> Option<TraitPathResolution>

pub(super) fn unify( &mut self, actual: &Type, expected: &Type, make_error: impl FnOnce() -> TypeCheckError, )

pub fn unify_without_applying_bindings( &mut self, actual: &Type, expected: &Type, make_error: impl FnOnce() -> TypeCheckError, )

Do not apply type bindings even after a successful unification. This function is used by the interpreter for some comptime code which can change types e.g. on each iteration of a for loop.

pub(super) fn unify_with_coercions( &mut self, actual: &Type, expected: &Type, expression: ExprId, location: Location, make_error: impl FnOnce() -> CompilationError, )

Wrapper of Type::unify_with_coercions using self.errors

pub(super) fn polymorphic_integer_or_field(&mut self) -> Type

Return a fresh integer or field type variable and log it in self.type_variables to default it later.

pub(super) fn polymorphic_integer(&mut self) -> Type

Return a fresh integer type variable and log it in self.type_variables to default it later.

pub(super) fn type_variable_with_kind(&mut self, type_var_kind: Kind) -> Type

Return a fresh integer type variable and log it in self.type_variables to default it later.

pub(super) fn resolve_inferred_type(&mut self, typ: UnresolvedType) -> Type

Translates a (possibly Unspecified) UnresolvedType to a Type. Any UnresolvedType::Unspecified encountered are replaced with fresh type variables.

pub(super) fn insert_auto_dereferences( &mut self, object: ExprId, typ: Type, ) -> (ExprId, Type)

Insert as many dereference operations as necessary to automatically dereference a method call object to its base value type T.

fn try_remove_implicit_dereference(&mut self, object: ExprId) -> Option<ExprId>

Given a method object: (*foo).bar of a method call (*foo).bar.baz(), remove the implicitly added dereference operator if one is found.

Returns Some(new_expr_id) if a dereference was removed and None otherwise.

fn bind_function_type_impl( &mut self, fn_params: &[Type], fn_ret: &Type, callsite_args: &[(Type, ExprId, Location)], location: Location, ) -> Type

pub(super) fn bind_function_type( &mut self, function: Type, args: Vec<(Type, ExprId, Location)>, location: Location, ) -> Type

pub(super) fn check_cast( &mut self, from_expr_id: &ExprId, from: &Type, to: &Type, location: Location, ) -> Type

fn comparator_operand_type_rules( &mut self, lhs_type: &Type, rhs_type: &Type, op: &HirBinaryOp, location: Location, ) -> Result<(Type, bool), TypeCheckError>

fn bind_type_variables_for_infix( &mut self, lhs_type: &Type, op: &HirBinaryOp, rhs_type: &Type, location: Location, ) -> bool

Handles the TypeVariable case for checking binary operators. Returns true if we should use the impl for the operator instead of the primitive version of it.

pub(super) fn infix_operand_type_rules( &mut self, lhs_type: &Type, op: &HirBinaryOp, rhs_type: &Type, location: Location, ) -> Result<(Type, bool), TypeCheckError>

pub(super) fn prefix_operand_type_rules( &mut self, op: &UnaryOp, rhs_type: &Type, location: Location, ) -> Result<(Type, bool), TypeCheckError>

pub(super) fn type_check_operator_method( &mut self, expr_id: ExprId, trait_method_id: TraitItemId, object_type: &Type, location: Location, )

Prerequisite: verify_trait_constraint of the operator’s trait constraint.

Although by this point the operator is expected to already have a trait impl, we still need to match the operator’s type against the method’s instantiated type to ensure the instantiation bindings are correct and the monomorphizer can re-apply the needed bindings.

pub(super) fn type_check_member_access( &mut self, access: HirMemberAccess, expr_id: ExprId, lhs_type: Type, location: Location, ) -> Type

pub(crate) fn lookup_method( &mut self, object_type: &Type, method_name: &str, location: Location, object_location: Location, check_self_param: bool, ) -> Option<HirMethodReference>

fn lookup_type_or_primitive_method( &mut self, object_type: &Type, method_name: &str, location: Location, object_location: Location, check_self_param: bool, ) -> Option<HirMethodReference>

fn return_trait_method_in_scope( &mut self, trait_methods: &[(FuncId, TraitId)], method_name: &str, location: Location, ) -> Option<HirMethodReference>

Given a list of functions and the trait they belong to, returns the one function that is in scope.

fn get_trait_method_in_scope( &mut self, trait_methods: &[(FuncId, TraitId)], method_name: &str, location: Location, ) -> (Option<HirMethodReference>, Option<PathResolutionError>)

fn trait_hir_method_reference( &self, trait_id: TraitId, trait_methods: &[(FuncId, TraitId)], method_name: &str, location: Location, ) -> HirMethodReference

fn lookup_method_in_trait_constraints( &mut self, object_type: &Type, method_name: &str, location: Location, object_location: Location, ) -> Option<HirMethodReference>

fn lookup_method_in_trait( &self, the_trait: &Trait, method_name: &str, trait_bound: &ResolvedTraitBound, starting_trait_id: TraitId, ) -> Option<HirMethodReference>

pub(super) fn type_check_call( &mut self, call: &HirCallExpression, func_type: Type, args: Vec<(Type, ExprId, Location)>, location: Location, ) -> Type

fn is_unconstrained_call(&self, expr: ExprId) -> bool

pub(super) fn try_add_mutable_reference_to_object( &mut self, function_type: &Type, object_type: &mut Type, object: &mut ExprId, )

Check if the given method type requires a mutable reference to the object type, and check if the given object type is already a mutable reference. If not, add one. This is used to automatically transform a method call: foo.bar() into a function call: bar(&mut foo).

A notable corner case of this function is where it interacts with auto-deref of .. If a field is being mutated e.g. foo.bar.mutate_bar() where foo: &mut Foo, the compiler will insert a dereference before bar (*foo).bar.mutate_bar() which would cause us to mutate a copy of bar rather than a reference to it. We must check for this corner case here and remove the implicitly added dereference operator if we find one.

pub fn type_check_function_body( &mut self, body_type: Type, meta: &FuncMeta, body_id: ExprId, )

fn function_info(&self, function_body_id: ExprId) -> (Location, bool)

pub fn verify_trait_constraint( &mut self, object_type: &Type, trait_id: TraitId, trait_generics: &[Type], associated_types: &[NamedType], function_ident_id: ExprId, select_impl: bool, location: Location, )

pub(super) fn push_trait_constraint_error( &mut self, object_type: &Type, error: ImplSearchErrorKind, location: Location, )

pub fn add_existing_generics( &mut self, unresolved_generics: &UnresolvedGenerics, generics: &Generics, )

pub fn add_existing_generic( &mut self, unresolved_generic: &UnresolvedGeneric, location: Location, resolved_generic: &ResolvedGeneric, )

pub fn bind_generics_from_trait_constraint( &self, constraint: &TraitConstraint, assumed: bool, bindings: &mut TypeBindings, )

pub fn bind_generics_from_trait_bound( &self, trait_bound: &ResolvedTraitBound, bindings: &mut TypeBindings, )

pub fn instantiate_parent_trait_bound( &self, trait_bound: &ResolvedTraitBound, parent_trait_bound: &ResolvedTraitBound, ) -> ResolvedTraitBound

pub(crate) fn fully_qualified_trait_path(&self, trait_: &Trait) -> String

impl Elaborator<'_>

pub fn find_unquoted_exprs_tokens(&mut self, tokens: Tokens) -> Tokens

Go through the given tokens looking for a ‘$’ token followed by a variable to unquote. Each time these two tokens are found, they are replaced by a new UnquoteMarker token containing the ExprId of the resolved variable to unquote.

impl<'context> Elaborator<'context>

pub fn new( interner: &'context mut NodeInterner, def_maps: &'context mut DefMaps, usage_tracker: &'context mut UsageTracker, crate_graph: &'context CrateGraph, interpreter_output: &'context Option<Rc<RefCell<dyn Write>>>, crate_id: CrateId, interpreter_call_stack: Vector<Location>, options: GenericOptions<'context, FileId>, elaborate_reasons: Vector<ElaborateReason>, ) -> Self

pub fn from_context( context: &'context mut Context<'_, '_>, crate_id: CrateId, options: GenericOptions<'context, FileId>, ) -> Self

pub fn elaborate( context: &'context mut Context<'_, '_>, crate_id: CrateId, items: CollectedItems, options: GenericOptions<'context, FileId>, ) -> Vec<CompilationError>

pub fn elaborate_and_return_self( context: &'context mut Context<'_, '_>, crate_id: CrateId, items: CollectedItems, options: GenericOptions<'context, FileId>, ) -> Self

pub(crate) fn elaborate_items(&mut self, items: CollectedItems)

pub fn pedantic_solving(&self) -> bool

True if we should use pedantic ACVM solving

fn recover_generics<T>(&mut self, f: impl FnOnce(&mut Self) -> T) -> T

Runs f and if it modifies self.generics, self.generics is truncated back to the previous length.

fn elaborate_functions(&mut self, functions: UnresolvedFunctions)

fn introduce_generics_into_scope(&mut self, all_generics: Vec<ResolvedGeneric>)

pub(crate) fn elaborate_function(&mut self, id: FuncId)

fn desugar_impl_trait_arg( &mut self, trait_path: Path, trait_generics: GenericTypeArgs, generics: &mut Vec<TypeVariable>, trait_constraints: &mut Vec<TraitConstraint>, ) -> Type

This turns function parameters of the form: fn foo(x: impl Bar)

into fn foo<T0_impl_Bar>(x: T0_impl_Bar) where T0_impl_Bar: Bar although the fresh type variable is not named internally.

pub fn add_generics(&mut self, generics: &UnresolvedGenerics) -> Generics

Add the given generics to scope. Each generic will have a fresh Shared<TypeBinding> associated with it.

fn resolve_generic( &mut self, generic: &UnresolvedGeneric, ) -> Result<(TypeVariable, Rc<String>), ResolverError>

pub(crate) fn resolve_generic_kind( &mut self, generic: &UnresolvedGeneric, ) -> Kind

Return the kind of an unresolved generic. If a numeric generic has been specified, resolve the annotated type to make sure only primitive numeric types are being used.

pub(crate) fn push_err(&mut self, error: impl Into<CompilationError>)

pub(crate) fn push_errors( &mut self, errors: impl IntoIterator<Item = CompilationError>, )

fn run_lint( &mut self, lint: impl Fn(&Elaborator<'_>) -> Option<CompilationError>, )

pub(crate) fn resolve_module_by_path( &mut self, path: TypedPath, ) -> Option<ModuleId>

fn resolve_trait_by_path(&mut self, path: TypedPath) -> Option<TraitId>

fn resolve_trait_constraints( &mut self, where_clause: &[UnresolvedTraitConstraint], ) -> Vec<TraitConstraint>

Resolve the given trait constraints and add them to scope as we go. This second step is necessary to resolve subsequent constraints such as <T as Foo>::Bar: Eq which may lookup an impl which was assumed by a previous constraint.

If these constraints are unwanted afterward they should be manually removed from the interner.

fn desugar_trait_constraints( &mut self, where_clause: &mut [UnresolvedTraitConstraint], ) -> Vec<(ResolvedGeneric, Vec<ResolvedTraitBound>)>

Expands any traits in a where clause to mention all associated types if they were elided by the user. See add_missing_named_generics for more detail.

Returns all newly created generics to be added to this function/trait/impl.

fn add_missing_named_generics( &mut self, object: &UnresolvedType, bound: &mut TraitBound, ) -> Vec<(ResolvedGeneric, Vec<ResolvedTraitBound>)>

For each associated type that isn’t mentioned in a trait bound, this adds the type as an implicit generic to the where clause and returns the newly created generics in a vector to add to the function/trait/impl later. For example, this will turn a function using a trait with 2 associated types:

fn foo<T>() where T: Foo { ... }

into: fn foo<T>() where T: Foo<Bar = A, Baz = B> { ... }

with a vector of <A, B> returned so that the caller can then modify the function to: fn foo<T, A, B>() where T: Foo<Bar = A, Baz = B> { ... }

fn resolve_trait_constraint( &mut self, constraint: &UnresolvedTraitConstraint, ) -> Option<TraitConstraint>

Resolves a trait constraint and adds it to scope as an assumed impl. This second step is necessary to resolve subsequent constraints such as <T as Foo>::Bar: Eq which may lookup an impl which was assumed by a previous constraint.

pub fn resolve_trait_bound( &mut self, bound: &TraitBound, ) -> Option<ResolvedTraitBound>

pub fn use_trait_bound( &mut self, bound: &TraitBound, ) -> Option<ResolvedTraitBound>

fn resolve_trait_bound_inner( &mut self, bound: &TraitBound, mode: PathResolutionMode, ) -> Option<ResolvedTraitBound>

fn define_function_meta( &mut self, func: &mut NoirFunction, func_id: FuncId, trait_id: Option<TraitId>, extra_trait_constraints: &[(TraitConstraint, Location)], )

Extract metadata from a NoirFunction to be used in analysis and intern the function parameters Prerequisite: any implicit generics, including any generics from the impl, have already been added to scope via self.add_generics.

fn mark_type_as_used(&mut self, typ: &Type)

fn run_function_lints(&mut self, func: &FuncMeta, modifiers: &FunctionModifiers)

fn check_if_type_is_valid_for_program_input( &mut self, typ: &Type, is_entry_point: bool, has_inline_attribute: bool, location: Location, )

Only sized types are valid to be used as main’s parameters or the parameters to a contract function. If the given type is not sized (e.g. contains a slice or NamedGeneric type), an error is issued.

fn pub_allowed(&self, func: &NoirFunction, in_contract: bool) -> bool

True if the pub keyword is allowed on parameters in this function pub on function parameters is only allowed for entry point functions

fn in_contract(&self) -> bool

Returns true if the current module is a contract.

This is usually determined by self.module_id(), but it can be overridden for impls. Impls are an odd case since the methods within resolve as if they’re in the parent module, but should be placed in a child module. Since they should be within a child module, they should be elaborated as if in_contract is false so we can still resolve them in the parent module without them being in a contract.

pub(crate) fn module_is_contract(&self, module_id: ModuleId) -> bool

fn is_entry_point_function( &self, func: &NoirFunction, in_contract: bool, ) -> bool

fn add_trait_constraints_to_scope<'a>( &mut self, constraints: impl Iterator<Item = &'a TraitConstraint>, location: Location, )

fn remove_trait_constraints_from_scope<'a>( &mut self, constraints: impl Iterator<Item = &'a TraitConstraint>, )

fn add_trait_bound_to_scope( &mut self, location: Location, object: &Type, trait_bound: &ResolvedTraitBound, starting_trait_id: TraitId, )

fn elaborate_traits(&mut self, traits: BTreeMap<TraitId, UnresolvedTrait>)

fn elaborate_impls( &mut self, impls: Vec<(UnresolvedGenerics, Location, UnresolvedFunctions)>, )

fn elaborate_trait_impl(&mut self, trait_impl: UnresolvedTraitImpl)

fn add_trait_impl_assumed_trait_implementations( &mut self, impl_id: Option<TraitImplId>, )

fn remove_trait_impl_assumed_trait_implementations( &mut self, impl_id: Option<TraitImplId>, )

fn check_trait_impl_where_clause_matches_trait_where_clause( &mut self, trait_impl: &UnresolvedTraitImpl, )

fn check_parent_traits_are_implemented( &mut self, trait_impl: &UnresolvedTraitImpl, )

fn collect_impls( &mut self, module: LocalModuleId, impls: &mut [(UnresolvedGenerics, Location, UnresolvedFunctions)], self_type: &UnresolvedType, )

fn collect_trait_impl(&mut self, trait_impl: &mut UnresolvedTraitImpl)

pub fn get_module(&self, module: ModuleId) -> &ModuleData

fn get_module_mut(def_maps: &mut DefMaps, module: ModuleId) -> &mut ModuleData

fn declare_methods_on_struct( &mut self, trait_id: Option<TraitId>, functions: &mut UnresolvedFunctions, location: Location, )

fn declare_methods(&mut self, self_type: &Type, function_ids: &[FuncId])

fn define_type_alias( &mut self, alias_id: TypeAliasId, alias: UnresolvedTypeAlias, )

fn find_struct_visibility( &self, struct_type: &DataType, ) -> Option<ItemVisibility>

Find the struct in the parent module so we can know its visibility

fn should_check_function_visibility( &self, func_meta: &FuncMeta, modifiers: &FunctionModifiers, ) -> bool

Check whether a functions return value and args should be checked for private type visibility.

fn check_type_is_not_more_private_then_item( &mut self, name: &Ident, visibility: ItemVisibility, typ: &Type, location: Location, )

Check that an item such as a struct field or type alias is not more visible than the type it refers to.

fn collect_struct_definitions( &mut self, structs: &BTreeMap<TypeId, UnresolvedStruct>, )

pub fn resolve_struct_fields( &mut self, unresolved: &NoirStruct, struct_id: TypeId, ) -> Vec<StructField>

fn collect_enum_definitions(&mut self, enums: &BTreeMap<TypeId, UnresolvedEnum>)

fn elaborate_global(&mut self, global: UnresolvedGlobal)

fn elaborate_comptime_global(&mut self, global_id: GlobalId)

fn elaborate_global_if_unresolved(&mut self, global_id: &GlobalId) -> bool

If the given global is unresolved, elaborate it and return true

fn define_function_metas( &mut self, functions: &mut [UnresolvedFunctions], impls: &mut FxHashMap<(UnresolvedType, LocalModuleId), Vec<(UnresolvedGenerics, Location, UnresolvedFunctions)>>, trait_impls: &mut [UnresolvedTraitImpl], )

fn define_function_metas_for_functions( &mut self, function_set: &mut UnresolvedFunctions, extra_constraints: &[(TraitConstraint, Location)], )

fn in_constrained_function(&self) -> bool

True if we’re currently within a constrained function. Defaults to true if the current function is unknown.

fn check_generics_appear_in_type( &mut self, generics: &[UnresolvedGeneric], self_type: &UnresolvedType, )

Check that all the generics show up in self_type (if they don’t, we produce an error)

pub fn use_unstable_feature( &mut self, feature: UnstableFeature, location: Location, )

Register a use of the given unstable feature. Errors if the feature has not been explicitly enabled in this package.

pub fn errors_occurred_in<T>( &mut self, f: impl FnOnce(&mut Self) -> T, ) -> (bool, T)

Run the given function using the resolver and return true if any errors (not warnings) occurred while running it.