noirc_frontend::hir
Module printer

Modules

items

This module is responsible for building a list of items that represent the final code that is going to be monomorphized and turned into SSA. This final code has all macros expanded and is mainly gathered from data inside NodeInterner, modules in DefMaps and function bodies from HirExpressions.

Structs

ItemPrinter 🔒

Functions

crate_to_module

display_crate

Returns the HIR as human-readable code for the given crate. At this point it is expected that all macros and comptime blocks have been expanded.