noirc_frontend
Module monomorphization

Coming after type checking, monomorphization is the last pass in Noir’s frontend. It accepts the type checked HIR as input and produces a monomorphized AST as output. This file implements the pass itself, while the AST is defined in the ast module.

Unlike the HIR, which is stored within the NodeInterner, the monomorphized AST is self-contained and does not need an external context struct. As a result, the NodeInterner can be safely discarded after monomorphization.

The entry point to this pass is the monomorphize function which, starting from a given function, will monomorphize the entire reachable program.

Modules

ast

debug 🔒

debug_types

errors

printer

This module implements printing of the monomorphized AST, for debugging purposes.

proxies

Implement a post-monomorphization pass where builtin/intrinsic/oracle functions used as values are wrapped in a proxy function which replaces them as a value and forwards calls to them.

visitor

Structs

LambdaContext 🔒

Monomorphizer

The context struct for the monomorphization pass.

Enums

TraitItem 🔒

Functions

append_printable_type_info_for_type

Extend the arguments to print (which is a bool to show if newline is needed and value to be printed itself) with a JSON serialized PrintableType to describe the value, and another bool to show if the print is using a format string, or a raw value.

append_printable_type_info_inner 🔒

bind_trait_impl_func_generics_to_trait_func_generics 🔒

Binds direct generics on a trait impl function to those on a corresponding trait function.

monomorphize

Starting from the given main function, monomorphize the entire program, replacing all references to type variables and NamedGenerics with concrete types, duplicating definitions as necessary to do so.

monomorphize_debug

perform_impl_bindings

Call sites are instantiated against the trait method, but when an impl is later selected, the corresponding method in the impl will have a different set of generics. perform_impl_bindings is needed to apply the generics from the trait method to the impl method. Without this, static method references to generic impls (e.g. Eq::eq for [T; N]) will fail to re-apply the correct type bindings during monomorphization.

perform_instantiation_bindings

resolve_trait_item 🔒

resolve_trait_item_impl 🔒

Resolve a trait item to a particular impl

undo_instantiation_bindings

unwrap_enum_type 🔒

unwrap_struct_type 🔒

unwrap_tuple_type 🔒

Type Aliases

Functions 🔒

Using nested HashMaps here lets us avoid cloning HirTypes when calling .get()

HirType 🔒