noirc_frontend::monomorphization
Module ast

Structs

ArrayLiteral

Assign

Binary

BinaryStatement

Call

Cast

For

FuncId

A function ID corresponds directly to an index of Program::functions

Function

GlobalId

A function ID corresponds directly to an index of Program::globals

Ident

IdentId

Each identifier is given a unique ID to distinguish different uses of identifiers. This is used, for example, in last use analysis to determine which identifiers represent the last use of their definition and can thus be moved instead of cloned.

If

Index

Let

Rather than a Pattern containing possibly several variables, Let now defines a single variable with the given LocalId. By the time this is produced in monomorphization, let-statements with tuple and struct patterns:

LocalId

ID of a local definition, e.g. from a let binding or function parameter that should be compiled before it is referenced.

Match

MatchCase

Program

Unary

While

Enums

Definition

A definition is either a local (variable), function, or is a built-in function that will be generated or referenced by the compiler later.

Expression

The monomorphized AST is expression-based, all statements are also folded into this expression enum. Compared to the HIR, the monomorphized AST has several differences:

InlineType

Represents how an Acir function should be inlined. This type is only relevant for ACIR functions as we do not inline any Brillig functions

LValue

Represents an Ast form that can be assigned to

Literal

Type

Compared to hir_def::types::Type, this monomorphized Type has:

Type Aliases

BinaryOp

Parameters