noirc_frontend
Module node_interner

Modules

dependency 🔒

function 🔒

globals 🔒

ids 🔒

methods 🔒

pusher

Helper constructs to ensure that no parts are forgotten when pushing expressions and statements piecemeal.

reexports 🔒

trait_impl 🔒

Structs

DefinitionId

DefinitionInfo

ExprId

FuncId

FunctionModifiers

All the information from a function that is filled out during definition collection rather than name resolution. As a result, if information about a function is needed during name resolution, this is the only place where it is safe to retrieve it (where all fields are guaranteed to be initialized).

GlobalId

An ID for a global value

ImplMethod

InternedExpressionKind

The ID of an crate::ast::ExpressionKind that’s been pushed into the NodeInterner.

InternedPattern

InternedStatementKind

The ID of a crate::ast::StatementKind that’s been pushed into the NodeInterner.

InternedUnresolvedTypeData

The ID of a crate::ast::UnresolvedTypeData that’s been pushed into the NodeInterner.

Methods

Represents the methods on a given type that each share the same name.

ModuleAttributes

NodeInterner

The node interner is the central storage location of all nodes in Noir’s Hir (the various node types can be found in hir_def). The interner is also used to collect extra information about the Hir, such as the type of each node, information about each definition or struct, etc. Because it is used on the Hir, the NodeInterner is useful in passes where the Hir is used - name resolution, type checking, and monomorphization - and it is not useful afterward.

QuotedTypeId

Reexport

Captures a reexport that happens inside a module. For example:

StmtId

TraitAssociatedTypeId

TraitId

TraitImplId

TraitItemId

TypeAliasId

TypeId

A TypeId wraps a ModuleId, because types are represented by an anonymous module.

Enums

DefinitionKind

DependencyId

A dependency in the dependency graph may be a type or a definition. Types can depend on definitions too. E.g. Foo depends on COUNT in:

GlobalValue

ImplSearchErrorKind

When searching for a trait impl, these are the types of errors we can expect

Node 🔒

A Definition enum specifies anything that we can intern in the NodeInterner We use one Arena for all types that can be interned as that has better cache locality This data structure is never accessed directly, so API wise there is no difference between using Multiple arenas and a single Arena

ReferenceId

A reference to a module, struct, trait, etc., mainly used by the LSP code to keep track of how symbols reference each other.

TraitImplKind

A trait implementation is either a normal implementation that is present in the source program via an impl block, or it is assumed to exist from a where clause or similar.

TypeMethodKey 🔒

These are the primitive type variants that we support adding methods to

Statics

TOP_LEVEL_MODULE_ATTRIBUTES 🔒

Functions

get_type_method_key 🔒

Type Aliases

TypeAttributes 🔒