noirc_frontend::ownership

Module last_uses

This module contains the last use analysis pass which is run on each function before the ownership pass.

The purpose of this pass is to find which instance of a variable is the variable’s last use. Note that a variable may have multiple last uses. This can happen if the variable’s last use is within an if expression or similar. It could be last used in one place in the then branch and in another place in the else branch as long as no code after the if expression uses the same variable.

This pass works by tracking the last use location for each variable with a “loop index” and a Branches enumeration.

The loop index tracks the loop level the variable was declared in. A variable must be cloned in any loop with an index greater than the variable’s own loop index. Note that “loop index” refers to how nested we are within loops. A function starts in index 0, and when we enter the body of a loop {} or for _ in a..b {}, the index increments by 1. So b in loop { for _ in 0..1 { b } } would have index 2.

The Branches enumeration holds each current last use of a variable.

In the common case this will be Branches::Direct(ident_id) indicating that ident_id is the last use of its variable and it was not moved into an if or match.

When the variable is used within an if or match its last use will have a value of Branches::IfOrMatch(cases) with the given nested last uses in each case of the if/match.

This pass is not sophisticated with regard to struct and tuple fields. It currently ignores these entirely and counts each use as a use of the entire variable. This is an area for future optimization. E.g. the program a.b.c; a.e.f will result in a being cloned in its entirety in the first statement. Note that this is lessened in the overall ownership pass such that only .c is cloned but it is still an area for improvement.

Structs

BranchId 🔒

The ID for a single branch of an if or match.

IfOrMatchId 🔒

The ID of an if or match, used to distinguish multiple sequential ifs/matches from each other.

LastUseContext 🔒

Enums

Branches 🔒

A variable’s last use may be split into several branches. E.g:

Type Aliases

BranchesPath 🔒

A single path through a Branches enum.