noirc_frontend::ownership::last_uses
Struct LastUseContext

struct LastUseContext {
    current_loop_and_branch: Vec<Vec<(IfOrMatchId, BranchId)>>,
    next_id: u32,
    last_uses: FxHashMap<LocalId, (usize, Branches)>,
}

Fields
current_loop_and_branch: Vec<Vec<(IfOrMatchId, BranchId)>>

The outer Vec is each loop we’re currently in, while the BranchPath contains the path to overwrite the last use in any Branches enums of the variables we find. As a whole, this tracks the current control-flow of the function we’re in.

next_id: u32last_uses: FxHashMap<LocalId, (usize, Branches)>

Stores the location of each variable’s last use

Map from each local variable to the last instance of that variable. Separate uses of the same variable are differentiated by that identifier’s IdentId which is always different on separate identifiers, unlike the LocalId which is the same for any identifier referring to the same underlying definition.

Each definition is mapped to a loop index and a Branches enumeration.

The loop index tracks how many loops the variable was declared within. It may be moved within the same loop but cannot be moved within a nested loop. E.g:

fn foo() {
    let x = 2;
    for _ in 0 .. 2 {
        let b = true;
        println((x, b));
    }
}

In the snippet above, x will have loop index 0 which does not match its last use within the for loop (1 loop deep = loop index of 1). However, b has loop index 1 and thus can be moved into its last use in the loop, in this case the println call.

The Branches enumeration holds each last use of the variable. This is usually only one use but can be multiple if the last use is spread across several if or match branches. E.g:

fn bar() {
    let x = 2;
    if true {
        println(x);
    } else {
        assert(x < 5);
    }
}

x above has two last uses, one in each if branch.

Implementations

impl LastUseContext

fn push_loop_scope(&mut self)

fn pop_loop_scope(&mut self)

fn branch(&mut self, id: IfOrMatchId, branch_id: u32)

fn next_if_or_match_id(&mut self) -> IfOrMatchId

fn unbranch(&mut self)

fn loop_index(&self) -> usize

fn declare_variable(&mut self, id: LocalId)

fn remember_use_of_variable(&mut self, id: LocalId, variable: IdentId)

Remember a new use of the given variable, possibly overwriting or adding to the previous last use depending on the current position in if/match branches or the loop index.

If the loop index is equal to the variable’s when it was defined we can overwrite the last use, but if it is greater we have to set the last use to None. This is because variable’s cannot be moved within loops unless it was defined within the same loop.

fn remember_use_of_variable_rec( branches: &mut Branches, path: &[(IfOrMatchId, BranchId)], variable: IdentId, )

fn get_variables_to_move(self) -> HashMap<LocalId, Vec<IdentId>>

fn track_variables_in_expression(&mut self, expr: &Expression)

fn track_variables_in_ident(&mut self, ident: &Ident)

fn track_variables_in_literal(&mut self, literal: &Literal)

fn track_variables_in_unary(&mut self, unary: &Unary)

fn track_variables_in_binary(&mut self, binary: &Binary)

fn track_variables_in_index(&mut self, index: &Index)

fn track_variables_in_cast(&mut self, cast: &Cast)

fn track_variables_in_for(&mut self, for_expr: &For)

fn track_variables_in_while(&mut self, while_expr: &While)

fn track_variables_in_loop(&mut self, loop_body: &Expression)

fn track_variables_in_if(&mut self, if_expr: &If)

fn track_variables_in_match(&mut self, match_expr: &Match)

fn track_variables_in_tuple(&mut self, elements: &[Expression])

fn track_variables_in_call(&mut self, call: &Call)

fn track_variables_in_let(&mut self, let_expr: &Let)

fn track_variables_in_constrain( &mut self, boolean: &Expression, msg: &Option<Box<(Expression, Type)>>, )

fn track_variables_in_assign(&mut self, assign: &Assign)

fn track_variables_in_lvalue(&mut self, lvalue: &LValue, nested: bool)

A variable in an lvalue position is never moved (otherwise you wouldn’t be able to access the variable you assigned to afterward). However, the index in an array expression a[i] = ... is an arbitrary expression that is actually in an rvalue position and can thus be moved.

Subtle point: since we don’t track identifier uses here at all this means if the last use of one was just before it is assigned, it can actually be moved before it is assigned. This should be fine because we move out of the binding, and the binding isn’t used until it is set to a new value.

The nested parameter indicates whether this l-value is nested inside another l-value. For top-level identifiers there’s nothing to track, but for an identifier happening as part of an index (ident[index] = ...) we do want to consider ident as moved.