noirc_frontend
Module parser

The parser is the second pass of the noir compiler. The parser’s job is to take the output of the lexer (a stream of tokens) and parse it into a valid Abstract Syntax Tree (Ast). During this, the parser validates the grammar of the program and returns parsing errors for any syntactically invalid constructs (such as fn fn fn).

This file is mostly helper functions and types for the parser. For the parser itself, see parser.rs. The definition of the abstract syntax tree can be found in the ast folder.

Modules

errors 🔒

labels 🔒

parser 🔒

Structs

Item

ParsedModule

A ParsedModule contains an entire Ast for one file.

ParsedSubModule

A submodule defined via mod name { contents } in some larger file. These submodules always share the same file as some larger ParsedModule

Parser

ParserError

Represents a parsing error, or a parsing error in the making.

SortedModule

SortedSubModule

Enums

ItemKind

ParserErrorReason

StatementOrExpressionOrLValue

Functions

parse_program

Entry function for the parser - also handles lexing internally.

parse_program_with_dummy_file